Struct aws_sdk_ec2::client::fluent_builders::AssociateAddress

pub struct AssociateAddress<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> AssociateAddress<C>

pub async fn send(
self
) -> Result<AssociateAddressOutput, SdkError<AssociateAddressError>> where
C: SmithyConnector,

pub fn allocation_id(self, input: impl Into<String>) -> Self

[EC2-VPC] The allocation ID. This is required for EC2-VPC.

pub fn set_allocation_id(self, input: Option<String>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the instance. The instance must have exactly one attached network interface. For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both. For EC2-Classic, you must specify an instance ID and the instance must be in the running state.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn public_ip(self, input: impl Into<String>) -> Self

[EC2-Classic] The Elastic IP address to associate with the instance. This is required for EC2-Classic.

pub fn set_public_ip(self, input: Option<String>) -> Self

pub fn allow_reassociation(self, input: bool) -> Self

[EC2-VPC] For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address that is already associated with an instance or network interface to be reassociated with the specified instance or network interface. Otherwise, the operation fails. In a VPC in an EC2-VPC-only account, reassociation is automatic, therefore you can specify false to ensure the operation fails if the Elastic IP address is already associated with another resource.

pub fn set_allow_reassociation(self, input: Option<bool>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn network_interface_id(self, input: impl Into<String>) -> Self

[EC2-VPC] The ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.

For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both.

pub fn set_network_interface_id(self, input: Option<String>) -> Self

pub fn private_ip_address(self, input: impl Into<String>) -> Self

[EC2-VPC] The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

pub fn set_private_ip_address(self, input: Option<String>) -> Self