Struct aws_sdk_ec2::client::fluent_builders::AuthorizeSecurityGroupIngress

pub struct AuthorizeSecurityGroupIngress<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> AuthorizeSecurityGroupIngress<C>

pub async fn send(
self
) -> Result<AuthorizeSecurityGroupIngressOutput, SdkError<AuthorizeSecurityGroupIngressError>> where
C: SmithyConnector,

pub fn cidr_ip(self, input: impl Into<String>) -> Self

The IPv4 address range, in CIDR format. You can't specify this parameter when specifying a source security group. To specify an IPv6 address range, use a set of IP permissions.

Alternatively, use a set of IP permissions to specify multiple rules and a description for the rule.

pub fn set_cidr_ip(self, input: Option<String>) -> Self

pub fn from_port(self, input: i32) -> Self

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use -1 to specify all types. If you specify all ICMP types, you must specify all codes.

Alternatively, use a set of IP permissions to specify multiple rules and a description for the rule.

pub fn set_from_port(self, input: Option<i32>) -> Self

pub fn group_id(self, input: impl Into<String>) -> Self

The ID of the security group. You must specify either the security group ID or the security group name in the request. For security groups in a nondefault VPC, you must specify the security group ID.

pub fn set_group_id(self, input: Option<String>) -> Self

pub fn group_name(self, input: impl Into<String>) -> Self

[EC2-Classic, default VPC] The name of the security group. You must specify either the security group ID or the security group name in the request.

pub fn set_group_name(self, input: Option<String>) -> Self

pub fn ip_permissions(self, inp: impl Into<IpPermission>) -> Self

The sets of IP permissions.

pub fn set_ip_permissions(self, input: Option<Vec<IpPermission>>) -> Self

pub fn ip_protocol(self, input: impl Into<String>) -> Self

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). To specify icmpv6, use a set of IP permissions.

[VPC only] Use -1 to specify all protocols. If you specify -1 or a protocol other than tcp, udp, or icmp, traffic on all ports is allowed, regardless of any ports you specify.

Alternatively, use a set of IP permissions to specify multiple rules and a description for the rule.

pub fn set_ip_protocol(self, input: Option<String>) -> Self

pub fn source_security_group_name(self, input: impl Into<String>) -> Self

[EC2-Classic, default VPC] The name of the source security group. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the start of the port range, the IP protocol, and the end of the port range. Creates rules that grant full ICMP, UDP, and TCP access. To create a rule with a specific IP protocol and port range, use a set of IP permissions instead. For EC2-VPC, the source security group must be in the same VPC.

pub fn set_source_security_group_name(self, input: Option<String>) -> Self

pub fn source_security_group_owner_id(self, input: impl Into<String>) -> Self

[nondefault VPC] The AWS account ID for the source security group, if the source security group is in a different account. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the IP protocol, the start of the port range, and the end of the port range. Creates rules that grant full ICMP, UDP, and TCP access. To create a rule with a specific IP protocol and port range, use a set of IP permissions instead.

pub fn set_source_security_group_owner_id(self, input: Option<String>) -> Self

pub fn to_port(self, input: i32) -> Self

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use -1 to specify all codes. If you specify all ICMP types, you must specify all codes.

Alternatively, use a set of IP permissions to specify multiple rules and a description for the rule.

pub fn set_to_port(self, input: Option<i32>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self