Struct aws_sdk_ec2::client::fluent_builders::CreateCapacityReservation

pub struct CreateCapacityReservation<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateCapacityReservation<C>

pub async fn send(
self
) -> Result<CreateCapacityReservationOutput, SdkError<CreateCapacityReservationError>> where
C: SmithyConnector,

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensure Idempotency.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn instance_type(self, input: impl Into<String>) -> Self

The instance type for which to reserve capacity. For more information, see Instance types in the Amazon EC2 User Guide.

pub fn set_instance_type(self, input: Option<String>) -> Self

pub fn instance_platform(
self,
input: CapacityReservationInstancePlatform
) -> Self

The type of operating system for which to reserve capacity.

pub fn set_instance_platform(
self,
input: Option<CapacityReservationInstancePlatform>
) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone in which to create the Capacity Reservation.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn availability_zone_id(self, input: impl Into<String>) -> Self

The ID of the Availability Zone in which to create the Capacity Reservation.

pub fn set_availability_zone_id(self, input: Option<String>) -> Self

pub fn tenancy(self, input: CapacityReservationTenancy) -> Self

Indicates the tenancy of the Capacity Reservation. A Capacity Reservation can have one of the following tenancy settings:

default - The Capacity Reservation is created on hardware that is shared with other accounts.

dedicated - The Capacity Reservation is created on single-tenant hardware that is dedicated to a single account.

pub fn set_tenancy(self, input: Option<CapacityReservationTenancy>) -> Self

pub fn instance_count(self, input: i32) -> Self

The number of instances for which to reserve capacity.

pub fn set_instance_count(self, input: Option<i32>) -> Self

pub fn ebs_optimized(self, input: bool) -> Self

Indicates whether the Capacity Reservation supports EBS-optimized instances. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS- optimized instance.

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

pub fn ephemeral_storage(self, input: bool) -> Self

Indicates whether the Capacity Reservation supports instances with temporary, block-level storage.

pub fn set_ephemeral_storage(self, input: Option<bool>) -> Self

pub fn end_date(self, input: Instant) -> Self

The date and time at which the Capacity Reservation expires. When a Capacity Reservation expires, the reserved capacity is released and you can no longer launch instances into it. The Capacity Reservation's state changes to expired when it reaches its end date and time.

You must provide an EndDate value if EndDateType is limited. Omit EndDate if EndDateType is unlimited.

If the EndDateType is limited, the Capacity Reservation is cancelled within an hour from the specified time. For example, if you specify 5/31/2019, 13:30:55, the Capacity Reservation is guaranteed to end between 13:30:55 and 14:30:55 on 5/31/2019.

pub fn set_end_date(self, input: Option<Instant>) -> Self

pub fn end_date_type(self, input: EndDateType) -> Self

Indicates the way in which the Capacity Reservation ends. A Capacity Reservation can have one of the following end types:

unlimited - The Capacity Reservation remains active until you explicitly cancel it. Do not provide an EndDate if the EndDateType is unlimited.

limited - The Capacity Reservation expires automatically at a specified date and time. You must provide an EndDate value if the EndDateType value is limited.

pub fn set_end_date_type(self, input: Option<EndDateType>) -> Self

pub fn instance_match_criteria(self, input: InstanceMatchCriteria) -> Self

Indicates the type of instance launches that the Capacity Reservation accepts. The options include:

open - The Capacity Reservation automatically matches all instances that have matching attributes (instance type, platform, and Availability Zone). Instances that have matching attributes run in the Capacity Reservation automatically without specifying any additional parameters.

targeted - The Capacity Reservation only accepts instances that have matching attributes (instance type, platform, and Availability Zone), and explicitly target the Capacity Reservation. This ensures that only permitted instances can use the reserved capacity.

Default: open

pub fn set_instance_match_criteria(
self,
input: Option<InstanceMatchCriteria>
) -> Self

pub fn tag_specifications(self, inp: impl Into<TagSpecification>) -> Self

The tags to apply to the Capacity Reservation during launch.

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn outpost_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Outpost on which to create the Capacity Reservation.

pub fn set_outpost_arn(self, input: Option<String>) -> Self