Struct aws_sdk_ec2::client::fluent_builders::CreateClientVpnEndpoint

pub struct CreateClientVpnEndpoint<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateClientVpnEndpoint<C>

pub async fn send(
self
) -> Result<CreateClientVpnEndpointOutput, SdkError<CreateClientVpnEndpointError>> where
C: SmithyConnector,

pub fn client_cidr_block(self, input: impl Into<String>) -> Self

The IPv4 address range, in CIDR notation, from which to assign client IP addresses. The address range cannot overlap with the local CIDR of the VPC in which the associated subnet is located, or the routes that you add manually. The address range cannot be changed after the Client VPN endpoint has been created. The CIDR block should be /22 or greater.

pub fn set_client_cidr_block(self, input: Option<String>) -> Self

pub fn server_certificate_arn(self, input: impl Into<String>) -> Self

The ARN of the server certificate. For more information, see the AWS Certificate Manager User Guide.

pub fn set_server_certificate_arn(self, input: Option<String>) -> Self

pub fn authentication_options(
self,
inp: impl Into<ClientVpnAuthenticationRequest>
) -> Self

Information about the authentication method to be used to authenticate clients.

pub fn set_authentication_options(
self,
input: Option<Vec<ClientVpnAuthenticationRequest>>
) -> Self

pub fn connection_log_options(self, input: ConnectionLogOptions) -> Self

Information about the client connection logging options.

If you enable client connection logging, data about client connections is sent to a Cloudwatch Logs log stream. The following information is logged:

Client connection requests

Client connection results (successful and unsuccessful)

Reasons for unsuccessful client connection requests

Client connection termination time

pub fn set_connection_log_options(
self,
input: Option<ConnectionLogOptions>
) -> Self

pub fn dns_servers(self, inp: impl Into<String>) -> Self

Information about the DNS servers to be used for DNS resolution. A Client VPN endpoint can have up to two DNS servers. If no DNS server is specified, the DNS address configured on the device is used for the DNS server.

pub fn set_dns_servers(self, input: Option<Vec<String>>) -> Self

pub fn transport_protocol(self, input: TransportProtocol) -> Self

The transport protocol to be used by the VPN session.

Default value: udp

pub fn set_transport_protocol(self, input: Option<TransportProtocol>) -> Self

pub fn vpn_port(self, input: i32) -> Self

The port number to assign to the Client VPN endpoint for TCP and UDP traffic.

Valid Values: 443 | 1194

Default Value: 443

pub fn set_vpn_port(self, input: Option<i32>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A brief description of the Client VPN endpoint.

pub fn set_description(self, input: Option<String>) -> Self

pub fn split_tunnel(self, input: bool) -> Self

Indicates whether split-tunnel is enabled on the AWS Client VPN endpoint.

By default, split-tunnel on a VPN endpoint is disabled.

For information about split-tunnel VPN endpoints, see Split-Tunnel AWS Client VPN Endpoint in the AWS Client VPN Administrator Guide.

pub fn set_split_tunnel(self, input: Option<bool>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn tag_specifications(self, inp: impl Into<TagSpecification>) -> Self

The tags to apply to the Client VPN endpoint during creation.

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn security_group_ids(self, inp: impl Into<String>) -> Self

The IDs of one or more security groups to apply to the target network. You must also specify the ID of the VPC that contains the security groups.

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The ID of the VPC to associate with the Client VPN endpoint. If no security group IDs are specified in the request, the default security group for the VPC is applied.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn self_service_portal(self, input: SelfServicePortal) -> Self

Specify whether to enable the self-service portal for the Client VPN endpoint.

Default Value: enabled

pub fn set_self_service_portal(self, input: Option<SelfServicePortal>) -> Self

pub fn client_connect_options(self, input: ClientConnectOptions) -> Self

The options for managing connection authorization for new client connections.

pub fn set_client_connect_options(
self,
input: Option<ClientConnectOptions>
) -> Self