Struct aws_sdk_ec2::client::fluent_builders::CreateFlowLogs

pub struct CreateFlowLogs<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateFlowLogs<C>

pub async fn send(
self
) -> Result<CreateFlowLogsOutput, SdkError<CreateFlowLogsError>> where
C: SmithyConnector,

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn deliver_logs_permission_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits Amazon EC2 to publish flow logs to a CloudWatch Logs log group in your account.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

pub fn set_deliver_logs_permission_arn(self, input: Option<String>) -> Self

pub fn log_group_name(self, input: impl Into<String>) -> Self

The name of a new or existing CloudWatch Logs log group where Amazon EC2 publishes your flow logs.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

pub fn set_log_group_name(self, input: Option<String>) -> Self

pub fn resource_ids(self, inp: impl Into<String>) -> Self

The ID of the subnet, network interface, or VPC for which you want to create a flow log.

Constraints: Maximum of 1000 resources

pub fn set_resource_ids(self, input: Option<Vec<String>>) -> Self

pub fn resource_type(self, input: FlowLogsResourceType) -> Self

The type of resource for which to create the flow log. For example, if you specified a VPC ID for the ResourceId property, specify VPC for this property.

pub fn set_resource_type(self, input: Option<FlowLogsResourceType>) -> Self

pub fn traffic_type(self, input: TrafficType) -> Self

The type of traffic to log. You can log traffic that the resource accepts or rejects, or all traffic.

pub fn set_traffic_type(self, input: Option<TrafficType>) -> Self

pub fn log_destination_type(self, input: LogDestinationType) -> Self

Specifies the type of destination to which the flow log data is to be published. Flow log data can be published to CloudWatch Logs or Amazon S3. To publish flow log data to CloudWatch Logs, specify cloud-watch-logs. To publish flow log data to Amazon S3, specify s3.

If you specify LogDestinationType as s3, do not specify DeliverLogsPermissionArn or LogGroupName.

Default: cloud-watch-logs

pub fn set_log_destination_type(self, input: Option<LogDestinationType>) -> Self

pub fn log_destination(self, input: impl Into<String>) -> Self

Specifies the destination to which the flow log data is to be published. Flow log data can be published to a CloudWatch Logs log group or an Amazon S3 bucket. The value specified for this parameter depends on the value specified for LogDestinationType.

If LogDestinationType is not specified or cloud-watch-logs, specify the Amazon Resource Name (ARN) of the CloudWatch Logs log group. For example, to publish to a log group called my-logs, specify arn:aws:logs:us-east-1:123456789012:log-group:my-logs. Alternatively, use LogGroupName instead.

If LogDestinationType is s3, specify the ARN of the Amazon S3 bucket. You can also specify a subfolder in the bucket. To specify a subfolder in the bucket, use the following ARN format: bucket_ARN/subfolder_name/. For example, to specify a subfolder named my-logs in a bucket named my-bucket, use the following ARN: arn:aws:s3:::my-bucket/my-logs/. You cannot use AWSLogs as a subfolder name. This is a reserved term.

pub fn set_log_destination(self, input: Option<String>) -> Self

pub fn log_format(self, input: impl Into<String>) -> Self

The fields to include in the flow log record, in the order in which they should appear. For a list of available fields, see Flow Log Records. If you omit this parameter, the flow log is created using the default format. If you specify this parameter, you must specify at least one field.

Specify the fields using the ${field-id} format, separated by spaces. For the AWS CLI, use single quotation marks (' ') to surround the parameter value.

pub fn set_log_format(self, input: Option<String>) -> Self

pub fn tag_specifications(self, inp: impl Into<TagSpecification>) -> Self

The tags to apply to the flow logs.

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn max_aggregation_interval(self, input: i32) -> Self

The maximum interval of time during which a flow of packets is captured and aggregated into a flow log record. You can specify 60 seconds (1 minute) or 600 seconds (10 minutes).

When a network interface is attached to a Nitro-based instance, the aggregation interval is always 60 seconds or less, regardless of the value that you specify.

Default: 600

pub fn set_max_aggregation_interval(self, input: Option<i32>) -> Self