Struct aws_sdk_ec2::client::fluent_builders::CreateImage

pub struct CreateImage<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateImage<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<CreateImageOutput, SdkError<CreateImageError>> where
R::Policy: SmithyRetryPolicy<CreateImageInputOperationOutputAlias, CreateImageOutput, CreateImageError, CreateImageInputOperationRetryAlias>,

pub fn block_device_mappings(self, inp: impl Into<BlockDeviceMapping>) -> Self

Appends an item to BlockDeviceMappings.

To override the contents of this collection use set_block_device_mappings.

The block device mappings. This parameter cannot be used to modify the encryption status of existing volumes or snapshots. To create an AMI with encrypted snapshots, use the CopyImage action.

pub fn set_block_device_mappings(
self,
input: Option<Vec<BlockDeviceMapping>>
) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description for the new image.

pub fn set_description(self, input: Option<String>) -> Self

pub fn dry_run(self, inp: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn instance_id(self, inp: impl Into<String>) -> Self

The ID of the instance.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

A name for the new image.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

pub fn set_name(self, input: Option<String>) -> Self

pub fn no_reboot(self, inp: bool) -> Self

By default, Amazon EC2 attempts to shut down and reboot the instance before creating the image. If the No Reboot option is set, Amazon EC2 doesn't shut down the instance before creating the image. Without a reboot, the AMI will be crash consistent (all the volumes are snapshotted at the same time), but not application consistent (all the operating system buffers are not flushed to disk before the snapshots are created).

pub fn set_no_reboot(self, input: Option<bool>) -> Self

pub fn tag_specifications(self, inp: impl Into<TagSpecification>) -> Self

Appends an item to TagSpecifications.

To override the contents of this collection use set_tag_specifications.

The tags to apply to the AMI and snapshots on creation. You can tag the AMI, the snapshots, or both.

To tag the AMI, the value for ResourceType must be image.

To tag the snapshots that are created of the root volume and of other Amazon EBS volumes that are attached to the instance, the value for ResourceType must be snapshot. The same tag is applied to all of the snapshots that are created.

If you specify other values for ResourceType, the request fails.

To tag an AMI or snapshot after it has been created, see CreateTags.

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self