Struct aws_sdk_ec2::client::fluent_builders::CreateTransitGatewayConnectPeer

pub struct CreateTransitGatewayConnectPeer<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateTransitGatewayConnectPeer<C>

pub async fn send(
self
) -> Result<CreateTransitGatewayConnectPeerOutput, SdkError<CreateTransitGatewayConnectPeerError>> where
C: SmithyConnector,

pub fn transit_gateway_attachment_id(self, input: impl Into<String>) -> Self

The ID of the Connect attachment.

pub fn set_transit_gateway_attachment_id(self, input: Option<String>) -> Self

pub fn transit_gateway_address(self, input: impl Into<String>) -> Self

The peer IP address (GRE outer IP address) on the transit gateway side of the Connect peer, which must be specified from a transit gateway CIDR block. If not specified, Amazon automatically assigns the first available IP address from the transit gateway CIDR block.

pub fn set_transit_gateway_address(self, input: Option<String>) -> Self

pub fn peer_address(self, input: impl Into<String>) -> Self

The peer IP address (GRE outer IP address) on the appliance side of the Connect peer.

pub fn set_peer_address(self, input: Option<String>) -> Self

pub fn bgp_options(self, input: TransitGatewayConnectRequestBgpOptions) -> Self

The BGP options for the Connect peer.

pub fn set_bgp_options(
self,
input: Option<TransitGatewayConnectRequestBgpOptions>
) -> Self

pub fn inside_cidr_blocks(self, inp: impl Into<String>) -> Self

The range of inside IP addresses that are used for BGP peering. You must specify a size /29 IPv4 CIDR block from the 169.254.0.0/16 range. The first address from the range must be configured on the appliance as the BGP IP address. You can also optionally specify a size /125 IPv6 CIDR block from the fd00::/8 range.

pub fn set_inside_cidr_blocks(self, input: Option<Vec<String>>) -> Self

pub fn tag_specifications(self, inp: impl Into<TagSpecification>) -> Self

The tags to apply to the Connect peer.

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self