Struct aws_sdk_ec2::client::fluent_builders::CreateVolume

pub struct CreateVolume<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateVolume<C>

pub async fn send(
self
) -> Result<CreateVolumeOutput, SdkError<CreateVolumeError>> where
C: SmithyConnector,

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone in which to create the volume.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn encrypted(self, input: bool) -> Self

Indicates whether the volume should be encrypted. The effect of setting the encryption state to true depends on the volume origin (new or from a snapshot), starting encryption state, ownership, and whether encryption by default is enabled. For more information, see Encryption by default in the Amazon Elastic Compute Cloud User Guide.

Encrypted Amazon EBS volumes must be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types.

pub fn set_encrypted(self, input: Option<bool>) -> Self

pub fn iops(self, input: i32) -> Self

The number of I/O operations per second (IOPS). For gp3, io1, and io2 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

io2: 100-64,000 IOPS

For io1 and io2 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

This parameter is required for io1 and io2 volumes. The default for gp3 volumes is 3,000 IOPS. This parameter is not supported for gp2, st1, sc1, or standard volumes.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The identifier of the AWS Key Management Service (AWS KMS) customer master key (CMK) to use for Amazon EBS encryption. If this parameter is not specified, your AWS managed CMK for EBS is used. If KmsKeyId is specified, the encrypted state must be true.

You can specify the CMK using any of the following:

Key ID. For example, 1234abcd-12ab-34cd-56ef-1234567890ab.

Key alias. For example, alias/ExampleAlias.

Key ARN. For example, arn:aws:kms:us-east-1:012345678910:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Alias ARN. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

AWS authenticates the CMK asynchronously. Therefore, if you specify an ID, alias, or ARN that is not valid, the action can appear to complete, but eventually fails.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn outpost_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Outpost.

pub fn set_outpost_arn(self, input: Option<String>) -> Self

pub fn size(self, input: i32) -> Self

The size of the volume, in GiBs. You must specify either a snapshot ID or a volume size. If you specify a snapshot, the default is the snapshot size. You can specify a volume size that is equal to or larger than the snapshot size.

The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1 and io2: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

pub fn set_size(self, input: Option<i32>) -> Self

pub fn snapshot_id(self, input: impl Into<String>) -> Self

The snapshot from which to create the volume. You must specify either a snapshot ID or a volume size.

pub fn set_snapshot_id(self, input: Option<String>) -> Self

pub fn volume_type(self, input: VolumeType) -> Self

The volume type. This parameter can be one of the following values:

General Purpose SSD: gp2 | gp3

Provisioned IOPS SSD: io1 | io2

Throughput Optimized HDD: st1

Cold HDD: sc1

Magnetic: standard

For more information, see Amazon EBS volume types in the Amazon Elastic Compute Cloud User Guide.

Default: gp2

pub fn set_volume_type(self, input: Option<VolumeType>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn tag_specifications(self, inp: impl Into<TagSpecification>) -> Self

The tags to apply to the volume during creation.

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn multi_attach_enabled(self, input: bool) -> Self

Indicates whether to enable Amazon EBS Multi-Attach. If you enable Multi-Attach, you can attach the volume to up to 16 Instances built on the Nitro System in the same Availability Zone. This parameter is supported with io1 and io2 volumes only. For more information, see Amazon EBS Multi-Attach in the Amazon Elastic Compute Cloud User Guide.

pub fn set_multi_attach_enabled(self, input: Option<bool>) -> Self

pub fn throughput(self, input: i32) -> Self

The throughput to provision for a volume, with a maximum of 1,000 MiB/s.

This parameter is valid only for gp3 volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.

pub fn set_throughput(self, input: Option<i32>) -> Self