Struct aws_sdk_ec2::client::fluent_builders::DescribeClassicLinkInstances

pub struct DescribeClassicLinkInstances<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> DescribeClassicLinkInstances<C>

pub async fn send(
self
) -> Result<DescribeClassicLinkInstancesOutput, SdkError<DescribeClassicLinkInstancesError>> where
C: SmithyConnector,

pub fn filters(self, inp: impl Into<Filter>) -> Self

One or more filters.

group-id - The ID of a VPC security group that's associated with the instance.

instance-id - The ID of the instance.

tag: - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

vpc-id - The ID of the VPC to which the instance is linked.

vpc-id - The ID of the VPC that the instance is linked to.

pub fn set_filters(self, input: Option<Vec<Filter>>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn instance_ids(self, inp: impl Into<String>) -> Self

One or more instance IDs. Must be instances linked to a VPC through ClassicLink.

pub fn set_instance_ids(self, input: Option<Vec<String>>) -> Self

pub fn max_results(self, input: i32) -> Self

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

Constraint: If the value is greater than 1000, we return only 1000 items.

pub fn set_max_results(self, input: Option<i32>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

The token for the next page of results.

pub fn set_next_token(self, input: Option<String>) -> Self