Struct aws_sdk_ec2::client::fluent_builders::DescribeReservedInstancesOfferings

pub struct DescribeReservedInstancesOfferings<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> DescribeReservedInstancesOfferings<C>

pub async fn send(
self
) -> Result<DescribeReservedInstancesOfferingsOutput, SdkError<DescribeReservedInstancesOfferingsError>> where
C: SmithyConnector,

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone in which the Reserved Instance can be used.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn filters(self, inp: impl Into<Filter>) -> Self

One or more filters.

availability-zone - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (for example, one year or three years), in seconds (31536000 | 94608000).

fixed-price - The purchase price of the Reserved Instance (for example, 9800.0).

instance-type - The instance type that is covered by the reservation.

marketplace - Set to true to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from both Amazon Web Services and the Reserved Instance Marketplace are listed.

product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC. (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Red Hat Enterprise Linux with HA (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC))

reserved-instances-offering-id - The Reserved Instances offering ID.

scope - The scope of the Reserved Instance (Availability Zone or Region).

usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

pub fn set_filters(self, input: Option<Vec<Filter>>) -> Self

pub fn include_marketplace(self, input: bool) -> Self

Include Reserved Instance Marketplace offerings in the response.

pub fn set_include_marketplace(self, input: Option<bool>) -> Self

pub fn instance_type(self, input: InstanceType) -> Self

The instance type that the reservation will cover (for example, m1.small). For more information, see Instance types in the Amazon EC2 User Guide.

pub fn set_instance_type(self, input: Option<InstanceType>) -> Self

pub fn max_duration(self, input: i64) -> Self

The maximum duration (in seconds) to filter when searching for offerings.

Default: 94608000 (3 years)

pub fn set_max_duration(self, input: Option<i64>) -> Self

pub fn max_instance_count(self, input: i32) -> Self

The maximum number of instances to filter when searching for offerings.

Default: 20

pub fn set_max_instance_count(self, input: Option<i32>) -> Self

pub fn min_duration(self, input: i64) -> Self

The minimum duration (in seconds) to filter when searching for offerings.

Default: 2592000 (1 month)

pub fn set_min_duration(self, input: Option<i64>) -> Self

pub fn offering_class(self, input: OfferingClassType) -> Self

The offering class of the Reserved Instance. Can be standard or convertible.

pub fn set_offering_class(self, input: Option<OfferingClassType>) -> Self

pub fn product_description(self, input: RiProductDescription) -> Self

The Reserved Instance product platform description. Instances that include (Amazon VPC) in the description are for use with Amazon VPC.

pub fn set_product_description(
self,
input: Option<RiProductDescription>
) -> Self

pub fn reserved_instances_offering_ids(self, inp: impl Into<String>) -> Self

One or more Reserved Instances offering IDs.

pub fn set_reserved_instances_offering_ids(
self,
input: Option<Vec<String>>
) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn instance_tenancy(self, input: Tenancy) -> Self

The tenancy of the instances covered by the reservation. A Reserved Instance with a tenancy of dedicated is applied to instances that run in a VPC on single-tenant hardware (i.e., Dedicated Instances).

Important: The host value cannot be used with this parameter. Use the default or dedicated values only.

Default: default

pub fn set_instance_tenancy(self, input: Option<Tenancy>) -> Self

pub fn max_results(self, input: i32) -> Self

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. The maximum is 100.

Default: 100

pub fn set_max_results(self, input: Option<i32>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

The token to retrieve the next page of results.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn offering_type(self, input: OfferingTypeValues) -> Self

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

pub fn set_offering_type(self, input: Option<OfferingTypeValues>) -> Self