Struct aws_sdk_ec2::client::fluent_builders::DescribeSnapshots

pub struct DescribeSnapshots<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> DescribeSnapshots<C>

pub async fn send(
self
) -> Result<DescribeSnapshotsOutput, SdkError<DescribeSnapshotsError>> where
C: SmithyConnector,

pub fn filters(self, inp: impl Into<Filter>) -> Self

The filters.

description - A description of the snapshot.

encrypted - Indicates whether the snapshot is encrypted (true | false)

owner-alias - The owner alias, from an Amazon-maintained list (amazon). This is not the user-configured AWS account alias set using the IAM console. We recommend that you use the related parameter instead of this filter.

owner-id - The AWS account ID of the owner. We recommend that you use the related parameter instead of this filter.

progress - The progress of the snapshot, as a percentage (for example, 80%).

snapshot-id - The snapshot ID.

start-time - The time stamp when the snapshot was initiated.

status - The status of the snapshot (pending | completed | error).

tag: - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key Owner and the value TeamA, specify tag:Owner for the filter name and TeamA for the filter value.

tag-key - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.

volume-id - The ID of the volume the snapshot is for.

volume-size - The size of the volume, in GiB.

pub fn set_filters(self, input: Option<Vec<Filter>>) -> Self

pub fn max_results(self, input: i32) -> Self

The maximum number of snapshot results returned by DescribeSnapshots in paginated output. When this parameter is used, DescribeSnapshots only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeSnapshots request with the returned NextToken value. This value can be between 5 and 1,000; if MaxResults is given a value larger than 1,000, only 1,000 results are returned. If this parameter is not used, then DescribeSnapshots returns all results. You cannot specify this parameter and the snapshot IDs parameter in the same request.

pub fn set_max_results(self, input: Option<i32>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

The NextToken value returned from a previous paginated DescribeSnapshots request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn owner_ids(self, inp: impl Into<String>) -> Self

Scopes the results to snapshots with the specified owners. You can specify a combination of AWS account IDs, self, and amazon.

pub fn set_owner_ids(self, input: Option<Vec<String>>) -> Self

pub fn restorable_by_user_ids(self, inp: impl Into<String>) -> Self

The IDs of the AWS accounts that can create volumes from the snapshot.

pub fn set_restorable_by_user_ids(self, input: Option<Vec<String>>) -> Self

pub fn snapshot_ids(self, inp: impl Into<String>) -> Self

The snapshot IDs.

Default: Describes the snapshots for which you have create volume permissions.

pub fn set_snapshot_ids(self, input: Option<Vec<String>>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self