Struct aws_sdk_ec2::client::fluent_builders::ImportImage

pub struct ImportImage<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> ImportImage<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<ImportImageOutput, SdkError<ImportImageError>> where
R::Policy: SmithyRetryPolicy<ImportImageInputOperationOutputAlias, ImportImageOutput, ImportImageError, ImportImageInputOperationRetryAlias>,

pub fn architecture(self, inp: impl Into<String>) -> Self

The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

pub fn set_architecture(self, input: Option<String>) -> Self

pub fn client_data(self, inp: ClientData) -> Self

The client-specific data.

pub fn set_client_data(self, input: Option<ClientData>) -> Self

pub fn client_token(self, inp: impl Into<String>) -> Self

The token to enable idempotency for VM import requests.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description string for the import image task.

pub fn set_description(self, input: Option<String>) -> Self

pub fn disk_containers(self, inp: impl Into<ImageDiskContainer>) -> Self

Appends an item to DiskContainers.

To override the contents of this collection use set_disk_containers.

Information about the disk containers.

pub fn set_disk_containers(self, input: Option<Vec<ImageDiskContainer>>) -> Self

pub fn dry_run(self, inp: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn encrypted(self, inp: bool) -> Self

Specifies whether the destination AMI of the imported image should be encrypted. The default KMS key for EBS is used unless you specify a non-default KMS key using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

pub fn set_encrypted(self, input: Option<bool>) -> Self

pub fn hypervisor(self, inp: impl Into<String>) -> Self

The target hypervisor platform.

Valid values: xen

pub fn set_hypervisor(self, input: Option<String>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

An identifier for the symmetric KMS key to use when creating the encrypted AMI. This parameter is only required if you want to use a non-default KMS key; if this parameter is not specified, the default KMS key for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The KMS key identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the key namespace, and then the key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the key, the Amazon Web Services account ID of the key owner, the alias namespace, and then the key alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

Amazon Web Services parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified KMS key must exist in the Region that the AMI is being copied to.

Amazon EBS does not support asymmetric KMS keys.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn license_type(self, inp: impl Into<String>) -> Self

The license type to be used for the Amazon Machine Image (AMI) after importing.

By default, we detect the source-system operating system (OS) and apply the appropriate license. Specify AWS to replace the source-system license with an Amazon Web Services license, if appropriate. Specify BYOL to retain the source-system license, if appropriate.

To use BYOL, you must have existing licenses with rights to use these licenses in a third party cloud, such as Amazon Web Services. For more information, see Prerequisites in the VM Import/Export User Guide.

pub fn set_license_type(self, input: Option<String>) -> Self

pub fn platform(self, inp: impl Into<String>) -> Self

The operating system of the virtual machine.

Valid values: Windows | Linux

pub fn set_platform(self, input: Option<String>) -> Self

pub fn role_name(self, inp: impl Into<String>) -> Self

The name of the role to use when not using the default role, 'vmimport'.

pub fn set_role_name(self, input: Option<String>) -> Self

pub fn license_specifications(
self,
inp: impl Into<ImportImageLicenseConfigurationRequest>
) -> Self

Appends an item to LicenseSpecifications.

To override the contents of this collection use set_license_specifications.

The ARNs of the license configurations.

pub fn set_license_specifications(
self,
input: Option<Vec<ImportImageLicenseConfigurationRequest>>
) -> Self

pub fn tag_specifications(self, inp: impl Into<TagSpecification>) -> Self

Appends an item to TagSpecifications.

To override the contents of this collection use set_tag_specifications.

The tags to apply to the import image task during creation.

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn usage_operation(self, inp: impl Into<String>) -> Self

The usage operation value. For more information, see AMI billing information fields in the Amazon Elastic Compute Cloud User Guide.

pub fn set_usage_operation(self, input: Option<String>) -> Self

pub fn boot_mode(self, inp: BootModeValues) -> Self

The boot mode of the virtual machine.

pub fn set_boot_mode(self, input: Option<BootModeValues>) -> Self