Struct aws_sdk_ec2::client::fluent_builders::ModifyInstanceAttribute

pub struct ModifyInstanceAttribute<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> ModifyInstanceAttribute<C>

pub async fn send(
self
) -> Result<ModifyInstanceAttributeOutput, SdkError<ModifyInstanceAttributeError>> where
C: SmithyConnector,

pub fn source_dest_check(self, input: AttributeBooleanValue) -> Self

Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

pub fn set_source_dest_check(self, input: Option<AttributeBooleanValue>) -> Self

pub fn attribute(self, input: InstanceAttributeName) -> Self

The name of the attribute.

pub fn set_attribute(self, input: Option<InstanceAttributeName>) -> Self

pub fn block_device_mappings(
self,
inp: impl Into<InstanceBlockDeviceMappingSpecification>
) -> Self

Modifies the DeleteOnTermination attribute for volumes that are currently attached. The volume must be owned by the caller. If no value is specified for DeleteOnTermination, the default is true and the volume is deleted when the instance is terminated.

To add instance store volumes to an Amazon EBS-backed instance, you must add them when you launch the instance. For more information, see Updating the block device mapping when launching an instance in the Amazon EC2 User Guide.

pub fn set_block_device_mappings(
self,
input: Option<Vec<InstanceBlockDeviceMappingSpecification>>
) -> Self

pub fn disable_api_termination(self, input: AttributeBooleanValue) -> Self

If the value is true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. You cannot use this parameter for Spot Instances.

pub fn set_disable_api_termination(
self,
input: Option<AttributeBooleanValue>
) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn ebs_optimized(self, input: AttributeBooleanValue) -> Self

Specifies whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

pub fn set_ebs_optimized(self, input: Option<AttributeBooleanValue>) -> Self

pub fn ena_support(self, input: AttributeBooleanValue) -> Self

Set to true to enable enhanced networking with ENA for the instance.

This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

pub fn set_ena_support(self, input: Option<AttributeBooleanValue>) -> Self

pub fn groups(self, inp: impl Into<String>) -> Self

[EC2-VPC] Replaces the security groups of the instance with the specified security groups. You must specify at least one security group, even if it's just the default security group for the VPC. You must specify the security group ID, not the security group name.

pub fn set_groups(self, input: Option<Vec<String>>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the instance.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn instance_initiated_shutdown_behavior(self, input: AttributeValue) -> Self

Specifies whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

pub fn set_instance_initiated_shutdown_behavior(
self,
input: Option<AttributeValue>
) -> Self

pub fn instance_type(self, input: AttributeValue) -> Self

Changes the instance type to the specified value. For more information, see Instance types in the Amazon EC2 User Guide. If the instance type is not valid, the error returned is InvalidInstanceAttributeValue.

pub fn set_instance_type(self, input: Option<AttributeValue>) -> Self

pub fn kernel(self, input: AttributeValue) -> Self

Changes the instance's kernel to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB.

pub fn set_kernel(self, input: Option<AttributeValue>) -> Self

pub fn ramdisk(self, input: AttributeValue) -> Self

Changes the instance's RAM disk to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB.

pub fn set_ramdisk(self, input: Option<AttributeValue>) -> Self

pub fn sriov_net_support(self, input: AttributeValue) -> Self

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the instance.

There is no way to disable enhanced networking with the Intel 82599 Virtual Function interface at this time.

This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

pub fn set_sriov_net_support(self, input: Option<AttributeValue>) -> Self

pub fn user_data(self, input: BlobAttributeValue) -> Self

Changes the instance's user data to the specified value. If you are using an Amazon Web Services SDK or command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text.

pub fn set_user_data(self, input: Option<BlobAttributeValue>) -> Self

pub fn value(self, input: impl Into<String>) -> Self

A new value for the attribute. Use only with the kernel, ramdisk, userData, disableApiTermination, or instanceInitiatedShutdownBehavior attribute.

pub fn set_value(self, input: Option<String>) -> Self