Struct aws_sdk_ec2::client::fluent_builders::ModifyManagedPrefixList

pub struct ModifyManagedPrefixList<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> ModifyManagedPrefixList<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<ModifyManagedPrefixListOutput, SdkError<ModifyManagedPrefixListError>> where
R::Policy: SmithyRetryPolicy<ModifyManagedPrefixListInputOperationOutputAlias, ModifyManagedPrefixListOutput, ModifyManagedPrefixListError, ModifyManagedPrefixListInputOperationRetryAlias>,

pub fn dry_run(self, inp: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn prefix_list_id(self, inp: impl Into<String>) -> Self

The ID of the prefix list.

pub fn set_prefix_list_id(self, input: Option<String>) -> Self

pub fn current_version(self, inp: i64) -> Self

The current version of the prefix list.

pub fn set_current_version(self, input: Option<i64>) -> Self

pub fn prefix_list_name(self, inp: impl Into<String>) -> Self

A name for the prefix list.

pub fn set_prefix_list_name(self, input: Option<String>) -> Self

pub fn add_entries(self, inp: impl Into<AddPrefixListEntry>) -> Self

Appends an item to AddEntries.

To override the contents of this collection use set_add_entries.

One or more entries to add to the prefix list.

pub fn set_add_entries(self, input: Option<Vec<AddPrefixListEntry>>) -> Self

pub fn remove_entries(self, inp: impl Into<RemovePrefixListEntry>) -> Self

Appends an item to RemoveEntries.

To override the contents of this collection use set_remove_entries.

One or more entries to remove from the prefix list.

pub fn set_remove_entries(
self,
input: Option<Vec<RemovePrefixListEntry>>
) -> Self

pub fn max_entries(self, inp: i32) -> Self

The maximum number of entries for the prefix list. You cannot modify the entries of a prefix list and modify the size of a prefix list at the same time.

If any of the resources that reference the prefix list cannot support the new maximum size, the modify operation fails. Check the state message for the IDs of the first ten resources that do not support the new maximum size.

pub fn set_max_entries(self, input: Option<i32>) -> Self