Struct aws_sdk_ec2::client::fluent_builders::RegisterImage

pub struct RegisterImage<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RegisterImage<C>

pub async fn send(
self
) -> Result<RegisterImageOutput, SdkError<RegisterImageError>> where
C: SmithyConnector,

pub fn image_location(self, input: impl Into<String>) -> Self

The full path to your AMI manifest in Amazon S3 storage. The specified bucket must have the aws-exec-read canned access control list (ACL) to ensure that it can be accessed by Amazon EC2. For more information, see Canned ACLs in the Amazon S3 Service Developer Guide.

pub fn set_image_location(self, input: Option<String>) -> Self

pub fn architecture(self, input: ArchitectureValues) -> Self

The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

pub fn set_architecture(self, input: Option<ArchitectureValues>) -> Self

pub fn block_device_mappings(self, inp: impl Into<BlockDeviceMapping>) -> Self

The block device mapping entries.

If you specify an EBS volume using the ID of an EBS snapshot, you can't specify the encryption state of the volume.

If you create an AMI on an Outpost, then all backing snapshots must be on the same Outpost or in the Region of that Outpost. AMIs on an Outpost that include local snapshots can be used to launch instances on the same Outpost only. For more information, Amazon EBS local snapshots on Outposts in the Amazon Elastic Compute Cloud User Guide.

pub fn set_block_device_mappings(
self,
input: Option<Vec<BlockDeviceMapping>>
) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description for your AMI.

pub fn set_description(self, input: Option<String>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn ena_support(self, input: bool) -> Self

Set to true to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

pub fn set_ena_support(self, input: Option<bool>) -> Self

pub fn kernel_id(self, input: impl Into<String>) -> Self

The ID of the kernel.

pub fn set_kernel_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

pub fn set_name(self, input: Option<String>) -> Self

pub fn billing_products(self, inp: impl Into<String>) -> Self

The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the AWS Marketplace to bill for the use of an AMI.

pub fn set_billing_products(self, input: Option<Vec<String>>) -> Self

pub fn ramdisk_id(self, input: impl Into<String>) -> Self

The ID of the RAM disk.

pub fn set_ramdisk_id(self, input: Option<String>) -> Self

pub fn root_device_name(self, input: impl Into<String>) -> Self

The device name of the root device volume (for example, /dev/sda1).

pub fn set_root_device_name(self, input: Option<String>) -> Self

pub fn sriov_net_support(self, input: impl Into<String>) -> Self

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI.

There is no way to disable sriovNetSupport at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

pub fn set_sriov_net_support(self, input: Option<String>) -> Self

pub fn virtualization_type(self, input: impl Into<String>) -> Self

The type of virtualization (hvm | paravirtual).

Default: paravirtual

pub fn set_virtualization_type(self, input: Option<String>) -> Self

pub fn boot_mode(self, input: BootModeValues) -> Self

The boot mode of the AMI. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

pub fn set_boot_mode(self, input: Option<BootModeValues>) -> Self