Struct aws_sdk_ec2::client::fluent_builders::RevokeSecurityGroupEgress

pub struct RevokeSecurityGroupEgress<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RevokeSecurityGroupEgress<C>

pub async fn send(
self
) -> Result<RevokeSecurityGroupEgressOutput, SdkError<RevokeSecurityGroupEgressError>> where
C: SmithyConnector,

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn group_id(self, input: impl Into<String>) -> Self

The ID of the security group.

pub fn set_group_id(self, input: Option<String>) -> Self

pub fn ip_permissions(self, inp: impl Into<IpPermission>) -> Self

The sets of IP permissions. You can't specify a destination security group and a CIDR IP address range in the same set of permissions.

pub fn set_ip_permissions(self, input: Option<Vec<IpPermission>>) -> Self

pub fn cidr_ip(self, input: impl Into<String>) -> Self

Not supported. Use a set of IP permissions to specify the CIDR.

pub fn set_cidr_ip(self, input: Option<String>) -> Self

pub fn from_port(self, input: i32) -> Self

Not supported. Use a set of IP permissions to specify the port.

pub fn set_from_port(self, input: Option<i32>) -> Self

pub fn ip_protocol(self, input: impl Into<String>) -> Self

Not supported. Use a set of IP permissions to specify the protocol name or number.

pub fn set_ip_protocol(self, input: Option<String>) -> Self

pub fn to_port(self, input: i32) -> Self

Not supported. Use a set of IP permissions to specify the port.

pub fn set_to_port(self, input: Option<i32>) -> Self

pub fn source_security_group_name(self, input: impl Into<String>) -> Self

Not supported. Use a set of IP permissions to specify a destination security group.

pub fn set_source_security_group_name(self, input: Option<String>) -> Self

pub fn source_security_group_owner_id(self, input: impl Into<String>) -> Self

Not supported. Use a set of IP permissions to specify a destination security group.

pub fn set_source_security_group_owner_id(self, input: Option<String>) -> Self