Struct aws_sdk_ec2::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonEC2.

This client allows ergonomic access to a AmazonEC2-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_reserved_instances_exchange_quote(
&self
) -> AcceptReservedInstancesExchangeQuote<C, M, R>

pub fn accept_transit_gateway_multicast_domain_associations(
&self
) -> AcceptTransitGatewayMulticastDomainAssociations<C, M, R>

pub fn accept_transit_gateway_peering_attachment(
&self
) -> AcceptTransitGatewayPeeringAttachment<C, M, R>

pub fn accept_transit_gateway_vpc_attachment(
&self
) -> AcceptTransitGatewayVpcAttachment<C, M, R>

pub fn accept_vpc_endpoint_connections(
&self
) -> AcceptVpcEndpointConnections<C, M, R>

pub fn accept_vpc_peering_connection(
&self
) -> AcceptVpcPeeringConnection<C, M, R>

pub fn advertise_byoip_cidr(&self) -> AdvertiseByoipCidr<C, M, R>

pub fn allocate_address(&self) -> AllocateAddress<C, M, R>

pub fn allocate_hosts(&self) -> AllocateHosts<C, M, R>

pub fn apply_security_groups_to_client_vpn_target_network(
&self
) -> ApplySecurityGroupsToClientVpnTargetNetwork<C, M, R>

pub fn assign_ipv6_addresses(&self) -> AssignIpv6Addresses<C, M, R>

pub fn assign_private_ip_addresses(&self) -> AssignPrivateIpAddresses<C, M, R>

pub fn associate_address(&self) -> AssociateAddress<C, M, R>

pub fn associate_client_vpn_target_network(
&self
) -> AssociateClientVpnTargetNetwork<C, M, R>

pub fn associate_dhcp_options(&self) -> AssociateDhcpOptions<C, M, R>

pub fn associate_enclave_certificate_iam_role(
&self
) -> AssociateEnclaveCertificateIamRole<C, M, R>

pub fn associate_iam_instance_profile(
&self
) -> AssociateIamInstanceProfile<C, M, R>

pub fn associate_instance_event_window(
&self
) -> AssociateInstanceEventWindow<C, M, R>

pub fn associate_route_table(&self) -> AssociateRouteTable<C, M, R>

pub fn associate_subnet_cidr_block(&self) -> AssociateSubnetCidrBlock<C, M, R>

pub fn associate_transit_gateway_multicast_domain(
&self
) -> AssociateTransitGatewayMulticastDomain<C, M, R>

pub fn associate_transit_gateway_route_table(
&self
) -> AssociateTransitGatewayRouteTable<C, M, R>

pub fn associate_trunk_interface(&self) -> AssociateTrunkInterface<C, M, R>

pub fn associate_vpc_cidr_block(&self) -> AssociateVpcCidrBlock<C, M, R>

pub fn attach_classic_link_vpc(&self) -> AttachClassicLinkVpc<C, M, R>

pub fn attach_internet_gateway(&self) -> AttachInternetGateway<C, M, R>

pub fn attach_network_interface(&self) -> AttachNetworkInterface<C, M, R>

pub fn attach_volume(&self) -> AttachVolume<C, M, R>

pub fn attach_vpn_gateway(&self) -> AttachVpnGateway<C, M, R>

pub fn authorize_client_vpn_ingress(&self) -> AuthorizeClientVpnIngress<C, M, R>

pub fn authorize_security_group_egress(
&self
) -> AuthorizeSecurityGroupEgress<C, M, R>

pub fn authorize_security_group_ingress(
&self
) -> AuthorizeSecurityGroupIngress<C, M, R>

pub fn bundle_instance(&self) -> BundleInstance<C, M, R>

pub fn cancel_bundle_task(&self) -> CancelBundleTask<C, M, R>

pub fn cancel_capacity_reservation(&self) -> CancelCapacityReservation<C, M, R>

pub fn cancel_conversion_task(&self) -> CancelConversionTask<C, M, R>

pub fn cancel_export_task(&self) -> CancelExportTask<C, M, R>

pub fn cancel_import_task(&self) -> CancelImportTask<C, M, R>

pub fn cancel_reserved_instances_listing(
&self
) -> CancelReservedInstancesListing<C, M, R>

pub fn cancel_spot_fleet_requests(&self) -> CancelSpotFleetRequests<C, M, R>

pub fn cancel_spot_instance_requests(
&self
) -> CancelSpotInstanceRequests<C, M, R>

pub fn confirm_product_instance(&self) -> ConfirmProductInstance<C, M, R>

pub fn copy_fpga_image(&self) -> CopyFpgaImage<C, M, R>

pub fn copy_image(&self) -> CopyImage<C, M, R>

pub fn copy_snapshot(&self) -> CopySnapshot<C, M, R>

pub fn create_capacity_reservation(&self) -> CreateCapacityReservation<C, M, R>

pub fn create_carrier_gateway(&self) -> CreateCarrierGateway<C, M, R>

pub fn create_client_vpn_endpoint(&self) -> CreateClientVpnEndpoint<C, M, R>

pub fn create_client_vpn_route(&self) -> CreateClientVpnRoute<C, M, R>

pub fn create_customer_gateway(&self) -> CreateCustomerGateway<C, M, R>

pub fn create_default_subnet(&self) -> CreateDefaultSubnet<C, M, R>

pub fn create_default_vpc(&self) -> CreateDefaultVpc<C, M, R>

pub fn create_dhcp_options(&self) -> CreateDhcpOptions<C, M, R>

pub fn create_egress_only_internet_gateway(
&self
) -> CreateEgressOnlyInternetGateway<C, M, R>

pub fn create_fleet(&self) -> CreateFleet<C, M, R>

pub fn create_flow_logs(&self) -> CreateFlowLogs<C, M, R>

pub fn create_fpga_image(&self) -> CreateFpgaImage<C, M, R>

pub fn create_image(&self) -> CreateImage<C, M, R>

pub fn create_instance_event_window(&self) -> CreateInstanceEventWindow<C, M, R>

pub fn create_instance_export_task(&self) -> CreateInstanceExportTask<C, M, R>

pub fn create_internet_gateway(&self) -> CreateInternetGateway<C, M, R>

pub fn create_key_pair(&self) -> CreateKeyPair<C, M, R>

pub fn create_launch_template(&self) -> CreateLaunchTemplate<C, M, R>

pub fn create_launch_template_version(
&self
) -> CreateLaunchTemplateVersion<C, M, R>

pub fn create_local_gateway_route(&self) -> CreateLocalGatewayRoute<C, M, R>

pub fn create_local_gateway_route_table_vpc_association(
&self
) -> CreateLocalGatewayRouteTableVpcAssociation<C, M, R>

pub fn create_managed_prefix_list(&self) -> CreateManagedPrefixList<C, M, R>

pub fn create_nat_gateway(&self) -> CreateNatGateway<C, M, R>

pub fn create_network_acl(&self) -> CreateNetworkAcl<C, M, R>

pub fn create_network_acl_entry(&self) -> CreateNetworkAclEntry<C, M, R>

pub fn create_network_insights_path(&self) -> CreateNetworkInsightsPath<C, M, R>

pub fn create_network_interface(&self) -> CreateNetworkInterface<C, M, R>

pub fn create_network_interface_permission(
&self
) -> CreateNetworkInterfacePermission<C, M, R>

pub fn create_placement_group(&self) -> CreatePlacementGroup<C, M, R>

pub fn create_replace_root_volume_task(
&self
) -> CreateReplaceRootVolumeTask<C, M, R>

pub fn create_reserved_instances_listing(
&self
) -> CreateReservedInstancesListing<C, M, R>

pub fn create_restore_image_task(&self) -> CreateRestoreImageTask<C, M, R>

pub fn create_route(&self) -> CreateRoute<C, M, R>

pub fn create_route_table(&self) -> CreateRouteTable<C, M, R>

pub fn create_security_group(&self) -> CreateSecurityGroup<C, M, R>

pub fn create_snapshot(&self) -> CreateSnapshot<C, M, R>

pub fn create_snapshots(&self) -> CreateSnapshots<C, M, R>

pub fn create_spot_datafeed_subscription(
&self
) -> CreateSpotDatafeedSubscription<C, M, R>

pub fn create_store_image_task(&self) -> CreateStoreImageTask<C, M, R>

pub fn create_subnet(&self) -> CreateSubnet<C, M, R>

pub fn create_subnet_cidr_reservation(
&self
) -> CreateSubnetCidrReservation<C, M, R>

pub fn create_tags(&self) -> CreateTags<C, M, R>

pub fn create_traffic_mirror_filter(&self) -> CreateTrafficMirrorFilter<C, M, R>

pub fn create_traffic_mirror_filter_rule(
&self
) -> CreateTrafficMirrorFilterRule<C, M, R>

pub fn create_traffic_mirror_session(
&self
) -> CreateTrafficMirrorSession<C, M, R>

pub fn create_traffic_mirror_target(&self) -> CreateTrafficMirrorTarget<C, M, R>

pub fn create_transit_gateway(&self) -> CreateTransitGateway<C, M, R>

pub fn create_transit_gateway_connect(
&self
) -> CreateTransitGatewayConnect<C, M, R>

pub fn create_transit_gateway_connect_peer(
&self
) -> CreateTransitGatewayConnectPeer<C, M, R>

pub fn create_transit_gateway_multicast_domain(
&self
) -> CreateTransitGatewayMulticastDomain<C, M, R>

pub fn create_transit_gateway_peering_attachment(
&self
) -> CreateTransitGatewayPeeringAttachment<C, M, R>

pub fn create_transit_gateway_prefix_list_reference(
&self
) -> CreateTransitGatewayPrefixListReference<C, M, R>

pub fn create_transit_gateway_route(&self) -> CreateTransitGatewayRoute<C, M, R>

pub fn create_transit_gateway_route_table(
&self
) -> CreateTransitGatewayRouteTable<C, M, R>

pub fn create_transit_gateway_vpc_attachment(
&self
) -> CreateTransitGatewayVpcAttachment<C, M, R>

pub fn create_volume(&self) -> CreateVolume<C, M, R>

pub fn create_vpc(&self) -> CreateVpc<C, M, R>

pub fn create_vpc_endpoint(&self) -> CreateVpcEndpoint<C, M, R>

pub fn create_vpc_endpoint_connection_notification(
&self
) -> CreateVpcEndpointConnectionNotification<C, M, R>

pub fn create_vpc_endpoint_service_configuration(
&self
) -> CreateVpcEndpointServiceConfiguration<C, M, R>

pub fn create_vpc_peering_connection(
&self
) -> CreateVpcPeeringConnection<C, M, R>

pub fn create_vpn_connection(&self) -> CreateVpnConnection<C, M, R>

pub fn create_vpn_connection_route(&self) -> CreateVpnConnectionRoute<C, M, R>

pub fn create_vpn_gateway(&self) -> CreateVpnGateway<C, M, R>

pub fn delete_carrier_gateway(&self) -> DeleteCarrierGateway<C, M, R>

pub fn delete_client_vpn_endpoint(&self) -> DeleteClientVpnEndpoint<C, M, R>

pub fn delete_client_vpn_route(&self) -> DeleteClientVpnRoute<C, M, R>

pub fn delete_customer_gateway(&self) -> DeleteCustomerGateway<C, M, R>

pub fn delete_dhcp_options(&self) -> DeleteDhcpOptions<C, M, R>

pub fn delete_egress_only_internet_gateway(
&self
) -> DeleteEgressOnlyInternetGateway<C, M, R>

pub fn delete_fleets(&self) -> DeleteFleets<C, M, R>

pub fn delete_flow_logs(&self) -> DeleteFlowLogs<C, M, R>

pub fn delete_fpga_image(&self) -> DeleteFpgaImage<C, M, R>

pub fn delete_instance_event_window(&self) -> DeleteInstanceEventWindow<C, M, R>

pub fn delete_internet_gateway(&self) -> DeleteInternetGateway<C, M, R>

pub fn delete_key_pair(&self) -> DeleteKeyPair<C, M, R>

pub fn delete_launch_template(&self) -> DeleteLaunchTemplate<C, M, R>

pub fn delete_launch_template_versions(
&self
) -> DeleteLaunchTemplateVersions<C, M, R>

pub fn delete_local_gateway_route(&self) -> DeleteLocalGatewayRoute<C, M, R>

pub fn delete_local_gateway_route_table_vpc_association(
&self
) -> DeleteLocalGatewayRouteTableVpcAssociation<C, M, R>

pub fn delete_managed_prefix_list(&self) -> DeleteManagedPrefixList<C, M, R>

pub fn delete_nat_gateway(&self) -> DeleteNatGateway<C, M, R>

pub fn delete_network_acl(&self) -> DeleteNetworkAcl<C, M, R>

pub fn delete_network_acl_entry(&self) -> DeleteNetworkAclEntry<C, M, R>

pub fn delete_network_insights_analysis(
&self
) -> DeleteNetworkInsightsAnalysis<C, M, R>

pub fn delete_network_insights_path(&self) -> DeleteNetworkInsightsPath<C, M, R>

pub fn delete_network_interface(&self) -> DeleteNetworkInterface<C, M, R>

pub fn delete_network_interface_permission(
&self
) -> DeleteNetworkInterfacePermission<C, M, R>

pub fn delete_placement_group(&self) -> DeletePlacementGroup<C, M, R>

pub fn delete_queued_reserved_instances(
&self
) -> DeleteQueuedReservedInstances<C, M, R>

pub fn delete_route(&self) -> DeleteRoute<C, M, R>

pub fn delete_route_table(&self) -> DeleteRouteTable<C, M, R>

pub fn delete_security_group(&self) -> DeleteSecurityGroup<C, M, R>

pub fn delete_snapshot(&self) -> DeleteSnapshot<C, M, R>

pub fn delete_spot_datafeed_subscription(
&self
) -> DeleteSpotDatafeedSubscription<C, M, R>

pub fn delete_subnet(&self) -> DeleteSubnet<C, M, R>

pub fn delete_subnet_cidr_reservation(
&self
) -> DeleteSubnetCidrReservation<C, M, R>

pub fn delete_tags(&self) -> DeleteTags<C, M, R>

pub fn delete_traffic_mirror_filter(&self) -> DeleteTrafficMirrorFilter<C, M, R>

pub fn delete_traffic_mirror_filter_rule(
&self
) -> DeleteTrafficMirrorFilterRule<C, M, R>

pub fn delete_traffic_mirror_session(
&self
) -> DeleteTrafficMirrorSession<C, M, R>

pub fn delete_traffic_mirror_target(&self) -> DeleteTrafficMirrorTarget<C, M, R>

pub fn delete_transit_gateway(&self) -> DeleteTransitGateway<C, M, R>

pub fn delete_transit_gateway_connect(
&self
) -> DeleteTransitGatewayConnect<C, M, R>

pub fn delete_transit_gateway_connect_peer(
&self
) -> DeleteTransitGatewayConnectPeer<C, M, R>

pub fn delete_transit_gateway_multicast_domain(
&self
) -> DeleteTransitGatewayMulticastDomain<C, M, R>

pub fn delete_transit_gateway_peering_attachment(
&self
) -> DeleteTransitGatewayPeeringAttachment<C, M, R>

pub fn delete_transit_gateway_prefix_list_reference(
&self
) -> DeleteTransitGatewayPrefixListReference<C, M, R>

pub fn delete_transit_gateway_route(&self) -> DeleteTransitGatewayRoute<C, M, R>

pub fn delete_transit_gateway_route_table(
&self
) -> DeleteTransitGatewayRouteTable<C, M, R>

pub fn delete_transit_gateway_vpc_attachment(
&self
) -> DeleteTransitGatewayVpcAttachment<C, M, R>

pub fn delete_volume(&self) -> DeleteVolume<C, M, R>

pub fn delete_vpc(&self) -> DeleteVpc<C, M, R>

pub fn delete_vpc_endpoint_connection_notifications(
&self
) -> DeleteVpcEndpointConnectionNotifications<C, M, R>

pub fn delete_vpc_endpoints(&self) -> DeleteVpcEndpoints<C, M, R>

pub fn delete_vpc_endpoint_service_configurations(
&self
) -> DeleteVpcEndpointServiceConfigurations<C, M, R>

pub fn delete_vpc_peering_connection(
&self
) -> DeleteVpcPeeringConnection<C, M, R>

pub fn delete_vpn_connection(&self) -> DeleteVpnConnection<C, M, R>

pub fn delete_vpn_connection_route(&self) -> DeleteVpnConnectionRoute<C, M, R>

pub fn delete_vpn_gateway(&self) -> DeleteVpnGateway<C, M, R>

pub fn deprovision_byoip_cidr(&self) -> DeprovisionByoipCidr<C, M, R>

pub fn deregister_image(&self) -> DeregisterImage<C, M, R>

pub fn deregister_instance_event_notification_attributes(
&self
) -> DeregisterInstanceEventNotificationAttributes<C, M, R>

pub fn deregister_transit_gateway_multicast_group_members(
&self
) -> DeregisterTransitGatewayMulticastGroupMembers<C, M, R>

pub fn deregister_transit_gateway_multicast_group_sources(
&self
) -> DeregisterTransitGatewayMulticastGroupSources<C, M, R>

pub fn describe_account_attributes(&self) -> DescribeAccountAttributes<C, M, R>

pub fn describe_addresses(&self) -> DescribeAddresses<C, M, R>

pub fn describe_addresses_attribute(
&self
) -> DescribeAddressesAttribute<C, M, R>

pub fn describe_aggregate_id_format(&self) -> DescribeAggregateIdFormat<C, M, R>

pub fn describe_availability_zones(&self) -> DescribeAvailabilityZones<C, M, R>

pub fn describe_bundle_tasks(&self) -> DescribeBundleTasks<C, M, R>

pub fn describe_byoip_cidrs(&self) -> DescribeByoipCidrs<C, M, R>

pub fn describe_capacity_reservations(
&self
) -> DescribeCapacityReservations<C, M, R>

pub fn describe_carrier_gateways(&self) -> DescribeCarrierGateways<C, M, R>

pub fn describe_classic_link_instances(
&self
) -> DescribeClassicLinkInstances<C, M, R>

pub fn describe_client_vpn_authorization_rules(
&self
) -> DescribeClientVpnAuthorizationRules<C, M, R>

pub fn describe_client_vpn_connections(
&self
) -> DescribeClientVpnConnections<C, M, R>

pub fn describe_client_vpn_endpoints(
&self
) -> DescribeClientVpnEndpoints<C, M, R>

pub fn describe_client_vpn_routes(&self) -> DescribeClientVpnRoutes<C, M, R>

pub fn describe_client_vpn_target_networks(
&self
) -> DescribeClientVpnTargetNetworks<C, M, R>

pub fn describe_coip_pools(&self) -> DescribeCoipPools<C, M, R>

pub fn describe_conversion_tasks(&self) -> DescribeConversionTasks<C, M, R>

pub fn describe_customer_gateways(&self) -> DescribeCustomerGateways<C, M, R>

pub fn describe_dhcp_options(&self) -> DescribeDhcpOptions<C, M, R>

pub fn describe_egress_only_internet_gateways(
&self
) -> DescribeEgressOnlyInternetGateways<C, M, R>

pub fn describe_elastic_gpus(&self) -> DescribeElasticGpus<C, M, R>

pub fn describe_export_image_tasks(&self) -> DescribeExportImageTasks<C, M, R>

pub fn describe_export_tasks(&self) -> DescribeExportTasks<C, M, R>

pub fn describe_fast_snapshot_restores(
&self
) -> DescribeFastSnapshotRestores<C, M, R>

pub fn describe_fleet_history(&self) -> DescribeFleetHistory<C, M, R>

pub fn describe_fleet_instances(&self) -> DescribeFleetInstances<C, M, R>

pub fn describe_fleets(&self) -> DescribeFleets<C, M, R>

pub fn describe_flow_logs(&self) -> DescribeFlowLogs<C, M, R>

pub fn describe_fpga_image_attribute(
&self
) -> DescribeFpgaImageAttribute<C, M, R>

pub fn describe_fpga_images(&self) -> DescribeFpgaImages<C, M, R>

pub fn describe_host_reservation_offerings(
&self
) -> DescribeHostReservationOfferings<C, M, R>

pub fn describe_host_reservations(&self) -> DescribeHostReservations<C, M, R>

pub fn describe_hosts(&self) -> DescribeHosts<C, M, R>

pub fn describe_iam_instance_profile_associations(
&self
) -> DescribeIamInstanceProfileAssociations<C, M, R>

pub fn describe_identity_id_format(&self) -> DescribeIdentityIdFormat<C, M, R>

pub fn describe_id_format(&self) -> DescribeIdFormat<C, M, R>

pub fn describe_image_attribute(&self) -> DescribeImageAttribute<C, M, R>

pub fn describe_images(&self) -> DescribeImages<C, M, R>

pub fn describe_import_image_tasks(&self) -> DescribeImportImageTasks<C, M, R>

pub fn describe_import_snapshot_tasks(
&self
) -> DescribeImportSnapshotTasks<C, M, R>

pub fn describe_instance_attribute(&self) -> DescribeInstanceAttribute<C, M, R>

pub fn describe_instance_credit_specifications(
&self
) -> DescribeInstanceCreditSpecifications<C, M, R>

pub fn describe_instance_event_notification_attributes(
&self
) -> DescribeInstanceEventNotificationAttributes<C, M, R>

pub fn describe_instance_event_windows(
&self
) -> DescribeInstanceEventWindows<C, M, R>

pub fn describe_instances(&self) -> DescribeInstances<C, M, R>

pub fn describe_instance_status(&self) -> DescribeInstanceStatus<C, M, R>

pub fn describe_instance_type_offerings(
&self
) -> DescribeInstanceTypeOfferings<C, M, R>

pub fn describe_instance_types(&self) -> DescribeInstanceTypes<C, M, R>

pub fn describe_internet_gateways(&self) -> DescribeInternetGateways<C, M, R>

pub fn describe_ipv6_pools(&self) -> DescribeIpv6Pools<C, M, R>

pub fn describe_key_pairs(&self) -> DescribeKeyPairs<C, M, R>

pub fn describe_launch_templates(&self) -> DescribeLaunchTemplates<C, M, R>

pub fn describe_launch_template_versions(
&self
) -> DescribeLaunchTemplateVersions<C, M, R>

pub fn describe_local_gateway_route_tables(
&self
) -> DescribeLocalGatewayRouteTables<C, M, R>

pub fn describe_local_gateway_route_table_virtual_interface_group_associations(
&self
) -> DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations<C, M, R>

pub fn describe_local_gateway_route_table_vpc_associations(
&self
) -> DescribeLocalGatewayRouteTableVpcAssociations<C, M, R>

pub fn describe_local_gateways(&self) -> DescribeLocalGateways<C, M, R>

pub fn describe_local_gateway_virtual_interface_groups(
&self
) -> DescribeLocalGatewayVirtualInterfaceGroups<C, M, R>

pub fn describe_local_gateway_virtual_interfaces(
&self
) -> DescribeLocalGatewayVirtualInterfaces<C, M, R>

pub fn describe_managed_prefix_lists(
&self
) -> DescribeManagedPrefixLists<C, M, R>

pub fn describe_moving_addresses(&self) -> DescribeMovingAddresses<C, M, R>

pub fn describe_nat_gateways(&self) -> DescribeNatGateways<C, M, R>

pub fn describe_network_acls(&self) -> DescribeNetworkAcls<C, M, R>

pub fn describe_network_insights_analyses(
&self
) -> DescribeNetworkInsightsAnalyses<C, M, R>

pub fn describe_network_insights_paths(
&self
) -> DescribeNetworkInsightsPaths<C, M, R>

pub fn describe_network_interface_attribute(
&self
) -> DescribeNetworkInterfaceAttribute<C, M, R>

pub fn describe_network_interface_permissions(
&self
) -> DescribeNetworkInterfacePermissions<C, M, R>

pub fn describe_network_interfaces(&self) -> DescribeNetworkInterfaces<C, M, R>

pub fn describe_placement_groups(&self) -> DescribePlacementGroups<C, M, R>

pub fn describe_prefix_lists(&self) -> DescribePrefixLists<C, M, R>

pub fn describe_principal_id_format(&self) -> DescribePrincipalIdFormat<C, M, R>

pub fn describe_public_ipv4_pools(&self) -> DescribePublicIpv4Pools<C, M, R>

pub fn describe_regions(&self) -> DescribeRegions<C, M, R>

pub fn describe_replace_root_volume_tasks(
&self
) -> DescribeReplaceRootVolumeTasks<C, M, R>

pub fn describe_reserved_instances(&self) -> DescribeReservedInstances<C, M, R>

pub fn describe_reserved_instances_listings(
&self
) -> DescribeReservedInstancesListings<C, M, R>

pub fn describe_reserved_instances_modifications(
&self
) -> DescribeReservedInstancesModifications<C, M, R>

pub fn describe_reserved_instances_offerings(
&self
) -> DescribeReservedInstancesOfferings<C, M, R>

pub fn describe_route_tables(&self) -> DescribeRouteTables<C, M, R>

pub fn describe_scheduled_instance_availability(
&self
) -> DescribeScheduledInstanceAvailability<C, M, R>

pub fn describe_scheduled_instances(
&self
) -> DescribeScheduledInstances<C, M, R>

pub fn describe_security_group_references(
&self
) -> DescribeSecurityGroupReferences<C, M, R>

pub fn describe_security_group_rules(
&self
) -> DescribeSecurityGroupRules<C, M, R>

pub fn describe_security_groups(&self) -> DescribeSecurityGroups<C, M, R>

pub fn describe_snapshot_attribute(&self) -> DescribeSnapshotAttribute<C, M, R>

pub fn describe_snapshots(&self) -> DescribeSnapshots<C, M, R>

pub fn describe_spot_datafeed_subscription(
&self
) -> DescribeSpotDatafeedSubscription<C, M, R>

pub fn describe_spot_fleet_instances(
&self
) -> DescribeSpotFleetInstances<C, M, R>

pub fn describe_spot_fleet_request_history(
&self
) -> DescribeSpotFleetRequestHistory<C, M, R>

pub fn describe_spot_fleet_requests(&self) -> DescribeSpotFleetRequests<C, M, R>

pub fn describe_spot_instance_requests(
&self
) -> DescribeSpotInstanceRequests<C, M, R>

pub fn describe_spot_price_history(&self) -> DescribeSpotPriceHistory<C, M, R>

pub fn describe_stale_security_groups(
&self
) -> DescribeStaleSecurityGroups<C, M, R>

pub fn describe_store_image_tasks(&self) -> DescribeStoreImageTasks<C, M, R>

pub fn describe_subnets(&self) -> DescribeSubnets<C, M, R>

pub fn describe_tags(&self) -> DescribeTags<C, M, R>

pub fn describe_traffic_mirror_filters(
&self
) -> DescribeTrafficMirrorFilters<C, M, R>

pub fn describe_traffic_mirror_sessions(
&self
) -> DescribeTrafficMirrorSessions<C, M, R>

pub fn describe_traffic_mirror_targets(
&self
) -> DescribeTrafficMirrorTargets<C, M, R>

pub fn describe_transit_gateway_attachments(
&self
) -> DescribeTransitGatewayAttachments<C, M, R>

pub fn describe_transit_gateway_connect_peers(
&self
) -> DescribeTransitGatewayConnectPeers<C, M, R>

pub fn describe_transit_gateway_connects(
&self
) -> DescribeTransitGatewayConnects<C, M, R>

pub fn describe_transit_gateway_multicast_domains(
&self
) -> DescribeTransitGatewayMulticastDomains<C, M, R>

pub fn describe_transit_gateway_peering_attachments(
&self
) -> DescribeTransitGatewayPeeringAttachments<C, M, R>

pub fn describe_transit_gateway_route_tables(
&self
) -> DescribeTransitGatewayRouteTables<C, M, R>

pub fn describe_transit_gateways(&self) -> DescribeTransitGateways<C, M, R>

pub fn describe_transit_gateway_vpc_attachments(
&self
) -> DescribeTransitGatewayVpcAttachments<C, M, R>

pub fn describe_trunk_interface_associations(
&self
) -> DescribeTrunkInterfaceAssociations<C, M, R>

pub fn describe_volume_attribute(&self) -> DescribeVolumeAttribute<C, M, R>

pub fn describe_volumes(&self) -> DescribeVolumes<C, M, R>

pub fn describe_volumes_modifications(
&self
) -> DescribeVolumesModifications<C, M, R>

pub fn describe_volume_status(&self) -> DescribeVolumeStatus<C, M, R>

pub fn describe_vpc_attribute(&self) -> DescribeVpcAttribute<C, M, R>

pub fn describe_vpc_classic_link(&self) -> DescribeVpcClassicLink<C, M, R>

pub fn describe_vpc_classic_link_dns_support(
&self
) -> DescribeVpcClassicLinkDnsSupport<C, M, R>

pub fn describe_vpc_endpoint_connection_notifications(
&self
) -> DescribeVpcEndpointConnectionNotifications<C, M, R>

pub fn describe_vpc_endpoint_connections(
&self
) -> DescribeVpcEndpointConnections<C, M, R>

pub fn describe_vpc_endpoints(&self) -> DescribeVpcEndpoints<C, M, R>

pub fn describe_vpc_endpoint_service_configurations(
&self
) -> DescribeVpcEndpointServiceConfigurations<C, M, R>

pub fn describe_vpc_endpoint_service_permissions(
&self
) -> DescribeVpcEndpointServicePermissions<C, M, R>

pub fn describe_vpc_endpoint_services(
&self
) -> DescribeVpcEndpointServices<C, M, R>

pub fn describe_vpc_peering_connections(
&self
) -> DescribeVpcPeeringConnections<C, M, R>

pub fn describe_vpcs(&self) -> DescribeVpcs<C, M, R>

pub fn describe_vpn_connections(&self) -> DescribeVpnConnections<C, M, R>

pub fn describe_vpn_gateways(&self) -> DescribeVpnGateways<C, M, R>

pub fn detach_classic_link_vpc(&self) -> DetachClassicLinkVpc<C, M, R>

pub fn detach_internet_gateway(&self) -> DetachInternetGateway<C, M, R>

pub fn detach_network_interface(&self) -> DetachNetworkInterface<C, M, R>

pub fn detach_volume(&self) -> DetachVolume<C, M, R>

pub fn detach_vpn_gateway(&self) -> DetachVpnGateway<C, M, R>

pub fn disable_ebs_encryption_by_default(
&self
) -> DisableEbsEncryptionByDefault<C, M, R>

pub fn disable_fast_snapshot_restores(
&self
) -> DisableFastSnapshotRestores<C, M, R>

pub fn disable_image_deprecation(&self) -> DisableImageDeprecation<C, M, R>

pub fn disable_serial_console_access(
&self
) -> DisableSerialConsoleAccess<C, M, R>

pub fn disable_transit_gateway_route_table_propagation(
&self
) -> DisableTransitGatewayRouteTablePropagation<C, M, R>

pub fn disable_vgw_route_propagation(
&self
) -> DisableVgwRoutePropagation<C, M, R>

pub fn disable_vpc_classic_link(&self) -> DisableVpcClassicLink<C, M, R>

pub fn disable_vpc_classic_link_dns_support(
&self
) -> DisableVpcClassicLinkDnsSupport<C, M, R>

pub fn disassociate_address(&self) -> DisassociateAddress<C, M, R>

pub fn disassociate_client_vpn_target_network(
&self
) -> DisassociateClientVpnTargetNetwork<C, M, R>

pub fn disassociate_enclave_certificate_iam_role(
&self
) -> DisassociateEnclaveCertificateIamRole<C, M, R>

pub fn disassociate_iam_instance_profile(
&self
) -> DisassociateIamInstanceProfile<C, M, R>

pub fn disassociate_instance_event_window(
&self
) -> DisassociateInstanceEventWindow<C, M, R>

pub fn disassociate_route_table(&self) -> DisassociateRouteTable<C, M, R>

pub fn disassociate_subnet_cidr_block(
&self
) -> DisassociateSubnetCidrBlock<C, M, R>

pub fn disassociate_transit_gateway_multicast_domain(
&self
) -> DisassociateTransitGatewayMulticastDomain<C, M, R>

pub fn disassociate_transit_gateway_route_table(
&self
) -> DisassociateTransitGatewayRouteTable<C, M, R>

pub fn disassociate_trunk_interface(
&self
) -> DisassociateTrunkInterface<C, M, R>

pub fn disassociate_vpc_cidr_block(&self) -> DisassociateVpcCidrBlock<C, M, R>

pub fn enable_ebs_encryption_by_default(
&self
) -> EnableEbsEncryptionByDefault<C, M, R>

pub fn enable_fast_snapshot_restores(
&self
) -> EnableFastSnapshotRestores<C, M, R>

pub fn enable_image_deprecation(&self) -> EnableImageDeprecation<C, M, R>

pub fn enable_serial_console_access(&self) -> EnableSerialConsoleAccess<C, M, R>

pub fn enable_transit_gateway_route_table_propagation(
&self
) -> EnableTransitGatewayRouteTablePropagation<C, M, R>

pub fn enable_vgw_route_propagation(&self) -> EnableVgwRoutePropagation<C, M, R>

pub fn enable_volume_io(&self) -> EnableVolumeIO<C, M, R>

pub fn enable_vpc_classic_link(&self) -> EnableVpcClassicLink<C, M, R>

pub fn enable_vpc_classic_link_dns_support(
&self
) -> EnableVpcClassicLinkDnsSupport<C, M, R>

pub fn export_client_vpn_client_certificate_revocation_list(
&self
) -> ExportClientVpnClientCertificateRevocationList<C, M, R>

pub fn export_client_vpn_client_configuration(
&self
) -> ExportClientVpnClientConfiguration<C, M, R>

pub fn export_image(&self) -> ExportImage<C, M, R>

pub fn export_transit_gateway_routes(
&self
) -> ExportTransitGatewayRoutes<C, M, R>

pub fn get_associated_enclave_certificate_iam_roles(
&self
) -> GetAssociatedEnclaveCertificateIamRoles<C, M, R>

pub fn get_associated_ipv6_pool_cidrs(
&self
) -> GetAssociatedIpv6PoolCidrs<C, M, R>

pub fn get_capacity_reservation_usage(
&self
) -> GetCapacityReservationUsage<C, M, R>

pub fn get_coip_pool_usage(&self) -> GetCoipPoolUsage<C, M, R>

pub fn get_console_output(&self) -> GetConsoleOutput<C, M, R>

pub fn get_console_screenshot(&self) -> GetConsoleScreenshot<C, M, R>

pub fn get_default_credit_specification(
&self
) -> GetDefaultCreditSpecification<C, M, R>

pub fn get_ebs_default_kms_key_id(&self) -> GetEbsDefaultKmsKeyId<C, M, R>

pub fn get_ebs_encryption_by_default(
&self
) -> GetEbsEncryptionByDefault<C, M, R>

pub fn get_flow_logs_integration_template(
&self
) -> GetFlowLogsIntegrationTemplate<C, M, R>

pub fn get_groups_for_capacity_reservation(
&self
) -> GetGroupsForCapacityReservation<C, M, R>

pub fn get_host_reservation_purchase_preview(
&self
) -> GetHostReservationPurchasePreview<C, M, R>

pub fn get_launch_template_data(&self) -> GetLaunchTemplateData<C, M, R>

pub fn get_managed_prefix_list_associations(
&self
) -> GetManagedPrefixListAssociations<C, M, R>

pub fn get_managed_prefix_list_entries(
&self
) -> GetManagedPrefixListEntries<C, M, R>

pub fn get_password_data(&self) -> GetPasswordData<C, M, R>

pub fn get_reserved_instances_exchange_quote(
&self
) -> GetReservedInstancesExchangeQuote<C, M, R>

pub fn get_serial_console_access_status(
&self
) -> GetSerialConsoleAccessStatus<C, M, R>

pub fn get_subnet_cidr_reservations(&self) -> GetSubnetCidrReservations<C, M, R>

pub fn get_transit_gateway_attachment_propagations(
&self
) -> GetTransitGatewayAttachmentPropagations<C, M, R>

pub fn get_transit_gateway_multicast_domain_associations(
&self
) -> GetTransitGatewayMulticastDomainAssociations<C, M, R>

pub fn get_transit_gateway_prefix_list_references(
&self
) -> GetTransitGatewayPrefixListReferences<C, M, R>

pub fn get_transit_gateway_route_table_associations(
&self
) -> GetTransitGatewayRouteTableAssociations<C, M, R>

pub fn get_transit_gateway_route_table_propagations(
&self
) -> GetTransitGatewayRouteTablePropagations<C, M, R>

pub fn import_client_vpn_client_certificate_revocation_list(
&self
) -> ImportClientVpnClientCertificateRevocationList<C, M, R>

pub fn import_image(&self) -> ImportImage<C, M, R>

pub fn import_instance(&self) -> ImportInstance<C, M, R>

pub fn import_key_pair(&self) -> ImportKeyPair<C, M, R>

pub fn import_snapshot(&self) -> ImportSnapshot<C, M, R>

pub fn import_volume(&self) -> ImportVolume<C, M, R>

pub fn modify_address_attribute(&self) -> ModifyAddressAttribute<C, M, R>

pub fn modify_availability_zone_group(
&self
) -> ModifyAvailabilityZoneGroup<C, M, R>

pub fn modify_capacity_reservation(&self) -> ModifyCapacityReservation<C, M, R>

pub fn modify_client_vpn_endpoint(&self) -> ModifyClientVpnEndpoint<C, M, R>

pub fn modify_default_credit_specification(
&self
) -> ModifyDefaultCreditSpecification<C, M, R>

pub fn modify_ebs_default_kms_key_id(&self) -> ModifyEbsDefaultKmsKeyId<C, M, R>

pub fn modify_fleet(&self) -> ModifyFleet<C, M, R>

pub fn modify_fpga_image_attribute(&self) -> ModifyFpgaImageAttribute<C, M, R>

pub fn modify_hosts(&self) -> ModifyHosts<C, M, R>

pub fn modify_identity_id_format(&self) -> ModifyIdentityIdFormat<C, M, R>

pub fn modify_id_format(&self) -> ModifyIdFormat<C, M, R>

pub fn modify_image_attribute(&self) -> ModifyImageAttribute<C, M, R>

pub fn modify_instance_attribute(&self) -> ModifyInstanceAttribute<C, M, R>

pub fn modify_instance_capacity_reservation_attributes(
&self
) -> ModifyInstanceCapacityReservationAttributes<C, M, R>

pub fn modify_instance_credit_specification(
&self
) -> ModifyInstanceCreditSpecification<C, M, R>

pub fn modify_instance_event_start_time(
&self
) -> ModifyInstanceEventStartTime<C, M, R>

pub fn modify_instance_event_window(&self) -> ModifyInstanceEventWindow<C, M, R>

pub fn modify_instance_metadata_options(
&self
) -> ModifyInstanceMetadataOptions<C, M, R>

pub fn modify_instance_placement(&self) -> ModifyInstancePlacement<C, M, R>

pub fn modify_launch_template(&self) -> ModifyLaunchTemplate<C, M, R>

pub fn modify_managed_prefix_list(&self) -> ModifyManagedPrefixList<C, M, R>

pub fn modify_network_interface_attribute(
&self
) -> ModifyNetworkInterfaceAttribute<C, M, R>

pub fn modify_reserved_instances(&self) -> ModifyReservedInstances<C, M, R>

pub fn modify_security_group_rules(&self) -> ModifySecurityGroupRules<C, M, R>

pub fn modify_snapshot_attribute(&self) -> ModifySnapshotAttribute<C, M, R>

pub fn modify_spot_fleet_request(&self) -> ModifySpotFleetRequest<C, M, R>

pub fn modify_subnet_attribute(&self) -> ModifySubnetAttribute<C, M, R>

pub fn modify_traffic_mirror_filter_network_services(
&self
) -> ModifyTrafficMirrorFilterNetworkServices<C, M, R>

pub fn modify_traffic_mirror_filter_rule(
&self
) -> ModifyTrafficMirrorFilterRule<C, M, R>

pub fn modify_traffic_mirror_session(
&self
) -> ModifyTrafficMirrorSession<C, M, R>

pub fn modify_transit_gateway(&self) -> ModifyTransitGateway<C, M, R>

pub fn modify_transit_gateway_prefix_list_reference(
&self
) -> ModifyTransitGatewayPrefixListReference<C, M, R>

pub fn modify_transit_gateway_vpc_attachment(
&self
) -> ModifyTransitGatewayVpcAttachment<C, M, R>

pub fn modify_volume(&self) -> ModifyVolume<C, M, R>

pub fn modify_volume_attribute(&self) -> ModifyVolumeAttribute<C, M, R>

pub fn modify_vpc_attribute(&self) -> ModifyVpcAttribute<C, M, R>

pub fn modify_vpc_endpoint(&self) -> ModifyVpcEndpoint<C, M, R>

pub fn modify_vpc_endpoint_connection_notification(
&self
) -> ModifyVpcEndpointConnectionNotification<C, M, R>

pub fn modify_vpc_endpoint_service_configuration(
&self
) -> ModifyVpcEndpointServiceConfiguration<C, M, R>

pub fn modify_vpc_endpoint_service_permissions(
&self
) -> ModifyVpcEndpointServicePermissions<C, M, R>

pub fn modify_vpc_peering_connection_options(
&self
) -> ModifyVpcPeeringConnectionOptions<C, M, R>

pub fn modify_vpc_tenancy(&self) -> ModifyVpcTenancy<C, M, R>

pub fn modify_vpn_connection(&self) -> ModifyVpnConnection<C, M, R>

pub fn modify_vpn_connection_options(
&self
) -> ModifyVpnConnectionOptions<C, M, R>

pub fn modify_vpn_tunnel_certificate(
&self
) -> ModifyVpnTunnelCertificate<C, M, R>

pub fn modify_vpn_tunnel_options(&self) -> ModifyVpnTunnelOptions<C, M, R>

pub fn monitor_instances(&self) -> MonitorInstances<C, M, R>

pub fn move_address_to_vpc(&self) -> MoveAddressToVpc<C, M, R>

pub fn provision_byoip_cidr(&self) -> ProvisionByoipCidr<C, M, R>

pub fn purchase_host_reservation(&self) -> PurchaseHostReservation<C, M, R>

pub fn purchase_reserved_instances_offering(
&self
) -> PurchaseReservedInstancesOffering<C, M, R>

pub fn purchase_scheduled_instances(
&self
) -> PurchaseScheduledInstances<C, M, R>

pub fn reboot_instances(&self) -> RebootInstances<C, M, R>

pub fn register_image(&self) -> RegisterImage<C, M, R>

pub fn register_instance_event_notification_attributes(
&self
) -> RegisterInstanceEventNotificationAttributes<C, M, R>

pub fn register_transit_gateway_multicast_group_members(
&self
) -> RegisterTransitGatewayMulticastGroupMembers<C, M, R>

pub fn register_transit_gateway_multicast_group_sources(
&self
) -> RegisterTransitGatewayMulticastGroupSources<C, M, R>

pub fn reject_transit_gateway_multicast_domain_associations(
&self
) -> RejectTransitGatewayMulticastDomainAssociations<C, M, R>

pub fn reject_transit_gateway_peering_attachment(
&self
) -> RejectTransitGatewayPeeringAttachment<C, M, R>

pub fn reject_transit_gateway_vpc_attachment(
&self
) -> RejectTransitGatewayVpcAttachment<C, M, R>

pub fn reject_vpc_endpoint_connections(
&self
) -> RejectVpcEndpointConnections<C, M, R>

pub fn reject_vpc_peering_connection(
&self
) -> RejectVpcPeeringConnection<C, M, R>

pub fn release_address(&self) -> ReleaseAddress<C, M, R>

pub fn release_hosts(&self) -> ReleaseHosts<C, M, R>

pub fn replace_iam_instance_profile_association(
&self
) -> ReplaceIamInstanceProfileAssociation<C, M, R>

pub fn replace_network_acl_association(
&self
) -> ReplaceNetworkAclAssociation<C, M, R>

pub fn replace_network_acl_entry(&self) -> ReplaceNetworkAclEntry<C, M, R>

pub fn replace_route(&self) -> ReplaceRoute<C, M, R>

pub fn replace_route_table_association(
&self
) -> ReplaceRouteTableAssociation<C, M, R>

pub fn replace_transit_gateway_route(
&self
) -> ReplaceTransitGatewayRoute<C, M, R>

pub fn report_instance_status(&self) -> ReportInstanceStatus<C, M, R>

pub fn request_spot_fleet(&self) -> RequestSpotFleet<C, M, R>

pub fn request_spot_instances(&self) -> RequestSpotInstances<C, M, R>

pub fn reset_address_attribute(&self) -> ResetAddressAttribute<C, M, R>

pub fn reset_ebs_default_kms_key_id(&self) -> ResetEbsDefaultKmsKeyId<C, M, R>

pub fn reset_fpga_image_attribute(&self) -> ResetFpgaImageAttribute<C, M, R>

pub fn reset_image_attribute(&self) -> ResetImageAttribute<C, M, R>

pub fn reset_instance_attribute(&self) -> ResetInstanceAttribute<C, M, R>

pub fn reset_network_interface_attribute(
&self
) -> ResetNetworkInterfaceAttribute<C, M, R>

pub fn reset_snapshot_attribute(&self) -> ResetSnapshotAttribute<C, M, R>

pub fn restore_address_to_classic(&self) -> RestoreAddressToClassic<C, M, R>

pub fn restore_managed_prefix_list_version(
&self
) -> RestoreManagedPrefixListVersion<C, M, R>

pub fn revoke_client_vpn_ingress(&self) -> RevokeClientVpnIngress<C, M, R>

pub fn revoke_security_group_egress(&self) -> RevokeSecurityGroupEgress<C, M, R>

pub fn revoke_security_group_ingress(
&self
) -> RevokeSecurityGroupIngress<C, M, R>

pub fn run_instances(&self) -> RunInstances<C, M, R>

pub fn run_scheduled_instances(&self) -> RunScheduledInstances<C, M, R>

pub fn search_local_gateway_routes(&self) -> SearchLocalGatewayRoutes<C, M, R>

pub fn search_transit_gateway_multicast_groups(
&self
) -> SearchTransitGatewayMulticastGroups<C, M, R>

pub fn search_transit_gateway_routes(
&self
) -> SearchTransitGatewayRoutes<C, M, R>

pub fn send_diagnostic_interrupt(&self) -> SendDiagnosticInterrupt<C, M, R>

pub fn start_instances(&self) -> StartInstances<C, M, R>

pub fn start_network_insights_analysis(
&self
) -> StartNetworkInsightsAnalysis<C, M, R>

pub fn start_vpc_endpoint_service_private_dns_verification(
&self
) -> StartVpcEndpointServicePrivateDnsVerification<C, M, R>

pub fn stop_instances(&self) -> StopInstances<C, M, R>

pub fn terminate_client_vpn_connections(
&self
) -> TerminateClientVpnConnections<C, M, R>

pub fn terminate_instances(&self) -> TerminateInstances<C, M, R>

pub fn unassign_ipv6_addresses(&self) -> UnassignIpv6Addresses<C, M, R>

pub fn unassign_private_ip_addresses(
&self
) -> UnassignPrivateIpAddresses<C, M, R>

pub fn unmonitor_instances(&self) -> UnmonitorInstances<C, M, R>

pub fn update_security_group_rule_descriptions_egress(
&self
) -> UpdateSecurityGroupRuleDescriptionsEgress<C, M, R>

pub fn update_security_group_rule_descriptions_ingress(
&self
) -> UpdateSecurityGroupRuleDescriptionsIngress<C, M, R>

pub fn withdraw_byoip_cidr(&self) -> WithdrawByoipCidr<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self