Struct aws_sdk_ec2::input::create_fleet_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateFleetInput

Implementations

impl Builder

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn spot_options(self, input: SpotOptionsRequest) -> Self

Describes the configuration of Spot Instances in an EC2 Fleet.

pub fn set_spot_options(self, input: Option<SpotOptionsRequest>) -> Self

pub fn on_demand_options(self, input: OnDemandOptionsRequest) -> Self

Describes the configuration of On-Demand Instances in an EC2 Fleet.

pub fn set_on_demand_options(
self,
input: Option<OnDemandOptionsRequest>
) -> Self

pub fn excess_capacity_termination_policy(
self,
input: FleetExcessCapacityTerminationPolicy
) -> Self

Indicates whether running instances should be terminated if the total target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

pub fn set_excess_capacity_termination_policy(
self,
input: Option<FleetExcessCapacityTerminationPolicy>
) -> Self

pub fn launch_template_configs(
self,
input: impl Into<FleetLaunchTemplateConfigRequest>
) -> Self

pub fn set_launch_template_configs(
self,
input: Option<Vec<FleetLaunchTemplateConfigRequest>>
) -> Self

pub fn target_capacity_specification(
self,
input: TargetCapacitySpecificationRequest
) -> Self

The number of units to request.

pub fn set_target_capacity_specification(
self,
input: Option<TargetCapacitySpecificationRequest>
) -> Self

pub fn terminate_instances_with_expiration(self, input: bool) -> Self

Indicates whether running instances should be terminated when the EC2 Fleet expires.

pub fn set_terminate_instances_with_expiration(
self,
input: Option<bool>
) -> Self

pub fn type(self, input: FleetType) -> Self

The fleet type. The default value is maintain.

maintain - The EC2 Fleet places an asynchronous request for your desired capacity, and continues to maintain your desired Spot capacity by replenishing interrupted Spot Instances.

request - The EC2 Fleet places an asynchronous one-time request for your desired capacity, but does submit Spot requests in alternative capacity pools if Spot capacity is unavailable, and does not maintain Spot capacity if Spot Instances are interrupted.

instant - The EC2 Fleet places a synchronous one-time request for your desired capacity, and returns errors for any instances that could not be launched.

For more information, see EC2 Fleet request types in the Amazon EC2 User Guide.

pub fn set_type(self, input: Option<FleetType>) -> Self

pub fn valid_from(self, input: Instant) -> Self

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

pub fn set_valid_from(self, input: Option<Instant>) -> Self

pub fn valid_until(self, input: Instant) -> Self

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new EC2 Fleet requests are placed or able to fulfill the request. If no value is specified, the request remains until you cancel it.

pub fn set_valid_until(self, input: Option<Instant>) -> Self

pub fn replace_unhealthy_instances(self, input: bool) -> Self

Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

pub fn set_replace_unhealthy_instances(self, input: Option<bool>) -> Self

pub fn tag_specifications(self, input: impl Into<TagSpecification>) -> Self

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn context(self, input: impl Into<String>) -> Self

Reserved.

pub fn set_context(self, input: Option<String>) -> Self

pub fn build(self) -> Result<CreateFleetInput, BuildError>

Consumes the builder and constructs a CreateFleetInput