Struct aws_sdk_ec2::input::create_vpc_endpoint_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateVpcEndpointInput

Implementations

impl Builder

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn vpc_endpoint_type(self, input: VpcEndpointType) -> Self

The type of endpoint.

Default: Gateway

pub fn set_vpc_endpoint_type(self, input: Option<VpcEndpointType>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The ID of the VPC in which the endpoint will be used.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn service_name(self, input: impl Into<String>) -> Self

The service name. To get a list of available services, use the DescribeVpcEndpointServices request, or get the name from the service provider.

pub fn set_service_name(self, input: Option<String>) -> Self

pub fn policy_document(self, input: impl Into<String>) -> Self

(Interface and gateway endpoints) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format. If this parameter is not specified, we attach a default policy that allows full access to the service.

pub fn set_policy_document(self, input: Option<String>) -> Self

pub fn route_table_ids(self, input: impl Into<String>) -> Self

pub fn set_route_table_ids(self, input: Option<Vec<String>>) -> Self

pub fn subnet_ids(self, input: impl Into<String>) -> Self

pub fn set_subnet_ids(self, input: Option<Vec<String>>) -> Self

pub fn security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn private_dns_enabled(self, input: bool) -> Self

(Interface endpoint) Indicates whether to associate a private hosted zone with the specified VPC. The private hosted zone contains a record set for the default public DNS name for the service for the Region (for example, kinesis.us-east-1.amazonaws.com), which resolves to the private IP addresses of the endpoint network interfaces in the VPC. This enables you to make requests to the default public DNS name for the service instead of the public DNS names that are automatically generated by the VPC endpoint service.

To use a private hosted zone, you must set the following VPC attributes to true: enableDnsHostnames and enableDnsSupport. Use ModifyVpcAttribute to set the VPC attributes.

Default: true

pub fn set_private_dns_enabled(self, input: Option<bool>) -> Self

pub fn tag_specifications(self, input: impl Into<TagSpecification>) -> Self

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn build(self) -> Result<CreateVpcEndpointInput, BuildError>

Consumes the builder and constructs a CreateVpcEndpointInput