Struct aws_sdk_ec2::input::import_image_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ImportImageInput

Implementations

impl Builder

pub fn architecture(self, input: impl Into<String>) -> Self

The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

pub fn set_architecture(self, input: Option<String>) -> Self

pub fn client_data(self, input: ClientData) -> Self

The client-specific data.

pub fn set_client_data(self, input: Option<ClientData>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

The token to enable idempotency for VM import requests.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description string for the import image task.

pub fn set_description(self, input: Option<String>) -> Self

pub fn disk_containers(self, input: impl Into<ImageDiskContainer>) -> Self

pub fn set_disk_containers(self, input: Option<Vec<ImageDiskContainer>>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn encrypted(self, input: bool) -> Self

Specifies whether the destination AMI of the imported image should be encrypted. The default CMK for EBS is used unless you specify a non-default AWS Key Management Service (AWS KMS) CMK using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

pub fn set_encrypted(self, input: Option<bool>) -> Self

pub fn hypervisor(self, input: impl Into<String>) -> Self

The target hypervisor platform.

Valid values: xen

pub fn set_hypervisor(self, input: Option<String>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

An identifier for the symmetric AWS Key Management Service (AWS KMS) customer master key (CMK) to use when creating the encrypted AMI. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The CMK identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the CMK, the AWS account ID of the CMK owner, the alias namespace, and then the CMK alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the CMK, the AWS account ID of the CMK owner, the alias namespace, and then the CMK alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

AWS parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified CMK must exist in the Region that the AMI is being copied to.

Amazon EBS does not support asymmetric CMKs.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn license_type(self, input: impl Into<String>) -> Self

The license type to be used for the Amazon Machine Image (AMI) after importing.

By default, we detect the source-system operating system (OS) and apply the appropriate license. Specify AWS to replace the source-system license with an AWS license, if appropriate. Specify BYOL to retain the source-system license, if appropriate.

To use BYOL, you must have existing licenses with rights to use these licenses in a third party cloud, such as AWS. For more information, see Prerequisites in the VM Import/Export User Guide.

pub fn set_license_type(self, input: Option<String>) -> Self

pub fn platform(self, input: impl Into<String>) -> Self

The operating system of the virtual machine.

Valid values: Windows | Linux

pub fn set_platform(self, input: Option<String>) -> Self

pub fn role_name(self, input: impl Into<String>) -> Self

The name of the role to use when not using the default role, 'vmimport'.

pub fn set_role_name(self, input: Option<String>) -> Self

pub fn license_specifications(
self,
input: impl Into<ImportImageLicenseConfigurationRequest>
) -> Self

pub fn set_license_specifications(
self,
input: Option<Vec<ImportImageLicenseConfigurationRequest>>
) -> Self

pub fn tag_specifications(self, input: impl Into<TagSpecification>) -> Self

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn build(self) -> Result<ImportImageInput, BuildError>

Consumes the builder and constructs a ImportImageInput