Struct aws_sdk_ec2::input::run_instances_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for RunInstancesInput

Implementations

impl Builder

pub fn block_device_mappings(self, input: impl Into<BlockDeviceMapping>) -> Self

pub fn set_block_device_mappings(
self,
input: Option<Vec<BlockDeviceMapping>>
) -> Self

pub fn image_id(self, input: impl Into<String>) -> Self

The ID of the AMI. An AMI ID is required to launch an instance and must be specified here or in a launch template.

pub fn set_image_id(self, input: Option<String>) -> Self

pub fn instance_type(self, input: InstanceType) -> Self

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

Default: m1.small

pub fn set_instance_type(self, input: Option<InstanceType>) -> Self

pub fn ipv6_address_count(self, input: i32) -> Self

[EC2-VPC] The number of IPv6 addresses to associate with the primary network interface. Amazon EC2 chooses the IPv6 addresses from the range of your subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

You cannot specify this option and the network interfaces option in the same request.

pub fn set_ipv6_address_count(self, input: Option<i32>) -> Self

pub fn ipv6_addresses(self, input: impl Into<InstanceIpv6Address>) -> Self

pub fn set_ipv6_addresses(self, input: Option<Vec<InstanceIpv6Address>>) -> Self

pub fn kernel_id(self, input: impl Into<String>) -> Self

The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

pub fn set_kernel_id(self, input: Option<String>) -> Self

pub fn key_name(self, input: impl Into<String>) -> Self

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

pub fn set_key_name(self, input: Option<String>) -> Self

pub fn max_count(self, input: i32) -> Self

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 FAQ.

pub fn set_max_count(self, input: Option<i32>) -> Self

pub fn min_count(self, input: i32) -> Self

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

pub fn set_min_count(self, input: Option<i32>) -> Self

pub fn monitoring(self, input: RunInstancesMonitoringEnabled) -> Self

Specifies whether detailed monitoring is enabled for the instance.

pub fn set_monitoring(
self,
input: Option<RunInstancesMonitoringEnabled>
) -> Self

pub fn placement(self, input: Placement) -> Self

The placement for the instance.

pub fn set_placement(self, input: Option<Placement>) -> Self

pub fn ramdisk_id(self, input: impl Into<String>) -> Self

The ID of the RAM disk to select. Some kernels require additional drivers at launch. Check the kernel requirements for information about whether you need to specify a RAM disk. To find kernel requirements, go to the Amazon Web Services Resource Center and search for the kernel ID.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon EC2 User Guide.

pub fn set_ramdisk_id(self, input: Option<String>) -> Self

pub fn security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn security_groups(self, input: impl Into<String>) -> Self

pub fn set_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn subnet_id(self, input: impl Into<String>) -> Self

[EC2-VPC] The ID of the subnet to launch the instance into.

If you specify a network interface, you must specify any subnets as part of the network interface.

pub fn set_subnet_id(self, input: Option<String>) -> Self

pub fn user_data(self, input: impl Into<String>) -> Self

The user data to make available to the instance. For more information, see Running commands on your Linux instance at launch (Linux) and Adding User Data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

pub fn set_user_data(self, input: Option<String>) -> Self

pub fn additional_info(self, input: impl Into<String>) -> Self

Reserved.

pub fn set_additional_info(self, input: Option<String>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn disable_api_termination(self, input: bool) -> Self

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

Default: false

pub fn set_disable_api_termination(self, input: Option<bool>) -> Self

pub fn dry_run(self, input: bool) -> Self

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

pub fn set_dry_run(self, input: Option<bool>) -> Self

pub fn ebs_optimized(self, input: bool) -> Self

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

pub fn iam_instance_profile(
self,
input: IamInstanceProfileSpecification
) -> Self

The name or Amazon Resource Name (ARN) of an IAM instance profile.

pub fn set_iam_instance_profile(
self,
input: Option<IamInstanceProfileSpecification>
) -> Self

pub fn instance_initiated_shutdown_behavior(
self,
input: ShutdownBehavior
) -> Self

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

pub fn set_instance_initiated_shutdown_behavior(
self,
input: Option<ShutdownBehavior>
) -> Self

pub fn network_interfaces(
self,
input: impl Into<InstanceNetworkInterfaceSpecification>
) -> Self

pub fn set_network_interfaces(
self,
input: Option<Vec<InstanceNetworkInterfaceSpecification>>
) -> Self

pub fn private_ip_address(self, input: impl Into<String>) -> Self

[EC2-VPC] The primary IPv4 address. You must specify a value from the IPv4 address range of the subnet.

Only one private IP address can be designated as primary. You can't specify this option if you've specified the option to designate a private IP address as the primary IP address in a network interface specification. You cannot specify this option if you're launching more than one instance in the request.

You cannot specify this option and the network interfaces option in the same request.

pub fn set_private_ip_address(self, input: Option<String>) -> Self

pub fn elastic_gpu_specification(
self,
input: impl Into<ElasticGpuSpecification>
) -> Self

pub fn set_elastic_gpu_specification(
self,
input: Option<Vec<ElasticGpuSpecification>>
) -> Self

pub fn elastic_inference_accelerators(
self,
input: impl Into<ElasticInferenceAccelerator>
) -> Self

pub fn set_elastic_inference_accelerators(
self,
input: Option<Vec<ElasticInferenceAccelerator>>
) -> Self

pub fn tag_specifications(self, input: impl Into<TagSpecification>) -> Self

pub fn set_tag_specifications(
self,
input: Option<Vec<TagSpecification>>
) -> Self

pub fn launch_template(self, input: LaunchTemplateSpecification) -> Self

The launch template to use to launch the instances. Any parameters that you specify in RunInstances override the same parameters in the launch template. You can specify either the name or ID of a launch template, but not both.

pub fn set_launch_template(
self,
input: Option<LaunchTemplateSpecification>
) -> Self

pub fn instance_market_options(
self,
input: InstanceMarketOptionsRequest
) -> Self

The market (purchasing) option for the instances.

For RunInstances, persistent Spot Instance requests are only supported when InstanceInterruptionBehavior is set to either hibernate or stop.

pub fn set_instance_market_options(
self,
input: Option<InstanceMarketOptionsRequest>
) -> Self

pub fn credit_specification(self, input: CreditSpecificationRequest) -> Self

The credit option for CPU usage of the burstable performance instance. Valid values are standard and unlimited. To change this attribute after launch, use ModifyInstanceCreditSpecification. For more information, see Burstable performance instances in the Amazon EC2 User Guide.

Default: standard (T2 instances) or unlimited (T3/T3a instances)

pub fn set_credit_specification(
self,
input: Option<CreditSpecificationRequest>
) -> Self

pub fn cpu_options(self, input: CpuOptionsRequest) -> Self

The CPU options for the instance. For more information, see Optimizing CPU options in the Amazon EC2 User Guide.

pub fn set_cpu_options(self, input: Option<CpuOptionsRequest>) -> Self

pub fn capacity_reservation_specification(
self,
input: CapacityReservationSpecification
) -> Self

Information about the Capacity Reservation targeting option. If you do not specify this parameter, the instance's Capacity Reservation preference defaults to open, which enables it to run in any open Capacity Reservation that has matching attributes (instance type, platform, Availability Zone).

pub fn set_capacity_reservation_specification(
self,
input: Option<CapacityReservationSpecification>
) -> Self

pub fn hibernation_options(self, input: HibernationOptionsRequest) -> Self

Indicates whether an instance is enabled for hibernation. For more information, see Hibernate your instance in the Amazon EC2 User Guide.

You can't enable hibernation and Amazon Web Services Nitro Enclaves on the same instance.

pub fn set_hibernation_options(
self,
input: Option<HibernationOptionsRequest>
) -> Self

pub fn license_specifications(
self,
input: impl Into<LicenseConfigurationRequest>
) -> Self

pub fn set_license_specifications(
self,
input: Option<Vec<LicenseConfigurationRequest>>
) -> Self

pub fn metadata_options(self, input: InstanceMetadataOptionsRequest) -> Self

The metadata options for the instance. For more information, see Instance metadata and user data.

pub fn set_metadata_options(
self,
input: Option<InstanceMetadataOptionsRequest>
) -> Self

pub fn enclave_options(self, input: EnclaveOptionsRequest) -> Self

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves. For more information, see What is Amazon Web Services Nitro Enclaves? in the Amazon Web Services Nitro Enclaves User Guide.

You can't enable Amazon Web Services Nitro Enclaves and hibernation on the same instance.

pub fn set_enclave_options(self, input: Option<EnclaveOptionsRequest>) -> Self

pub fn build(self) -> Result<RunInstancesInput, BuildError>

Consumes the builder and constructs a RunInstancesInput