Struct aws_sdk_ec2::input::ImportImageInput

#[non_exhaustive]
pub struct ImportImageInput {

    pub architecture: Option<String>,
    pub client_data: Option<ClientData>,
    pub client_token: Option<String>,
    pub description: Option<String>,
    pub disk_containers: Option<Vec<ImageDiskContainer>>,
    pub dry_run: Option<bool>,
    pub encrypted: Option<bool>,
    pub hypervisor: Option<String>,
    pub kms_key_id: Option<String>,
    pub license_type: Option<String>,
    pub platform: Option<String>,
    pub role_name: Option<String>,
    pub license_specifications: Option<Vec<ImportImageLicenseConfigurationRequest>>,
    pub tag_specifications: Option<Vec<TagSpecification>>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

architecture: Option<String>

The architecture of the virtual machine.

Valid values: i386 | x86_64 | arm64

client_data: Option<ClientData>

The client-specific data.

client_token: Option<String>

The token to enable idempotency for VM import requests.

description: Option<String>

A description string for the import image task.

disk_containers: Option<Vec<ImageDiskContainer>>

Information about the disk containers.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

encrypted: Option<bool>

Specifies whether the destination AMI of the imported image should be encrypted. The default CMK for EBS is used unless you specify a non-default AWS Key Management Service (AWS KMS) CMK using KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

hypervisor: Option<String>

The target hypervisor platform.

Valid values: xen

kms_key_id: Option<String>

An identifier for the symmetric AWS Key Management Service (AWS KMS) customer master key (CMK) to use when creating the encrypted AMI. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. If a KmsKeyId is specified, the Encrypted flag must also be set.

The CMK identifier may be provided in any of the following formats:

Key ID

Key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the CMK, the AWS account ID of the CMK owner, the alias namespace, and then the CMK alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

ARN using key ID. The ID ARN contains the arn:aws:kms namespace, followed by the Region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

ARN using key alias. The alias ARN contains the arn:aws:kms namespace, followed by the Region of the CMK, the AWS account ID of the CMK owner, the alias namespace, and then the CMK alias. For example, arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.

AWS parses KmsKeyId asynchronously, meaning that the action you call may appear to complete even though you provided an invalid identifier. This action will eventually report failure.

The specified CMK must exist in the Region that the AMI is being copied to.

Amazon EBS does not support asymmetric CMKs.

license_type: Option<String>

The license type to be used for the Amazon Machine Image (AMI) after importing.

By default, we detect the source-system operating system (OS) and apply the appropriate license. Specify AWS to replace the source-system license with an AWS license, if appropriate. Specify BYOL to retain the source-system license, if appropriate.

To use BYOL, you must have existing licenses with rights to use these licenses in a third party cloud, such as AWS. For more information, see Prerequisites in the VM Import/Export User Guide.

platform: Option<String>

The operating system of the virtual machine.

Valid values: Windows | Linux

role_name: Option<String>

The name of the role to use when not using the default role, 'vmimport'.

license_specifications: Option<Vec<ImportImageLicenseConfigurationRequest>>

The ARNs of the license configurations.

tag_specifications: Option<Vec<TagSpecification>>

The tags to apply to the import image task during creation.

Implementations

impl ImportImageInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<ImportImage, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<ImportImage>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ImportImageInput