Struct aws_sdk_ec2::model::client_vpn_endpoint::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ClientVpnEndpoint

Implementations

impl Builder

pub fn client_vpn_endpoint_id(self, input: impl Into<String>) -> Self

The ID of the Client VPN endpoint.

pub fn set_client_vpn_endpoint_id(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A brief description of the endpoint.

pub fn set_description(self, input: Option<String>) -> Self

pub fn status(self, input: ClientVpnEndpointStatus) -> Self

The current state of the Client VPN endpoint.

pub fn set_status(self, input: Option<ClientVpnEndpointStatus>) -> Self

pub fn creation_time(self, input: impl Into<String>) -> Self

The date and time the Client VPN endpoint was created.

pub fn set_creation_time(self, input: Option<String>) -> Self

pub fn deletion_time(self, input: impl Into<String>) -> Self

The date and time the Client VPN endpoint was deleted, if applicable.

pub fn set_deletion_time(self, input: Option<String>) -> Self

pub fn dns_name(self, input: impl Into<String>) -> Self

The DNS name to be used by clients when connecting to the Client VPN endpoint.

pub fn set_dns_name(self, input: Option<String>) -> Self

pub fn client_cidr_block(self, input: impl Into<String>) -> Self

The IPv4 address range, in CIDR notation, from which client IP addresses are assigned.

pub fn set_client_cidr_block(self, input: Option<String>) -> Self

pub fn dns_servers(self, input: impl Into<String>) -> Self

pub fn set_dns_servers(self, input: Option<Vec<String>>) -> Self

pub fn split_tunnel(self, input: bool) -> Self

Indicates whether split-tunnel is enabled in the AWS Client VPN endpoint.

For information about split-tunnel VPN endpoints, see Split-Tunnel AWS Client VPN Endpoint in the AWS Client VPN Administrator Guide.

pub fn set_split_tunnel(self, input: Option<bool>) -> Self

pub fn vpn_protocol(self, input: VpnProtocol) -> Self

The protocol used by the VPN session.

pub fn set_vpn_protocol(self, input: Option<VpnProtocol>) -> Self

pub fn transport_protocol(self, input: TransportProtocol) -> Self

The transport protocol used by the Client VPN endpoint.

pub fn set_transport_protocol(self, input: Option<TransportProtocol>) -> Self

pub fn vpn_port(self, input: i32) -> Self

The port number for the Client VPN endpoint.

pub fn set_vpn_port(self, input: Option<i32>) -> Self

pub fn associated_target_networks(
self,
input: impl Into<AssociatedTargetNetwork>
) -> Self

pub fn set_associated_target_networks(
self,
input: Option<Vec<AssociatedTargetNetwork>>
) -> Self

pub fn server_certificate_arn(self, input: impl Into<String>) -> Self

The ARN of the server certificate.

pub fn set_server_certificate_arn(self, input: Option<String>) -> Self

pub fn authentication_options(
self,
input: impl Into<ClientVpnAuthentication>
) -> Self

pub fn set_authentication_options(
self,
input: Option<Vec<ClientVpnAuthentication>>
) -> Self

pub fn connection_log_options(self, input: ConnectionLogResponseOptions) -> Self

Information about the client connection logging options for the Client VPN endpoint.

pub fn set_connection_log_options(
self,
input: Option<ConnectionLogResponseOptions>
) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The ID of the VPC.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn self_service_portal_url(self, input: impl Into<String>) -> Self

The URL of the self-service portal.

pub fn set_self_service_portal_url(self, input: Option<String>) -> Self

pub fn client_connect_options(self, input: ClientConnectResponseOptions) -> Self

The options for managing connection authorization for new client connections.

pub fn set_client_connect_options(
self,
input: Option<ClientConnectResponseOptions>
) -> Self

pub fn build(self) -> ClientVpnEndpoint

Consumes the builder and constructs a ClientVpnEndpoint