Struct aws_sdk_ec2::model::fleet_data::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for FleetData

Implementations

impl Builder

pub fn activity_status(self, input: FleetActivityStatus) -> Self

The progress of the EC2 Fleet. If there is an error, the status is error. After all requests are placed, the status is pending_fulfillment. If the size of the EC2 Fleet is equal to or greater than its target capacity, the status is fulfilled. If the size of the EC2 Fleet is decreased, the status is pending_termination while instances are terminating.

pub fn set_activity_status(self, input: Option<FleetActivityStatus>) -> Self

pub fn create_time(self, input: Instant) -> Self

The creation date and time of the EC2 Fleet.

pub fn set_create_time(self, input: Option<Instant>) -> Self

pub fn fleet_id(self, input: impl Into<String>) -> Self

The ID of the EC2 Fleet.

pub fn set_fleet_id(self, input: Option<String>) -> Self

pub fn fleet_state(self, input: FleetStateCode) -> Self

The state of the EC2 Fleet.

pub fn set_fleet_state(self, input: Option<FleetStateCode>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn excess_capacity_termination_policy(
self,
input: FleetExcessCapacityTerminationPolicy
) -> Self

Indicates whether running instances should be terminated if the target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.

pub fn set_excess_capacity_termination_policy(
self,
input: Option<FleetExcessCapacityTerminationPolicy>
) -> Self

pub fn fulfilled_capacity(self, input: f64) -> Self

The number of units fulfilled by this request compared to the set target capacity.

pub fn set_fulfilled_capacity(self, input: Option<f64>) -> Self

pub fn fulfilled_on_demand_capacity(self, input: f64) -> Self

The number of units fulfilled by this request compared to the set target On-Demand capacity.

pub fn set_fulfilled_on_demand_capacity(self, input: Option<f64>) -> Self

pub fn launch_template_configs(
self,
input: impl Into<FleetLaunchTemplateConfig>
) -> Self

pub fn set_launch_template_configs(
self,
input: Option<Vec<FleetLaunchTemplateConfig>>
) -> Self

pub fn target_capacity_specification(
self,
input: TargetCapacitySpecification
) -> Self

The number of units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

pub fn set_target_capacity_specification(
self,
input: Option<TargetCapacitySpecification>
) -> Self

pub fn terminate_instances_with_expiration(self, input: bool) -> Self

Indicates whether running instances should be terminated when the EC2 Fleet expires.

pub fn set_terminate_instances_with_expiration(
self,
input: Option<bool>
) -> Self

pub fn type(self, input: FleetType) -> Self

The type of request. Indicates whether the EC2 Fleet only requests the target capacity, or also attempts to maintain it. If you request a certain target capacity, EC2 Fleet only places the required requests; it does not attempt to replenish instances if capacity is diminished, and it does not submit requests in alternative capacity pools if capacity is unavailable. To maintain a certain target capacity, EC2 Fleet places the required requests to meet this target capacity. It also automatically replenishes any interrupted Spot Instances. Default: maintain.

pub fn set_type(self, input: Option<FleetType>) -> Self

pub fn valid_from(self, input: Instant) -> Self

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

pub fn set_valid_from(self, input: Option<Instant>) -> Self

pub fn valid_until(self, input: Instant) -> Self

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new instance requests are placed or able to fulfill the request. The default end date is 7 days from the current date.

pub fn set_valid_until(self, input: Option<Instant>) -> Self

pub fn replace_unhealthy_instances(self, input: bool) -> Self

Indicates whether EC2 Fleet should replace unhealthy Spot Instances. Supported only for fleets of type maintain. For more information, see EC2 Fleet health checks in the Amazon EC2 User Guide.

pub fn set_replace_unhealthy_instances(self, input: Option<bool>) -> Self

pub fn spot_options(self, input: SpotOptions) -> Self

The configuration of Spot Instances in an EC2 Fleet.

pub fn set_spot_options(self, input: Option<SpotOptions>) -> Self

pub fn on_demand_options(self, input: OnDemandOptions) -> Self

The allocation strategy of On-Demand Instances in an EC2 Fleet.

pub fn set_on_demand_options(self, input: Option<OnDemandOptions>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn errors(self, input: impl Into<DescribeFleetError>) -> Self

pub fn set_errors(self, input: Option<Vec<DescribeFleetError>>) -> Self

pub fn instances(self, input: impl Into<DescribeFleetsInstances>) -> Self

pub fn set_instances(self, input: Option<Vec<DescribeFleetsInstances>>) -> Self

pub fn context(self, input: impl Into<String>) -> Self

Reserved.

pub fn set_context(self, input: Option<String>) -> Self

pub fn build(self) -> FleetData

Consumes the builder and constructs a FleetData