Struct aws_sdk_ec2::model::image::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Image

Implementations

impl Builder

pub fn architecture(self, input: ArchitectureValues) -> Self

The architecture of the image.

pub fn set_architecture(self, input: Option<ArchitectureValues>) -> Self

pub fn creation_date(self, input: impl Into<String>) -> Self

The date and time the image was created.

pub fn set_creation_date(self, input: Option<String>) -> Self

pub fn image_id(self, input: impl Into<String>) -> Self

The ID of the AMI.

pub fn set_image_id(self, input: Option<String>) -> Self

pub fn image_location(self, input: impl Into<String>) -> Self

The location of the AMI.

pub fn set_image_location(self, input: Option<String>) -> Self

pub fn image_type(self, input: ImageTypeValues) -> Self

The type of image.

pub fn set_image_type(self, input: Option<ImageTypeValues>) -> Self

pub fn public(self, input: bool) -> Self

Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

pub fn set_public(self, input: Option<bool>) -> Self

pub fn kernel_id(self, input: impl Into<String>) -> Self

The kernel associated with the image, if any. Only applicable for machine images.

pub fn set_kernel_id(self, input: Option<String>) -> Self

pub fn owner_id(self, input: impl Into<String>) -> Self

The AWS account ID of the image owner.

pub fn set_owner_id(self, input: Option<String>) -> Self

pub fn platform(self, input: PlatformValues) -> Self

This value is set to windows for Windows AMIs; otherwise, it is blank.

pub fn set_platform(self, input: Option<PlatformValues>) -> Self

pub fn platform_details(self, input: impl Into<String>) -> Self

The platform details associated with the billing code of the AMI. For more information, see Obtaining Billing Information in the Amazon Elastic Compute Cloud User Guide.

pub fn set_platform_details(self, input: Option<String>) -> Self

pub fn usage_operation(self, input: impl Into<String>) -> Self

The operation of the Amazon EC2 instance and the billing code that is associated with the AMI. usageOperation corresponds to the lineitem/Operation column on your AWS Cost and Usage Report and in the AWS Price List API. For the list of UsageOperation codes, see Platform Details and Usage Operation Billing Codes in the Amazon Elastic Compute Cloud User Guide.

pub fn set_usage_operation(self, input: Option<String>) -> Self

pub fn product_codes(self, input: impl Into<ProductCode>) -> Self

pub fn set_product_codes(self, input: Option<Vec<ProductCode>>) -> Self

pub fn ramdisk_id(self, input: impl Into<String>) -> Self

The RAM disk associated with the image, if any. Only applicable for machine images.

pub fn set_ramdisk_id(self, input: Option<String>) -> Self

pub fn state(self, input: ImageState) -> Self

The current state of the AMI. If the state is available, the image is successfully registered and can be used to launch an instance.

pub fn set_state(self, input: Option<ImageState>) -> Self

pub fn block_device_mappings(self, input: impl Into<BlockDeviceMapping>) -> Self

pub fn set_block_device_mappings(
self,
input: Option<Vec<BlockDeviceMapping>>
) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description of the AMI that was provided during image creation.

pub fn set_description(self, input: Option<String>) -> Self

pub fn ena_support(self, input: bool) -> Self

Specifies whether enhanced networking with ENA is enabled.

pub fn set_ena_support(self, input: Option<bool>) -> Self

pub fn hypervisor(self, input: HypervisorType) -> Self

The hypervisor type of the image.

pub fn set_hypervisor(self, input: Option<HypervisorType>) -> Self

pub fn image_owner_alias(self, input: impl Into<String>) -> Self

The AWS account alias (for example, amazon, self) or the AWS account ID of the AMI owner.

pub fn set_image_owner_alias(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name of the AMI that was provided during image creation.

pub fn set_name(self, input: Option<String>) -> Self

pub fn root_device_name(self, input: impl Into<String>) -> Self

The device name of the root device volume (for example, /dev/sda1).

pub fn set_root_device_name(self, input: Option<String>) -> Self

pub fn root_device_type(self, input: DeviceType) -> Self

The type of root device used by the AMI. The AMI can use an EBS volume or an instance store volume.

pub fn set_root_device_type(self, input: Option<DeviceType>) -> Self

pub fn sriov_net_support(self, input: impl Into<String>) -> Self

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

pub fn set_sriov_net_support(self, input: Option<String>) -> Self

pub fn state_reason(self, input: StateReason) -> Self

The reason for the state change.

pub fn set_state_reason(self, input: Option<StateReason>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn virtualization_type(self, input: VirtualizationType) -> Self

The type of virtualization of the AMI.

pub fn set_virtualization_type(self, input: Option<VirtualizationType>) -> Self

pub fn boot_mode(self, input: BootModeValues) -> Self

The boot mode of the image. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

pub fn set_boot_mode(self, input: Option<BootModeValues>) -> Self

pub fn deprecation_time(self, input: impl Into<String>) -> Self

The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specified a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

pub fn set_deprecation_time(self, input: Option<String>) -> Self

pub fn build(self) -> Image

Consumes the builder and constructs a Image