Struct aws_sdk_ec2::model::instance::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Instance

Implementations

impl Builder

pub fn ami_launch_index(self, input: i32) -> Self

The AMI launch index, which can be used to find this instance in the launch group.

pub fn set_ami_launch_index(self, input: Option<i32>) -> Self

pub fn image_id(self, input: impl Into<String>) -> Self

The ID of the AMI used to launch the instance.

pub fn set_image_id(self, input: Option<String>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the instance.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn instance_type(self, input: InstanceType) -> Self

The instance type.

pub fn set_instance_type(self, input: Option<InstanceType>) -> Self

pub fn kernel_id(self, input: impl Into<String>) -> Self

The kernel associated with this instance, if applicable.

pub fn set_kernel_id(self, input: Option<String>) -> Self

pub fn key_name(self, input: impl Into<String>) -> Self

The name of the key pair, if this instance was launched with an associated key pair.

pub fn set_key_name(self, input: Option<String>) -> Self

pub fn launch_time(self, input: Instant) -> Self

The time the instance was launched.

pub fn set_launch_time(self, input: Option<Instant>) -> Self

pub fn monitoring(self, input: Monitoring) -> Self

The monitoring for the instance.

pub fn set_monitoring(self, input: Option<Monitoring>) -> Self

pub fn placement(self, input: Placement) -> Self

The location where the instance launched, if applicable.

pub fn set_placement(self, input: Option<Placement>) -> Self

pub fn platform(self, input: PlatformValues) -> Self

The value is Windows for Windows instances; otherwise blank.

pub fn set_platform(self, input: Option<PlatformValues>) -> Self

pub fn private_dns_name(self, input: impl Into<String>) -> Self

(IPv4 only) The private DNS hostname name assigned to the instance. This DNS hostname can only be used inside the Amazon EC2 network. This name is not available until the instance enters the running state.

[EC2-VPC] The Amazon-provided DNS server resolves Amazon-provided private DNS hostnames if you've enabled DNS resolution and DNS hostnames in your VPC. If you are not using the Amazon-provided DNS server in your VPC, your custom domain name servers must resolve the hostname as appropriate.

pub fn set_private_dns_name(self, input: Option<String>) -> Self

pub fn private_ip_address(self, input: impl Into<String>) -> Self

The private IPv4 address assigned to the instance.

pub fn set_private_ip_address(self, input: Option<String>) -> Self

pub fn product_codes(self, input: impl Into<ProductCode>) -> Self

pub fn set_product_codes(self, input: Option<Vec<ProductCode>>) -> Self

pub fn public_dns_name(self, input: impl Into<String>) -> Self

(IPv4 only) The public DNS name assigned to the instance. This name is not available until the instance enters the running state. For EC2-VPC, this name is only available if you've enabled DNS hostnames for your VPC.

pub fn set_public_dns_name(self, input: Option<String>) -> Self

pub fn public_ip_address(self, input: impl Into<String>) -> Self

The public IPv4 address, or the Carrier IP address assigned to the instance, if applicable.

A Carrier IP address only applies to an instance launched in a subnet associated with a Wavelength Zone.

pub fn set_public_ip_address(self, input: Option<String>) -> Self

pub fn ramdisk_id(self, input: impl Into<String>) -> Self

The RAM disk associated with this instance, if applicable.

pub fn set_ramdisk_id(self, input: Option<String>) -> Self

pub fn state(self, input: InstanceState) -> Self

The current state of the instance.

pub fn set_state(self, input: Option<InstanceState>) -> Self

pub fn state_transition_reason(self, input: impl Into<String>) -> Self

The reason for the most recent state transition. This might be an empty string.

pub fn set_state_transition_reason(self, input: Option<String>) -> Self

pub fn subnet_id(self, input: impl Into<String>) -> Self

[EC2-VPC] The ID of the subnet in which the instance is running.

pub fn set_subnet_id(self, input: Option<String>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

[EC2-VPC] The ID of the VPC in which the instance is running.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn architecture(self, input: ArchitectureValues) -> Self

The architecture of the image.

pub fn set_architecture(self, input: Option<ArchitectureValues>) -> Self

pub fn block_device_mappings(
self,
input: impl Into<InstanceBlockDeviceMapping>
) -> Self

pub fn set_block_device_mappings(
self,
input: Option<Vec<InstanceBlockDeviceMapping>>
) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

The idempotency token you provided when you launched the instance, if applicable.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn ebs_optimized(self, input: bool) -> Self

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

pub fn ena_support(self, input: bool) -> Self

Specifies whether enhanced networking with ENA is enabled.

pub fn set_ena_support(self, input: Option<bool>) -> Self

pub fn hypervisor(self, input: HypervisorType) -> Self

The hypervisor type of the instance. The value xen is used for both Xen and Nitro hypervisors.

pub fn set_hypervisor(self, input: Option<HypervisorType>) -> Self

pub fn iam_instance_profile(self, input: IamInstanceProfile) -> Self

The IAM instance profile associated with the instance, if applicable.

pub fn set_iam_instance_profile(self, input: Option<IamInstanceProfile>) -> Self

pub fn instance_lifecycle(self, input: InstanceLifecycleType) -> Self

Indicates whether this is a Spot Instance or a Scheduled Instance.

pub fn set_instance_lifecycle(
self,
input: Option<InstanceLifecycleType>
) -> Self

pub fn elastic_gpu_associations(
self,
input: impl Into<ElasticGpuAssociation>
) -> Self

pub fn set_elastic_gpu_associations(
self,
input: Option<Vec<ElasticGpuAssociation>>
) -> Self

pub fn elastic_inference_accelerator_associations(
self,
input: impl Into<ElasticInferenceAcceleratorAssociation>
) -> Self

pub fn set_elastic_inference_accelerator_associations(
self,
input: Option<Vec<ElasticInferenceAcceleratorAssociation>>
) -> Self

pub fn network_interfaces(
self,
input: impl Into<InstanceNetworkInterface>
) -> Self

pub fn set_network_interfaces(
self,
input: Option<Vec<InstanceNetworkInterface>>
) -> Self

pub fn outpost_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Outpost.

pub fn set_outpost_arn(self, input: Option<String>) -> Self

pub fn root_device_name(self, input: impl Into<String>) -> Self

The device name of the root device volume (for example, /dev/sda1).

pub fn set_root_device_name(self, input: Option<String>) -> Self

pub fn root_device_type(self, input: DeviceType) -> Self

The root device type used by the AMI. The AMI can use an EBS volume or an instance store volume.

pub fn set_root_device_type(self, input: Option<DeviceType>) -> Self

pub fn security_groups(self, input: impl Into<GroupIdentifier>) -> Self

pub fn set_security_groups(self, input: Option<Vec<GroupIdentifier>>) -> Self

pub fn source_dest_check(self, input: bool) -> Self

Indicates whether source/destination checking is enabled.

pub fn set_source_dest_check(self, input: Option<bool>) -> Self

pub fn spot_instance_request_id(self, input: impl Into<String>) -> Self

If the request is a Spot Instance request, the ID of the request.

pub fn set_spot_instance_request_id(self, input: Option<String>) -> Self

pub fn sriov_net_support(self, input: impl Into<String>) -> Self

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

pub fn set_sriov_net_support(self, input: Option<String>) -> Self

pub fn state_reason(self, input: StateReason) -> Self

The reason for the most recent state transition.

pub fn set_state_reason(self, input: Option<StateReason>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn virtualization_type(self, input: VirtualizationType) -> Self

The virtualization type of the instance.

pub fn set_virtualization_type(self, input: Option<VirtualizationType>) -> Self

pub fn cpu_options(self, input: CpuOptions) -> Self

The CPU options for the instance.

pub fn set_cpu_options(self, input: Option<CpuOptions>) -> Self

pub fn capacity_reservation_id(self, input: impl Into<String>) -> Self

The ID of the Capacity Reservation.

pub fn set_capacity_reservation_id(self, input: Option<String>) -> Self

pub fn capacity_reservation_specification(
self,
input: CapacityReservationSpecificationResponse
) -> Self

Information about the Capacity Reservation targeting option.

pub fn set_capacity_reservation_specification(
self,
input: Option<CapacityReservationSpecificationResponse>
) -> Self

pub fn hibernation_options(self, input: HibernationOptions) -> Self

Indicates whether the instance is enabled for hibernation.

pub fn set_hibernation_options(self, input: Option<HibernationOptions>) -> Self

pub fn licenses(self, input: impl Into<LicenseConfiguration>) -> Self

pub fn set_licenses(self, input: Option<Vec<LicenseConfiguration>>) -> Self

pub fn metadata_options(self, input: InstanceMetadataOptionsResponse) -> Self

The metadata options for the instance.

pub fn set_metadata_options(
self,
input: Option<InstanceMetadataOptionsResponse>
) -> Self

pub fn enclave_options(self, input: EnclaveOptions) -> Self

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

pub fn set_enclave_options(self, input: Option<EnclaveOptions>) -> Self

pub fn boot_mode(self, input: BootModeValues) -> Self

The boot mode of the instance. For more information, see Boot modes in the Amazon EC2 User Guide.

pub fn set_boot_mode(self, input: Option<BootModeValues>) -> Self

pub fn build(self) -> Instance

Consumes the builder and constructs a Instance