Struct aws_sdk_ec2::model::instance_network_interface_specification::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for InstanceNetworkInterfaceSpecification

Implementations

impl Builder

pub fn associate_public_ip_address(self, input: bool) -> Self

Indicates whether to assign a public IPv4 address to an instance you launch in a VPC. The public IP address can only be assigned to a network interface for eth0, and can only be assigned to a new network interface, not an existing one. You cannot specify more than one network interface in the request. If launching into a default subnet, the default value is true.

pub fn set_associate_public_ip_address(self, input: Option<bool>) -> Self

pub fn delete_on_termination(self, input: bool) -> Self

If set to true, the interface is deleted when the instance is terminated. You can specify true only if creating a new network interface when launching an instance.

pub fn set_delete_on_termination(self, input: Option<bool>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description of the network interface. Applies only if creating a network interface when launching an instance.

pub fn set_description(self, input: Option<String>) -> Self

pub fn device_index(self, input: i32) -> Self

The position of the network interface in the attachment order. A primary network interface has a device index of 0.

If you specify a network interface when launching an instance, you must specify the device index.

pub fn set_device_index(self, input: Option<i32>) -> Self

pub fn groups(self, input: impl Into<String>) -> Self

pub fn set_groups(self, input: Option<Vec<String>>) -> Self

pub fn ipv6_address_count(self, input: i32) -> Self

A number of IPv6 addresses to assign to the network interface. Amazon EC2 chooses the IPv6 addresses from the range of the subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

pub fn set_ipv6_address_count(self, input: Option<i32>) -> Self

pub fn ipv6_addresses(self, input: impl Into<InstanceIpv6Address>) -> Self

pub fn set_ipv6_addresses(self, input: Option<Vec<InstanceIpv6Address>>) -> Self

pub fn network_interface_id(self, input: impl Into<String>) -> Self

The ID of the network interface.

If you are creating a Spot Fleet, omit this parameter because you can’t specify a network interface ID in a launch specification.

pub fn set_network_interface_id(self, input: Option<String>) -> Self

pub fn private_ip_address(self, input: impl Into<String>) -> Self

The private IPv4 address of the network interface. Applies only if creating a network interface when launching an instance. You cannot specify this option if you're launching more than one instance in a RunInstances request.

pub fn set_private_ip_address(self, input: Option<String>) -> Self

pub fn private_ip_addresses(
self,
input: impl Into<PrivateIpAddressSpecification>
) -> Self

pub fn set_private_ip_addresses(
self,
input: Option<Vec<PrivateIpAddressSpecification>>
) -> Self

pub fn secondary_private_ip_address_count(self, input: i32) -> Self

The number of secondary private IPv4 addresses. You can't specify this option and specify more than one private IP address using the private IP addresses option. You cannot specify this option if you're launching more than one instance in a RunInstances request.

pub fn set_secondary_private_ip_address_count(self, input: Option<i32>) -> Self

pub fn subnet_id(self, input: impl Into<String>) -> Self

The ID of the subnet associated with the network interface. Applies only if creating a network interface when launching an instance.

pub fn set_subnet_id(self, input: Option<String>) -> Self

pub fn associate_carrier_ip_address(self, input: bool) -> Self

Indicates whether to assign a carrier IP address to the network interface.

You can only assign a carrier IP address to a network interface that is in a subnet in a Wavelength Zone. For more information about carrier IP addresses, see Carrier IP addresses in the Amazon Web Services Wavelength Developer Guide.

pub fn set_associate_carrier_ip_address(self, input: Option<bool>) -> Self

pub fn interface_type(self, input: impl Into<String>) -> Self

The type of network interface.

To create an Elastic Fabric Adapter (EFA), specify efa. For more information, see Elastic Fabric Adapter in the Amazon Elastic Compute Cloud User Guide.

Valid values: interface | efa

pub fn set_interface_type(self, input: Option<String>) -> Self

pub fn network_card_index(self, input: i32) -> Self

The index of the network card. Some instance types support multiple network cards. The primary network interface must be assigned to network card index 0. The default is network card index 0.

pub fn set_network_card_index(self, input: Option<i32>) -> Self

pub fn build(self) -> InstanceNetworkInterfaceSpecification

Consumes the builder and constructs a InstanceNetworkInterfaceSpecification