Struct aws_sdk_ec2::model::instance_type_info::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for InstanceTypeInfo

Implementations

impl Builder

pub fn instance_type(self, input: InstanceType) -> Self

The instance type. For more information, see Instance types in the Amazon EC2 User Guide.

pub fn set_instance_type(self, input: Option<InstanceType>) -> Self

pub fn current_generation(self, input: bool) -> Self

Indicates whether the instance type is current generation.

pub fn set_current_generation(self, input: Option<bool>) -> Self

pub fn free_tier_eligible(self, input: bool) -> Self

Indicates whether the instance type is eligible for the free tier.

pub fn set_free_tier_eligible(self, input: Option<bool>) -> Self

pub fn supported_usage_classes(self, input: impl Into<UsageClassType>) -> Self

pub fn set_supported_usage_classes(
self,
input: Option<Vec<UsageClassType>>
) -> Self

pub fn supported_root_device_types(
self,
input: impl Into<RootDeviceType>
) -> Self

pub fn set_supported_root_device_types(
self,
input: Option<Vec<RootDeviceType>>
) -> Self

pub fn supported_virtualization_types(
self,
input: impl Into<VirtualizationType>
) -> Self

pub fn set_supported_virtualization_types(
self,
input: Option<Vec<VirtualizationType>>
) -> Self

pub fn bare_metal(self, input: bool) -> Self

Indicates whether the instance is a bare metal instance type.

pub fn set_bare_metal(self, input: Option<bool>) -> Self

pub fn hypervisor(self, input: InstanceTypeHypervisor) -> Self

The hypervisor for the instance type.

pub fn set_hypervisor(self, input: Option<InstanceTypeHypervisor>) -> Self

pub fn processor_info(self, input: ProcessorInfo) -> Self

Describes the processor.

pub fn set_processor_info(self, input: Option<ProcessorInfo>) -> Self

pub fn v_cpu_info(self, input: VCpuInfo) -> Self

Describes the vCPU configurations for the instance type.

pub fn set_v_cpu_info(self, input: Option<VCpuInfo>) -> Self

pub fn memory_info(self, input: MemoryInfo) -> Self

Describes the memory for the instance type.

pub fn set_memory_info(self, input: Option<MemoryInfo>) -> Self

pub fn instance_storage_supported(self, input: bool) -> Self

Indicates whether instance storage is supported.

pub fn set_instance_storage_supported(self, input: Option<bool>) -> Self

pub fn instance_storage_info(self, input: InstanceStorageInfo) -> Self

Describes the instance storage for the instance type.

pub fn set_instance_storage_info(
self,
input: Option<InstanceStorageInfo>
) -> Self

pub fn ebs_info(self, input: EbsInfo) -> Self

Describes the Amazon EBS settings for the instance type.

pub fn set_ebs_info(self, input: Option<EbsInfo>) -> Self

pub fn network_info(self, input: NetworkInfo) -> Self

Describes the network settings for the instance type.

pub fn set_network_info(self, input: Option<NetworkInfo>) -> Self

pub fn gpu_info(self, input: GpuInfo) -> Self

Describes the GPU accelerator settings for the instance type.

pub fn set_gpu_info(self, input: Option<GpuInfo>) -> Self

pub fn fpga_info(self, input: FpgaInfo) -> Self

Describes the FPGA accelerator settings for the instance type.

pub fn set_fpga_info(self, input: Option<FpgaInfo>) -> Self

pub fn placement_group_info(self, input: PlacementGroupInfo) -> Self

Describes the placement group settings for the instance type.

pub fn set_placement_group_info(self, input: Option<PlacementGroupInfo>) -> Self

pub fn inference_accelerator_info(self, input: InferenceAcceleratorInfo) -> Self

Describes the Inference accelerator settings for the instance type.

pub fn set_inference_accelerator_info(
self,
input: Option<InferenceAcceleratorInfo>
) -> Self

pub fn hibernation_supported(self, input: bool) -> Self

Indicates whether On-Demand hibernation is supported.

pub fn set_hibernation_supported(self, input: Option<bool>) -> Self

pub fn burstable_performance_supported(self, input: bool) -> Self

Indicates whether the instance type is a burstable performance instance type.

pub fn set_burstable_performance_supported(self, input: Option<bool>) -> Self

pub fn dedicated_hosts_supported(self, input: bool) -> Self

Indicates whether Dedicated Hosts are supported on the instance type.

pub fn set_dedicated_hosts_supported(self, input: Option<bool>) -> Self

pub fn auto_recovery_supported(self, input: bool) -> Self

Indicates whether auto recovery is supported.

pub fn set_auto_recovery_supported(self, input: Option<bool>) -> Self

pub fn supported_boot_modes(self, input: impl Into<BootModeType>) -> Self

pub fn set_supported_boot_modes(self, input: Option<Vec<BootModeType>>) -> Self

pub fn build(self) -> InstanceTypeInfo

Consumes the builder and constructs a InstanceTypeInfo