Struct aws_sdk_ec2::model::modify_vpn_tunnel_options_specification::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ModifyVpnTunnelOptionsSpecification

Implementations

impl Builder

pub fn tunnel_inside_cidr(self, input: impl Into<String>) -> Self

The range of inside IPv4 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same virtual private gateway.

Constraints: A size /30 CIDR block from the 169.254.0.0/16 range. The following CIDR blocks are reserved and cannot be used:

169.254.0.0/30

169.254.1.0/30

169.254.2.0/30

169.254.3.0/30

169.254.4.0/30

169.254.5.0/30

169.254.169.252/30

pub fn set_tunnel_inside_cidr(self, input: Option<String>) -> Self

pub fn tunnel_inside_ipv6_cidr(self, input: impl Into<String>) -> Self

The range of inside IPv6 addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same transit gateway.

Constraints: A size /126 CIDR block from the local fd00::/8 range.

pub fn set_tunnel_inside_ipv6_cidr(self, input: Option<String>) -> Self

pub fn pre_shared_key(self, input: impl Into<String>) -> Self

The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and the customer gateway.

Constraints: Allowed characters are alphanumeric characters, periods (.), and underscores (_). Must be between 8 and 64 characters in length and cannot start with zero (0).

pub fn set_pre_shared_key(self, input: Option<String>) -> Self

pub fn phase1_lifetime_seconds(self, input: i32) -> Self

The lifetime for phase 1 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 28,800.

Default: 28800

pub fn set_phase1_lifetime_seconds(self, input: Option<i32>) -> Self

pub fn phase2_lifetime_seconds(self, input: i32) -> Self

The lifetime for phase 2 of the IKE negotiation, in seconds.

Constraints: A value between 900 and 3,600. The value must be less than the value for Phase1LifetimeSeconds.

Default: 3600

pub fn set_phase2_lifetime_seconds(self, input: Option<i32>) -> Self

pub fn rekey_margin_time_seconds(self, input: i32) -> Self

The margin time, in seconds, before the phase 2 lifetime expires, during which the AWS side of the VPN connection performs an IKE rekey. The exact time of the rekey is randomly selected based on the value for RekeyFuzzPercentage.

Constraints: A value between 60 and half of Phase2LifetimeSeconds.

Default: 540

pub fn set_rekey_margin_time_seconds(self, input: Option<i32>) -> Self

pub fn rekey_fuzz_percentage(self, input: i32) -> Self

The percentage of the rekey window (determined by RekeyMarginTimeSeconds) during which the rekey time is randomly selected.

Constraints: A value between 0 and 100.

Default: 100

pub fn set_rekey_fuzz_percentage(self, input: Option<i32>) -> Self

pub fn replay_window_size(self, input: i32) -> Self

The number of packets in an IKE replay window.

Constraints: A value between 64 and 2048.

Default: 1024

pub fn set_replay_window_size(self, input: Option<i32>) -> Self

pub fn dpd_timeout_seconds(self, input: i32) -> Self

The number of seconds after which a DPD timeout occurs.

Constraints: A value between 0 and 30.

Default: 30

pub fn set_dpd_timeout_seconds(self, input: Option<i32>) -> Self

pub fn dpd_timeout_action(self, input: impl Into<String>) -> Self

The action to take after DPD timeout occurs. Specify restart to restart the IKE initiation. Specify clear to end the IKE session.

Valid Values: clear | none | restart

Default: clear

pub fn set_dpd_timeout_action(self, input: Option<String>) -> Self

pub fn phase1_encryption_algorithms(
self,
input: impl Into<Phase1EncryptionAlgorithmsRequestListValue>
) -> Self

pub fn set_phase1_encryption_algorithms(
self,
input: Option<Vec<Phase1EncryptionAlgorithmsRequestListValue>>
) -> Self

pub fn phase2_encryption_algorithms(
self,
input: impl Into<Phase2EncryptionAlgorithmsRequestListValue>
) -> Self

pub fn set_phase2_encryption_algorithms(
self,
input: Option<Vec<Phase2EncryptionAlgorithmsRequestListValue>>
) -> Self

pub fn phase1_integrity_algorithms(
self,
input: impl Into<Phase1IntegrityAlgorithmsRequestListValue>
) -> Self

pub fn set_phase1_integrity_algorithms(
self,
input: Option<Vec<Phase1IntegrityAlgorithmsRequestListValue>>
) -> Self

pub fn phase2_integrity_algorithms(
self,
input: impl Into<Phase2IntegrityAlgorithmsRequestListValue>
) -> Self

pub fn set_phase2_integrity_algorithms(
self,
input: Option<Vec<Phase2IntegrityAlgorithmsRequestListValue>>
) -> Self

pub fn phase1_dh_group_numbers(
self,
input: impl Into<Phase1DhGroupNumbersRequestListValue>
) -> Self

pub fn set_phase1_dh_group_numbers(
self,
input: Option<Vec<Phase1DhGroupNumbersRequestListValue>>
) -> Self

pub fn phase2_dh_group_numbers(
self,
input: impl Into<Phase2DhGroupNumbersRequestListValue>
) -> Self

pub fn set_phase2_dh_group_numbers(
self,
input: Option<Vec<Phase2DhGroupNumbersRequestListValue>>
) -> Self

pub fn ike_versions(self, input: impl Into<IkeVersionsRequestListValue>) -> Self

pub fn set_ike_versions(
self,
input: Option<Vec<IkeVersionsRequestListValue>>
) -> Self

pub fn startup_action(self, input: impl Into<String>) -> Self

The action to take when the establishing the tunnel for the VPN connection. By default, your customer gateway device must initiate the IKE negotiation and bring up the tunnel. Specify start for AWS to initiate the IKE negotiation.

Valid Values: add | start

Default: add

pub fn set_startup_action(self, input: Option<String>) -> Self

pub fn build(self) -> ModifyVpnTunnelOptionsSpecification

Consumes the builder and constructs a ModifyVpnTunnelOptionsSpecification