Struct aws_sdk_ec2::model::network_interface::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for NetworkInterface

Implementations

impl Builder

pub fn association(self, input: NetworkInterfaceAssociation) -> Self

The association information for an Elastic IP address (IPv4) associated with the network interface.

pub fn set_association(self, input: Option<NetworkInterfaceAssociation>) -> Self

pub fn attachment(self, input: NetworkInterfaceAttachment) -> Self

The network interface attachment.

pub fn set_attachment(self, input: Option<NetworkInterfaceAttachment>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description.

pub fn set_description(self, input: Option<String>) -> Self

pub fn groups(self, input: impl Into<GroupIdentifier>) -> Self

pub fn set_groups(self, input: Option<Vec<GroupIdentifier>>) -> Self

pub fn interface_type(self, input: NetworkInterfaceType) -> Self

The type of network interface.

pub fn set_interface_type(self, input: Option<NetworkInterfaceType>) -> Self

pub fn ipv6_addresses(
self,
input: impl Into<NetworkInterfaceIpv6Address>
) -> Self

pub fn set_ipv6_addresses(
self,
input: Option<Vec<NetworkInterfaceIpv6Address>>
) -> Self

pub fn mac_address(self, input: impl Into<String>) -> Self

The MAC address.

pub fn set_mac_address(self, input: Option<String>) -> Self

pub fn network_interface_id(self, input: impl Into<String>) -> Self

The ID of the network interface.

pub fn set_network_interface_id(self, input: Option<String>) -> Self

pub fn outpost_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Outpost.

pub fn set_outpost_arn(self, input: Option<String>) -> Self

pub fn owner_id(self, input: impl Into<String>) -> Self

The account ID of the owner of the network interface.

pub fn set_owner_id(self, input: Option<String>) -> Self

pub fn private_dns_name(self, input: impl Into<String>) -> Self

The private DNS name.

pub fn set_private_dns_name(self, input: Option<String>) -> Self

pub fn private_ip_address(self, input: impl Into<String>) -> Self

The IPv4 address of the network interface within the subnet.

pub fn set_private_ip_address(self, input: Option<String>) -> Self

pub fn private_ip_addresses(
self,
input: impl Into<NetworkInterfacePrivateIpAddress>
) -> Self

pub fn set_private_ip_addresses(
self,
input: Option<Vec<NetworkInterfacePrivateIpAddress>>
) -> Self

pub fn requester_id(self, input: impl Into<String>) -> Self

The alias or account ID of the principal or service that created the network interface.

pub fn set_requester_id(self, input: Option<String>) -> Self

pub fn requester_managed(self, input: bool) -> Self

Indicates whether the network interface is being managed by Amazon Web Services.

pub fn set_requester_managed(self, input: Option<bool>) -> Self

pub fn source_dest_check(self, input: bool) -> Self

Indicates whether source/destination checking is enabled.

pub fn set_source_dest_check(self, input: Option<bool>) -> Self

pub fn status(self, input: NetworkInterfaceStatus) -> Self

The status of the network interface.

pub fn set_status(self, input: Option<NetworkInterfaceStatus>) -> Self

pub fn subnet_id(self, input: impl Into<String>) -> Self

The ID of the subnet.

pub fn set_subnet_id(self, input: Option<String>) -> Self

pub fn tag_set(self, input: impl Into<Tag>) -> Self

pub fn set_tag_set(self, input: Option<Vec<Tag>>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The ID of the VPC.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn build(self) -> NetworkInterface

Consumes the builder and constructs a NetworkInterface