Struct aws_sdk_ec2::model::spot_instance_request::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for SpotInstanceRequest

Implementations

impl Builder

pub fn actual_block_hourly_price(self, input: impl Into<String>) -> Self

Deprecated.

pub fn set_actual_block_hourly_price(self, input: Option<String>) -> Self

pub fn availability_zone_group(self, input: impl Into<String>) -> Self

The Availability Zone group. If you specify the same Availability Zone group for all Spot Instance requests, all Spot Instances are launched in the same Availability Zone.

pub fn set_availability_zone_group(self, input: Option<String>) -> Self

pub fn block_duration_minutes(self, input: i32) -> Self

Deprecated.

pub fn set_block_duration_minutes(self, input: Option<i32>) -> Self

pub fn create_time(self, input: Instant) -> Self

The date and time when the Spot Instance request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

pub fn set_create_time(self, input: Option<Instant>) -> Self

pub fn fault(self, input: SpotInstanceStateFault) -> Self

The fault codes for the Spot Instance request, if any.

pub fn set_fault(self, input: Option<SpotInstanceStateFault>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The instance ID, if an instance has been launched to fulfill the Spot Instance request.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn launch_group(self, input: impl Into<String>) -> Self

The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

pub fn set_launch_group(self, input: Option<String>) -> Self

pub fn launch_specification(self, input: LaunchSpecification) -> Self

Additional information for launching instances.

pub fn set_launch_specification(
self,
input: Option<LaunchSpecification>
) -> Self

pub fn launched_availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone in which the request is launched.

pub fn set_launched_availability_zone(self, input: Option<String>) -> Self

pub fn product_description(self, input: RiProductDescription) -> Self

The product description associated with the Spot Instance.

pub fn set_product_description(
self,
input: Option<RiProductDescription>
) -> Self

pub fn spot_instance_request_id(self, input: impl Into<String>) -> Self

The ID of the Spot Instance request.

pub fn set_spot_instance_request_id(self, input: Option<String>) -> Self

pub fn spot_price(self, input: impl Into<String>) -> Self

The maximum price per hour that you are willing to pay for a Spot Instance.

pub fn set_spot_price(self, input: Option<String>) -> Self

pub fn state(self, input: SpotInstanceState) -> Self

The state of the Spot Instance request. Spot status information helps track your Spot Instance requests. For more information, see Spot status in the Amazon EC2 User Guide for Linux Instances.

pub fn set_state(self, input: Option<SpotInstanceState>) -> Self

pub fn status(self, input: SpotInstanceStatus) -> Self

The status code and status message describing the Spot Instance request.

pub fn set_status(self, input: Option<SpotInstanceStatus>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn type(self, input: SpotInstanceType) -> Self

The Spot Instance request type.

pub fn set_type(self, input: Option<SpotInstanceType>) -> Self

pub fn valid_from(self, input: Instant) -> Self

The start date of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The request becomes active at this date and time.

pub fn set_valid_from(self, input: Option<Instant>) -> Self

pub fn valid_until(self, input: Instant) -> Self

The end date of the request, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

For a persistent request, the request remains active until the validUntil date and time is reached. Otherwise, the request remains active until you cancel it.

For a one-time request, the request remains active until all instances launch, the request is canceled, or the validUntil date and time is reached. By default, the request is valid for 7 days from the date the request was created.

pub fn set_valid_until(self, input: Option<Instant>) -> Self

pub fn instance_interruption_behavior(
self,
input: InstanceInterruptionBehavior
) -> Self

The behavior when a Spot Instance is interrupted.

pub fn set_instance_interruption_behavior(
self,
input: Option<InstanceInterruptionBehavior>
) -> Self

pub fn build(self) -> SpotInstanceRequest

Consumes the builder and constructs a SpotInstanceRequest