Struct aws_sdk_ec2::model::spot_options_request::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for SpotOptionsRequest

Implementations

impl Builder

pub fn allocation_strategy(self, input: SpotAllocationStrategy) -> Self

Indicates how to allocate the target Spot Instance capacity across the Spot Instance pools specified by the EC2 Fleet.

If the allocation strategy is lowest-price, EC2 Fleet launches instances from the Spot Instance pools with the lowest price. This is the default allocation strategy.

If the allocation strategy is diversified, EC2 Fleet launches instances from all of the Spot Instance pools that you specify.

If the allocation strategy is capacity-optimized (recommended), EC2 Fleet launches instances from Spot Instance pools with optimal capacity for the number of instances that are launching. To give certain instance types a higher chance of launching first, use capacity-optimized-prioritized. Set a priority for each instance type by using the Priority parameter for LaunchTemplateOverrides. You can assign the same priority to different LaunchTemplateOverrides. EC2 implements the priorities on a best-effort basis, but optimizes for capacity first. capacity-optimized-prioritized is supported only if your fleet uses a launch template. Note that if the On-Demand AllocationStrategy is set to prioritized, the same priority is applied when fulfilling On-Demand capacity.

pub fn set_allocation_strategy(
self,
input: Option<SpotAllocationStrategy>
) -> Self

pub fn maintenance_strategies(
self,
input: FleetSpotMaintenanceStrategiesRequest
) -> Self

The strategies for managing your Spot Instances that are at an elevated risk of being interrupted.

pub fn set_maintenance_strategies(
self,
input: Option<FleetSpotMaintenanceStrategiesRequest>
) -> Self

pub fn instance_interruption_behavior(
self,
input: SpotInstanceInterruptionBehavior
) -> Self

The behavior when a Spot Instance is interrupted. The default is terminate.

pub fn set_instance_interruption_behavior(
self,
input: Option<SpotInstanceInterruptionBehavior>
) -> Self

pub fn instance_pools_to_use_count(self, input: i32) -> Self

The number of Spot pools across which to allocate your target Spot capacity. Valid only when Spot AllocationStrategy is set to lowest-price. EC2 Fleet selects the cheapest Spot pools and evenly allocates your target Spot capacity across the number of Spot pools that you specify.

Note that EC2 Fleet attempts to draw Spot Instances from the number of pools that you specify on a best effort basis. If a pool runs out of Spot capacity before fulfilling your target capacity, EC2 Fleet will continue to fulfill your request by drawing from the next cheapest pool. To ensure that your target capacity is met, you might receive Spot Instances from more than the number of pools that you specified. Similarly, if most of the pools have no Spot capacity, you might receive your full target capacity from fewer than the number of pools that you specified.

pub fn set_instance_pools_to_use_count(self, input: Option<i32>) -> Self

pub fn single_instance_type(self, input: bool) -> Self

Indicates that the fleet uses a single instance type to launch all Spot Instances in the fleet. Supported only for fleets of type instant.

pub fn set_single_instance_type(self, input: Option<bool>) -> Self

pub fn single_availability_zone(self, input: bool) -> Self

Indicates that the fleet launches all Spot Instances into a single Availability Zone. Supported only for fleets of type instant.

pub fn set_single_availability_zone(self, input: Option<bool>) -> Self

pub fn min_target_capacity(self, input: i32) -> Self

The minimum target capacity for Spot Instances in the fleet. If the minimum target capacity is not reached, the fleet launches no instances.

pub fn set_min_target_capacity(self, input: Option<i32>) -> Self

pub fn max_total_price(self, input: impl Into<String>) -> Self

The maximum amount per hour for Spot Instances that you're willing to pay.

pub fn set_max_total_price(self, input: Option<String>) -> Self

pub fn build(self) -> SpotOptionsRequest

Consumes the builder and constructs a SpotOptionsRequest