Struct aws_sdk_ec2::model::AvailabilityZone

#[non_exhaustive]
pub struct AvailabilityZone {
    pub state: Option<AvailabilityZoneState>,
    pub opt_in_status: Option<AvailabilityZoneOptInStatus>,
    pub messages: Option<Vec<AvailabilityZoneMessage>>,
    pub region_name: Option<String>,
    pub zone_name: Option<String>,
    pub zone_id: Option<String>,
    pub group_name: Option<String>,
    pub network_border_group: Option<String>,
    pub zone_type: Option<String>,
    pub parent_zone_name: Option<String>,
    pub parent_zone_id: Option<String>,
}

Describes Availability Zones, Local Zones, and Wavelength Zones.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

state: Option<AvailabilityZoneState>

The state of the Availability Zone, Local Zone, or Wavelength Zone.

opt_in_status: Option<AvailabilityZoneOptInStatus>

For Availability Zones, this parameter always has the value of opt-in-not-required.

For Local Zones and Wavelength Zones, this parameter is the opt-in status. The possible values are opted-in, and not-opted-in.

messages: Option<Vec<AvailabilityZoneMessage>>

Any messages about the Availability Zone, Local Zone, or Wavelength Zone.

region_name: Option<String>

The name of the Region.

zone_name: Option<String>

The name of the Availability Zone, Local Zone, or Wavelength Zone.

zone_id: Option<String>

The ID of the Availability Zone, Local Zone, or Wavelength Zone.

group_name: Option<String>

For Availability Zones, this parameter has the same value as the Region name.

For Local Zones, the name of the associated group, for example us-west-2-lax-1.

For Wavelength Zones, the name of the associated group, for example us-east-1-wl1-bos-wlz-1.

network_border_group: Option<String>

The name of the network border group.

zone_type: Option<String>

The type of zone. The valid values are availability-zone, local-zone, and wavelength-zone.

parent_zone_name: Option<String>

The name of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

parent_zone_id: Option<String>

The ID of the zone that handles some of the Local Zone or Wavelength Zone control plane operations, such as API calls.

Implementations

impl AvailabilityZone

pub fn builder() -> Builder

Creates a new builder-style object to manufacture AvailabilityZone