Struct aws_sdk_ec2::model::Image

#[non_exhaustive]
pub struct Image {

    pub architecture: Option<ArchitectureValues>,
    pub creation_date: Option<String>,
    pub image_id: Option<String>,
    pub image_location: Option<String>,
    pub image_type: Option<ImageTypeValues>,
    pub public: Option<bool>,
    pub kernel_id: Option<String>,
    pub owner_id: Option<String>,
    pub platform: Option<PlatformValues>,
    pub platform_details: Option<String>,
    pub usage_operation: Option<String>,
    pub product_codes: Option<Vec<ProductCode>>,
    pub ramdisk_id: Option<String>,
    pub state: Option<ImageState>,
    pub block_device_mappings: Option<Vec<BlockDeviceMapping>>,
    pub description: Option<String>,
    pub ena_support: Option<bool>,
    pub hypervisor: Option<HypervisorType>,
    pub image_owner_alias: Option<String>,
    pub name: Option<String>,
    pub root_device_name: Option<String>,
    pub root_device_type: Option<DeviceType>,
    pub sriov_net_support: Option<String>,
    pub state_reason: Option<StateReason>,
    pub tags: Option<Vec<Tag>>,
    pub virtualization_type: Option<VirtualizationType>,
    pub boot_mode: Option<BootModeValues>,
    pub deprecation_time: Option<String>,

}

Describes an image.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

architecture: Option<ArchitectureValues>

The architecture of the image.

creation_date: Option<String>

The date and time the image was created.

image_id: Option<String>

The ID of the AMI.

image_location: Option<String>

The location of the AMI.

image_type: Option<ImageTypeValues>

The type of image.

public: Option<bool>

Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

kernel_id: Option<String>

The kernel associated with the image, if any. Only applicable for machine images.

owner_id: Option<String>

The ID of the Amazon Web Services account that owns the image.

platform: Option<PlatformValues>

This value is set to windows for Windows AMIs; otherwise, it is blank.

platform_details: Option<String>

The platform details associated with the billing code of the AMI. For more information, see Understanding AMI billing in the Amazon Elastic Compute Cloud User Guide.

usage_operation: Option<String>

The operation of the Amazon EC2 instance and the billing code that is associated with the AMI. usageOperation corresponds to the lineitem/Operation column on your Amazon Web Services Cost and Usage Report and in the Amazon Web Services Price List API. You can view these fields on the Instances or AMIs pages in the Amazon EC2 console, or in the responses that are returned by the DescribeImages command in the Amazon EC2 API, or the describe-images command in the CLI.

product_codes: Option<Vec<ProductCode>>

Any product codes associated with the AMI.

ramdisk_id: Option<String>

The RAM disk associated with the image, if any. Only applicable for machine images.

state: Option<ImageState>

The current state of the AMI. If the state is available, the image is successfully registered and can be used to launch an instance.

block_device_mappings: Option<Vec<BlockDeviceMapping>>

Any block device mapping entries.

description: Option<String>

The description of the AMI that was provided during image creation.

ena_support: Option<bool>

Specifies whether enhanced networking with ENA is enabled.

hypervisor: Option<HypervisorType>

The hypervisor type of the image.

image_owner_alias: Option<String>

The Amazon Web Services account alias (for example, amazon, self) or the Amazon Web Services account ID of the AMI owner.

name: Option<String>

The name of the AMI that was provided during image creation.

root_device_name: Option<String>

The device name of the root device volume (for example, /dev/sda1).

root_device_type: Option<DeviceType>

The type of root device used by the AMI. The AMI can use an Amazon EBS volume or an instance store volume.

sriov_net_support: Option<String>

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

state_reason: Option<StateReason>

The reason for the state change.

tags: Option<Vec<Tag>>

Any tags assigned to the image.

virtualization_type: Option<VirtualizationType>

The type of virtualization of the AMI.

boot_mode: Option<BootModeValues>

The boot mode of the image. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

deprecation_time: Option<String>

The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specified a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

Implementations

impl Image

pub fn architecture(&self) -> Option<&ArchitectureValues>

The architecture of the image.

pub fn creation_date(&self) -> Option<&str>

The date and time the image was created.

pub fn image_id(&self) -> Option<&str>

The ID of the AMI.

pub fn image_location(&self) -> Option<&str>

The location of the AMI.

pub fn image_type(&self) -> Option<&ImageTypeValues>

The type of image.

pub fn public(&self) -> Option<bool>

Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

pub fn kernel_id(&self) -> Option<&str>

The kernel associated with the image, if any. Only applicable for machine images.

pub fn owner_id(&self) -> Option<&str>

The ID of the Amazon Web Services account that owns the image.

pub fn platform(&self) -> Option<&PlatformValues>

This value is set to windows for Windows AMIs; otherwise, it is blank.

pub fn platform_details(&self) -> Option<&str>

The platform details associated with the billing code of the AMI. For more information, see Understanding AMI billing in the Amazon Elastic Compute Cloud User Guide.

pub fn usage_operation(&self) -> Option<&str>

The operation of the Amazon EC2 instance and the billing code that is associated with the AMI. usageOperation corresponds to the lineitem/Operation column on your Amazon Web Services Cost and Usage Report and in the Amazon Web Services Price List API. You can view these fields on the Instances or AMIs pages in the Amazon EC2 console, or in the responses that are returned by the DescribeImages command in the Amazon EC2 API, or the describe-images command in the CLI.

pub fn product_codes(&self) -> Option<&[ProductCode]>

Any product codes associated with the AMI.

pub fn ramdisk_id(&self) -> Option<&str>

The RAM disk associated with the image, if any. Only applicable for machine images.

pub fn state(&self) -> Option<&ImageState>

The current state of the AMI. If the state is available, the image is successfully registered and can be used to launch an instance.

pub fn block_device_mappings(&self) -> Option<&[BlockDeviceMapping]>

Any block device mapping entries.

pub fn description(&self) -> Option<&str>

The description of the AMI that was provided during image creation.

pub fn ena_support(&self) -> Option<bool>

Specifies whether enhanced networking with ENA is enabled.

pub fn hypervisor(&self) -> Option<&HypervisorType>

The hypervisor type of the image.

pub fn image_owner_alias(&self) -> Option<&str>

The Amazon Web Services account alias (for example, amazon, self) or the Amazon Web Services account ID of the AMI owner.

pub fn name(&self) -> Option<&str>

The name of the AMI that was provided during image creation.

pub fn root_device_name(&self) -> Option<&str>

The device name of the root device volume (for example, /dev/sda1).

pub fn root_device_type(&self) -> Option<&DeviceType>

The type of root device used by the AMI. The AMI can use an Amazon EBS volume or an instance store volume.

pub fn sriov_net_support(&self) -> Option<&str>

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

pub fn state_reason(&self) -> Option<&StateReason>

The reason for the state change.

pub fn tags(&self) -> Option<&[Tag]>

Any tags assigned to the image.

pub fn virtualization_type(&self) -> Option<&VirtualizationType>

The type of virtualization of the AMI.

pub fn boot_mode(&self) -> Option<&BootModeValues>

The boot mode of the image. For more information, see Boot modes in the Amazon Elastic Compute Cloud User Guide.

pub fn deprecation_time(&self) -> Option<&str>

The date and time to deprecate the AMI, in UTC, in the following format: YYYY-MM-DDTHH:MM:SSZ. If you specified a value for seconds, Amazon EC2 rounds the seconds to the nearest minute.

impl Image

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Image