Struct aws_sdk_ec2::model::Instance

#[non_exhaustive]
pub struct Instance {

    pub ami_launch_index: Option<i32>,
    pub image_id: Option<String>,
    pub instance_id: Option<String>,
    pub instance_type: Option<InstanceType>,
    pub kernel_id: Option<String>,
    pub key_name: Option<String>,
    pub launch_time: Option<Instant>,
    pub monitoring: Option<Monitoring>,
    pub placement: Option<Placement>,
    pub platform: Option<PlatformValues>,
    pub private_dns_name: Option<String>,
    pub private_ip_address: Option<String>,
    pub product_codes: Option<Vec<ProductCode>>,
    pub public_dns_name: Option<String>,
    pub public_ip_address: Option<String>,
    pub ramdisk_id: Option<String>,
    pub state: Option<InstanceState>,
    pub state_transition_reason: Option<String>,
    pub subnet_id: Option<String>,
    pub vpc_id: Option<String>,
    pub architecture: Option<ArchitectureValues>,
    pub block_device_mappings: Option<Vec<InstanceBlockDeviceMapping>>,
    pub client_token: Option<String>,
    pub ebs_optimized: Option<bool>,
    pub ena_support: Option<bool>,
    pub hypervisor: Option<HypervisorType>,
    pub iam_instance_profile: Option<IamInstanceProfile>,
    pub instance_lifecycle: Option<InstanceLifecycleType>,
    pub elastic_gpu_associations: Option<Vec<ElasticGpuAssociation>>,
    pub elastic_inference_accelerator_associations: Option<Vec<ElasticInferenceAcceleratorAssociation>>,
    pub network_interfaces: Option<Vec<InstanceNetworkInterface>>,
    pub outpost_arn: Option<String>,
    pub root_device_name: Option<String>,
    pub root_device_type: Option<DeviceType>,
    pub security_groups: Option<Vec<GroupIdentifier>>,
    pub source_dest_check: Option<bool>,
    pub spot_instance_request_id: Option<String>,
    pub sriov_net_support: Option<String>,
    pub state_reason: Option<StateReason>,
    pub tags: Option<Vec<Tag>>,
    pub virtualization_type: Option<VirtualizationType>,
    pub cpu_options: Option<CpuOptions>,
    pub capacity_reservation_id: Option<String>,
    pub capacity_reservation_specification: Option<CapacityReservationSpecificationResponse>,
    pub hibernation_options: Option<HibernationOptions>,
    pub licenses: Option<Vec<LicenseConfiguration>>,
    pub metadata_options: Option<InstanceMetadataOptionsResponse>,
    pub enclave_options: Option<EnclaveOptions>,
    pub boot_mode: Option<BootModeValues>,

}

Describes an instance.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

ami_launch_index: Option<i32>

The AMI launch index, which can be used to find this instance in the launch group.

image_id: Option<String>

The ID of the AMI used to launch the instance.

instance_id: Option<String>

The ID of the instance.

instance_type: Option<InstanceType>

The instance type.

kernel_id: Option<String>

The kernel associated with this instance, if applicable.

key_name: Option<String>

The name of the key pair, if this instance was launched with an associated key pair.

launch_time: Option<Instant>

The time the instance was launched.

monitoring: Option<Monitoring>

The monitoring for the instance.

placement: Option<Placement>

The location where the instance launched, if applicable.

platform: Option<PlatformValues>

The value is Windows for Windows instances; otherwise blank.

private_dns_name: Option<String>

(IPv4 only) The private DNS hostname name assigned to the instance. This DNS hostname can only be used inside the Amazon EC2 network. This name is not available until the instance enters the running state.

[EC2-VPC] The Amazon-provided DNS server resolves Amazon-provided private DNS hostnames if you've enabled DNS resolution and DNS hostnames in your VPC. If you are not using the Amazon-provided DNS server in your VPC, your custom domain name servers must resolve the hostname as appropriate.

private_ip_address: Option<String>

The private IPv4 address assigned to the instance.

product_codes: Option<Vec<ProductCode>>

The product codes attached to this instance, if applicable.

public_dns_name: Option<String>

(IPv4 only) The public DNS name assigned to the instance. This name is not available until the instance enters the running state. For EC2-VPC, this name is only available if you've enabled DNS hostnames for your VPC.

public_ip_address: Option<String>

The public IPv4 address, or the Carrier IP address assigned to the instance, if applicable.

A Carrier IP address only applies to an instance launched in a subnet associated with a Wavelength Zone.

ramdisk_id: Option<String>

The RAM disk associated with this instance, if applicable.

state: Option<InstanceState>

The current state of the instance.

state_transition_reason: Option<String>

The reason for the most recent state transition. This might be an empty string.

subnet_id: Option<String>

[EC2-VPC] The ID of the subnet in which the instance is running.

vpc_id: Option<String>

[EC2-VPC] The ID of the VPC in which the instance is running.

architecture: Option<ArchitectureValues>

The architecture of the image.

block_device_mappings: Option<Vec<InstanceBlockDeviceMapping>>

Any block device mapping entries for the instance.

client_token: Option<String>

The idempotency token you provided when you launched the instance, if applicable.

ebs_optimized: Option<bool>

Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

ena_support: Option<bool>

Specifies whether enhanced networking with ENA is enabled.

hypervisor: Option<HypervisorType>

The hypervisor type of the instance. The value xen is used for both Xen and Nitro hypervisors.

iam_instance_profile: Option<IamInstanceProfile>

The IAM instance profile associated with the instance, if applicable.

instance_lifecycle: Option<InstanceLifecycleType>

Indicates whether this is a Spot Instance or a Scheduled Instance.

elastic_gpu_associations: Option<Vec<ElasticGpuAssociation>>

The Elastic GPU associated with the instance.

elastic_inference_accelerator_associations: Option<Vec<ElasticInferenceAcceleratorAssociation>>

The elastic inference accelerator associated with the instance.

network_interfaces: Option<Vec<InstanceNetworkInterface>>

[EC2-VPC] The network interfaces for the instance.

outpost_arn: Option<String>

The Amazon Resource Name (ARN) of the Outpost.

root_device_name: Option<String>

The device name of the root device volume (for example, /dev/sda1).

root_device_type: Option<DeviceType>

The root device type used by the AMI. The AMI can use an EBS volume or an instance store volume.

security_groups: Option<Vec<GroupIdentifier>>

The security groups for the instance.

source_dest_check: Option<bool>

Indicates whether source/destination checking is enabled.

spot_instance_request_id: Option<String>

If the request is a Spot Instance request, the ID of the request.

sriov_net_support: Option<String>

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

state_reason: Option<StateReason>

The reason for the most recent state transition.

tags: Option<Vec<Tag>>

Any tags assigned to the instance.

virtualization_type: Option<VirtualizationType>

The virtualization type of the instance.

cpu_options: Option<CpuOptions>

The CPU options for the instance.

capacity_reservation_id: Option<String>

The ID of the Capacity Reservation.

capacity_reservation_specification: Option<CapacityReservationSpecificationResponse>

Information about the Capacity Reservation targeting option.

hibernation_options: Option<HibernationOptions>

Indicates whether the instance is enabled for hibernation.

licenses: Option<Vec<LicenseConfiguration>>

The license configurations.

metadata_options: Option<InstanceMetadataOptionsResponse>

The metadata options for the instance.

enclave_options: Option<EnclaveOptions>

Indicates whether the instance is enabled for Amazon Web Services Nitro Enclaves.

boot_mode: Option<BootModeValues>

The boot mode of the instance. For more information, see Boot modes in the Amazon EC2 User Guide.

Implementations

impl Instance

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Instance