Module aws_sdk_ec2::operation

Structs

AcceptReservedInstancesExchangeQuote

Accepts the Convertible Reserved Instance exchange quote described in the GetReservedInstancesExchangeQuote call.

AcceptTransitGatewayMulticastDomainAssociations

Accepts a request to associate subnets with a transit gateway multicast domain.

AcceptTransitGatewayPeeringAttachment

Accepts a transit gateway peering attachment request. The peering attachment must be

AcceptTransitGatewayVpcAttachment

Accepts a request to attach a VPC to a transit gateway.

AcceptVpcEndpointConnections

Accepts one or more interface VPC endpoint connection requests to your VPC endpoint

AcceptVpcPeeringConnection

Accept a VPC peering connection request. To accept a request, the VPC peering connection must

AdvertiseByoipCidr

Advertises an IPv4 or IPv6 address range that is provisioned for use with your Amazon Web Services resources through

AllocateAddress

Allocates an Elastic IP address to your account. After you allocate the Elastic IP address you can associate

AllocateHosts

Allocates a Dedicated Host to your account. At a minimum, specify the supported instance

ApplySecurityGroupsToClientVpnTargetNetwork

Applies a security group to the association between the target network and the Client VPN endpoint. This action replaces the existing

AssignIpv6Addresses

Assigns one or more IPv6 addresses to the specified network interface. You can

AssignPrivateIpAddresses

Assigns one or more secondary private IP addresses to the specified network interface.

AssociateAddress

Associates an Elastic IP address, or carrier IP address (for instances that are in

AssociateClientVpnTargetNetwork

Associates a target network with a Client VPN endpoint. A target network is a subnet in a VPC. You can associate multiple subnets from the same VPC with a Client VPN endpoint. You can associate only one subnet in each Availability Zone. We recommend that you associate at least two subnets to provide Availability Zone redundancy.

AssociateDhcpOptions

Associates a set of DHCP options (that you've previously created) with the specified VPC, or associates no DHCP options with the VPC.

AssociateEnclaveCertificateIamRole

Associates an AWS Identity and Access Management (IAM) role with an AWS Certificate Manager (ACM) certificate.

AssociateIamInstanceProfile

Associates an IAM instance profile with a running or stopped instance. You cannot

AssociateRouteTable

Associates a subnet in your VPC or an internet gateway or virtual private gateway

AssociateSubnetCidrBlock

Associates a CIDR block with your subnet. You can only associate a single IPv6 CIDR

AssociateTransitGatewayMulticastDomain

Associates the specified subnets and transit gateway attachments with the specified transit gateway multicast domain.

AssociateTransitGatewayRouteTable

Associates the specified attachment with the specified transit gateway route table. You can

AssociateTrunkInterface

Associates a branch network interface with a trunk network interface.

AssociateVpcCidrBlock

Associates a CIDR block with your VPC. You can associate a secondary IPv4 CIDR block,

AttachClassicLinkVpc

Links an EC2-Classic instance to a ClassicLink-enabled VPC through one or more of the VPC's

AttachInternetGateway

Attaches an internet gateway or a virtual private gateway to a VPC, enabling connectivity between the internet and

AttachNetworkInterface

Attaches a network interface to an instance.

AttachVolume

Attaches an EBS volume to a running or stopped instance and exposes it to the instance

AttachVpnGateway

Attaches a virtual private gateway to a VPC. You can attach one virtual private

AuthorizeClientVpnIngress

Adds an ingress authorization rule to a Client VPN endpoint. Ingress authorization rules act as

AuthorizeSecurityGroupEgress

[VPC only] Adds the specified egress rules to a security group for use with a VPC.

AuthorizeSecurityGroupIngress

Adds the specified ingress rules to a security group.

BundleInstance

Bundles an Amazon instance store-backed Windows instance.

CancelBundleTask

Cancels a bundling operation for an instance store-backed Windows instance.

CancelCapacityReservation

Cancels the specified Capacity Reservation, releases the reserved capacity, and changes the Capacity Reservation's state to

CancelConversionTask

Cancels an active conversion task. The task can be the import of an instance or volume. The action removes all

CancelExportTask

Cancels an active export task. The request removes all artifacts of the export, including any partially-created

CancelImportTask

Cancels an in-process import virtual machine or import snapshot task.

CancelReservedInstancesListing

Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace.

CancelSpotFleetRequests

Cancels the specified Spot Fleet requests.

CancelSpotInstanceRequests

Cancels one or more Spot Instance requests.

ConfirmProductInstance

Determines whether a product code is associated with an instance. This action can only

CopyFpgaImage

Copies the specified Amazon FPGA Image (AFI) to the current Region.

CopyImage

Initiates the copy of an AMI. You can copy an AMI from one Region to another, or from a

CopySnapshot

Copies a point-in-time snapshot of an EBS volume and stores it in Amazon S3. You can copy a

CreateCapacityReservation

Creates a new Capacity Reservation with the specified attributes.

CreateCarrierGateway

Creates a carrier gateway. For more information about carrier gateways, see Carrier gateways in the AWS Wavelength Developer

CreateClientVpnEndpoint

Creates a Client VPN endpoint. A Client VPN endpoint is the resource you create and configure to

CreateClientVpnRoute

Adds a route to a network to a Client VPN endpoint. Each Client VPN endpoint has a route table that describes the

CreateCustomerGateway

Provides information to AWS about your VPN customer gateway device. The customer

CreateDefaultSubnet

Creates a default subnet with a size /20 IPv4 CIDR block in the

CreateDefaultVpc

Creates a default VPC with a size /16 IPv4 CIDR block and a default subnet

CreateDhcpOptions

Creates a set of DHCP options for your VPC. After creating the set, you must

CreateEgressOnlyInternetGateway

[IPv6 only] Creates an egress-only internet gateway for your VPC. An egress-only

CreateFleet

Launches an EC2 Fleet.

CreateFlowLogs

Creates one or more flow logs to capture information about IP traffic for a specific network interface,

CreateFpgaImage

Creates an Amazon FPGA Image (AFI) from the specified design checkpoint (DCP).

CreateImage

Creates an Amazon EBS-backed AMI from an Amazon EBS-backed instance

CreateInstanceExportTask

Exports a running or stopped instance to an Amazon S3 bucket.

CreateInternetGateway

Creates an internet gateway for use with a VPC. After creating the internet gateway,

CreateKeyPair

Creates a 2048-bit RSA key pair with the specified name. Amazon EC2 stores the public

CreateLaunchTemplate

Creates a launch template. A launch template contains the parameters to launch an

CreateLaunchTemplateVersion

Creates a new version for a launch template. You can specify an existing version of

CreateLocalGatewayRoute

Creates a static route for the specified local gateway route table.

CreateLocalGatewayRouteTableVpcAssociation

Associates the specified VPC with the specified local gateway route table.

CreateManagedPrefixList

Creates a managed prefix list. You can specify one or more entries for the prefix list. Each entry consists of a CIDR block and an optional description.

CreateNatGateway

Creates a NAT gateway in the specified subnet. This action creates a network interface

CreateNetworkAcl

Creates a network ACL in a VPC. Network ACLs provide an optional layer of security (in addition to security groups) for the instances in your VPC.

CreateNetworkAclEntry

Creates an entry (a rule) in a network ACL with the specified rule number. Each network ACL has a set of numbered ingress rules

CreateNetworkInsightsPath

Creates a path to analyze for reachability.

CreateNetworkInterface

Creates a network interface in the specified subnet.

CreateNetworkInterfacePermission

Grants an Amazon Web Services-authorized account permission to attach the specified network interface to

CreatePlacementGroup

Creates a placement group in which to launch instances. The strategy of the placement

CreateReplaceRootVolumeTask

Creates a root volume replacement task for an Amazon EC2 instance. The root volume

CreateReservedInstancesListing

Creates a listing for Amazon EC2 Standard Reserved Instances to be sold in the Reserved Instance

CreateRestoreImageTask

Starts a task that restores an AMI from an S3 object that was previously created by using

CreateRoute

Creates a route in a route table within a VPC.

CreateRouteTable

Creates a route table for the specified VPC. After you create a route table, you can add routes and associate the table with a subnet.

CreateSecurityGroup

Creates a security group.

CreateSnapshot

Creates a snapshot of an EBS volume and stores it in Amazon S3. You can use snapshots for

CreateSnapshots

Creates crash-consistent snapshots of multiple EBS volumes and stores the data in S3.

CreateSpotDatafeedSubscription

Creates a data feed for Spot Instances, enabling you to view Spot Instance usage logs.

CreateStoreImageTask

Stores an AMI as a single object in an S3 bucket.

CreateSubnet

Creates a subnet in a specified VPC.

CreateTags

Adds or overwrites only the specified tags for the specified Amazon EC2 resource or

CreateTrafficMirrorFilter

Creates a Traffic Mirror filter.

CreateTrafficMirrorFilterRule

Creates a Traffic Mirror filter rule.

CreateTrafficMirrorSession

Creates a Traffic Mirror session.

CreateTrafficMirrorTarget

Creates a target for your Traffic Mirror session.

CreateTransitGateway

Creates a transit gateway.

CreateTransitGatewayConnect

Creates a Connect attachment from a specified transit gateway attachment. A Connect attachment is a GRE-based tunnel attachment that you can use to establish a connection between a transit gateway and an appliance.

CreateTransitGatewayConnectPeer

Creates a Connect peer for a specified transit gateway Connect attachment between a

CreateTransitGatewayMulticastDomain

Creates a multicast domain using the specified transit gateway.

CreateTransitGatewayPeeringAttachment

Requests a transit gateway peering attachment between the specified transit gateway

CreateTransitGatewayPrefixListReference

Creates a reference (route) to a prefix list in a specified transit gateway route table.

CreateTransitGatewayRoute

Creates a static route for the specified transit gateway route table.

CreateTransitGatewayRouteTable

Creates a route table for the specified transit gateway.

CreateTransitGatewayVpcAttachment

Attaches the specified VPC to the specified transit gateway.

CreateVolume

Creates an EBS volume that can be attached to an instance in the same Availability Zone.

CreateVpc

Creates a VPC with the specified IPv4 CIDR block. The smallest VPC you can create

CreateVpcEndpoint

Creates a VPC endpoint for a specified service. An endpoint enables you to create a

CreateVpcEndpointConnectionNotification

Creates a connection notification for a specified VPC endpoint or VPC endpoint

CreateVpcEndpointServiceConfiguration

Creates a VPC endpoint service configuration to which service consumers (AWS accounts,

CreateVpcPeeringConnection

Requests a VPC peering connection between two VPCs: a requester VPC that you own and

CreateVpnConnection

Creates a VPN connection between an existing virtual private gateway or transit gateway and a

CreateVpnConnectionRoute

Creates a static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway.

CreateVpnGateway

Creates a virtual private gateway. A virtual private gateway is the endpoint on the VPC side of your VPN connection. You can create a virtual private gateway before creating the VPC itself.

DeleteCarrierGateway

Deletes a carrier gateway.

DeleteClientVpnEndpoint

Deletes the specified Client VPN endpoint. You must disassociate all target networks before you

DeleteClientVpnRoute

Deletes a route from a Client VPN endpoint. You can only delete routes that you manually added using

DeleteCustomerGateway

Deletes the specified customer gateway. You must delete the VPN connection before you can delete the customer gateway.

DeleteDhcpOptions

Deletes the specified set of DHCP options. You must disassociate the set of DHCP options before you can delete it. You can disassociate the set of DHCP options by associating either a new set of options or the default set of options with the VPC.

DeleteEgressOnlyInternetGateway

Deletes an egress-only internet gateway.

DeleteFleets

Deletes the specified EC2 Fleet.

DeleteFlowLogs

Deletes one or more flow logs.

DeleteFpgaImage

Deletes the specified Amazon FPGA Image (AFI).

DeleteInternetGateway

Deletes the specified internet gateway. You must detach the internet gateway from the

DeleteKeyPair

Deletes the specified key pair, by removing the public key from Amazon EC2.

DeleteLaunchTemplate

Deletes a launch template. Deleting a launch template deletes all of its versions.

DeleteLaunchTemplateVersions

Deletes one or more versions of a launch template. You cannot delete the default

DeleteLocalGatewayRoute

Deletes the specified route from the specified local gateway route table.

DeleteLocalGatewayRouteTableVpcAssociation

Deletes the specified association between a VPC and local gateway route table.

DeleteManagedPrefixList

Deletes the specified managed prefix list. You must first remove all references to the prefix list in your resources.

DeleteNatGateway

Deletes the specified NAT gateway. Deleting a public NAT gateway disassociates its Elastic IP address,

DeleteNetworkAcl

Deletes the specified network ACL. You can't delete the ACL if it's associated with any subnets. You can't delete the default network ACL.

DeleteNetworkAclEntry

Deletes the specified ingress or egress entry (rule) from the specified network ACL.

DeleteNetworkInsightsAnalysis

Deletes the specified network insights analysis.

DeleteNetworkInsightsPath

Deletes the specified path.

DeleteNetworkInterface

Deletes the specified network interface. You must detach the network interface before you can delete it.

DeleteNetworkInterfacePermission

Deletes a permission for a network interface. By default, you cannot delete the

DeletePlacementGroup

Deletes the specified placement group. You must terminate all instances in the

DeleteQueuedReservedInstances

Deletes the queued purchases for the specified Reserved Instances.

DeleteRoute

Deletes the specified route from the specified route table.

DeleteRouteTable

Deletes the specified route table. You must disassociate the route table from any subnets before you can delete it. You can't delete the main route table.

DeleteSecurityGroup

Deletes a security group.

DeleteSnapshot

Deletes the specified snapshot.

DeleteSpotDatafeedSubscription

Deletes the data feed for Spot Instances.

DeleteSubnet

Deletes the specified subnet. You must terminate all running instances in the subnet before you can delete the subnet.

DeleteTags

Deletes the specified set of tags from the specified set of resources.

DeleteTrafficMirrorFilter

Deletes the specified Traffic Mirror filter.

DeleteTrafficMirrorFilterRule

Deletes the specified Traffic Mirror rule.

DeleteTrafficMirrorSession

Deletes the specified Traffic Mirror session.

DeleteTrafficMirrorTarget

Deletes the specified Traffic Mirror target.

DeleteTransitGateway

Deletes the specified transit gateway.

DeleteTransitGatewayConnect

Deletes the specified Connect attachment. You must first delete any Connect peers for

DeleteTransitGatewayConnectPeer

Deletes the specified Connect peer.

DeleteTransitGatewayMulticastDomain

Deletes the specified transit gateway multicast domain.

DeleteTransitGatewayPeeringAttachment

Deletes a transit gateway peering attachment.

DeleteTransitGatewayPrefixListReference

Deletes a reference (route) to a prefix list in a specified transit gateway route table.

DeleteTransitGatewayRoute

Deletes the specified route from the specified transit gateway route table.

DeleteTransitGatewayRouteTable

Deletes the specified transit gateway route table. You must disassociate the route table from any

DeleteTransitGatewayVpcAttachment

Deletes the specified VPC attachment.

DeleteVolume

Deletes the specified EBS volume. The volume must be in the available state

DeleteVpc

Deletes the specified VPC. You must detach or delete all gateways and resources that are associated with the VPC before you can delete it. For example, you must terminate all instances running in the VPC, delete all security groups associated with the VPC (except the default one), delete all route tables associated with the VPC (except the default one), and so on.

DeleteVpcEndpointConnectionNotifications

Deletes one or more VPC endpoint connection notifications.

DeleteVpcEndpointServiceConfigurations

Deletes one or more VPC endpoint service configurations in your account. Before you

DeleteVpcEndpoints

Deletes one or more specified VPC endpoints. You can delete any of the following types of VPC endpoints.

DeleteVpcPeeringConnection

Deletes a VPC peering connection. Either the owner of the requester VPC or the owner

DeleteVpnConnection

Deletes the specified VPN connection.

DeleteVpnConnectionRoute

Deletes the specified static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway.

DeleteVpnGateway

Deletes the specified virtual private gateway. You must first detach the virtual private gateway from the VPC. Note that you don't need to delete the virtual private gateway if you plan to delete and recreate the VPN connection between your VPC and your network.

DeprovisionByoipCidr

Releases the specified address range that you provisioned for use with your Amazon Web Services resources

DeregisterImage

Deregisters the specified AMI. After you deregister an AMI, it can't be used to launch

DeregisterInstanceEventNotificationAttributes

Deregisters tag keys to prevent tags that have the specified tag keys from being included

DeregisterTransitGatewayMulticastGroupMembers

Deregisters the specified members (network interfaces) from the transit gateway multicast group.

DeregisterTransitGatewayMulticastGroupSources

Deregisters the specified sources (network interfaces) from the transit gateway multicast group.

DescribeAccountAttributes

Describes attributes of your AWS account. The following are the supported account attributes:

DescribeAddresses

Describes the specified Elastic IP addresses or all of your Elastic IP addresses.

DescribeAddressesAttribute

Describes the attributes of the specified Elastic IP addresses. For requirements, see Using reverse DNS for email applications.

DescribeAggregateIdFormat

Describes the longer ID format settings for all resource types in a specific

DescribeAvailabilityZones

Describes the Availability Zones, Local Zones, and Wavelength Zones that are available to

DescribeBundleTasks

Describes the specified bundle tasks or all of your bundle tasks.

DescribeByoipCidrs

Describes the IP address ranges that were specified in calls to ProvisionByoipCidr.

DescribeCapacityReservations

Describes one or more of your Capacity Reservations. The results describe only the Capacity Reservations in the

DescribeCarrierGateways

Describes one or more of your carrier gateways.

DescribeClassicLinkInstances

Describes one or more of your linked EC2-Classic instances. This request only returns

DescribeClientVpnAuthorizationRules

Describes the authorization rules for a specified Client VPN endpoint.

DescribeClientVpnConnections

Describes active client connections and connections that have been terminated within the last 60

DescribeClientVpnEndpoints

Describes one or more Client VPN endpoints in the account.

DescribeClientVpnRoutes

Describes the routes for the specified Client VPN endpoint.

DescribeClientVpnTargetNetworks

Describes the target networks associated with the specified Client VPN endpoint.

DescribeCoipPools

Describes the specified customer-owned address pools or all of your customer-owned address pools.

DescribeConversionTasks

Describes the specified conversion tasks or all your conversion tasks. For more information, see the

DescribeCustomerGateways

Describes one or more of your VPN customer gateways.

DescribeDhcpOptions

Describes one or more of your DHCP options sets.

DescribeEgressOnlyInternetGateways

Describes one or more of your egress-only internet gateways.

DescribeElasticGpus

Describes the Elastic Graphics accelerator associated with your instances. For more information

DescribeExportImageTasks

Describes the specified export image tasks or all of your export image tasks.

DescribeExportTasks

Describes the specified export instance tasks or all of your export instance tasks.

DescribeFastSnapshotRestores

Describes the state of fast snapshot restores for your snapshots.

DescribeFleetHistory

Describes the events for the specified EC2 Fleet during the specified time.

DescribeFleetInstances

Describes the running instances for the specified EC2 Fleet.

DescribeFleets

Describes the specified EC2 Fleets or all of your EC2 Fleets.

DescribeFlowLogs

Describes one or more flow logs. To view the information in your flow logs (the log

DescribeFpgaImageAttribute

Describes the specified attribute of the specified Amazon FPGA Image (AFI).

DescribeFpgaImages

Describes the Amazon FPGA Images (AFIs) available to you. These include public AFIs,

DescribeHostReservationOfferings

Describes the Dedicated Host reservations that are available to purchase.

DescribeHostReservations

Describes reservations that are associated with Dedicated Hosts in your

DescribeHosts

Describes the specified Dedicated Hosts or all your Dedicated Hosts.

DescribeIamInstanceProfileAssociations

Describes your IAM instance profile associations.

DescribeIdFormat

Describes the ID format settings for your resources on a per-Region basis, for example, to view which resource types are enabled for longer IDs. This request only returns information about resource types whose ID formats can be modified; it does not return information about other resource types.

DescribeIdentityIdFormat

Describes the ID format settings for resources for the specified IAM user, IAM role, or root

DescribeImageAttribute

Describes the specified attribute of the specified AMI. You can specify only one attribute at a time.

DescribeImages

Describes the specified images (AMIs, AKIs, and ARIs) available to you or all of the images available to you.

DescribeImportImageTasks

Displays details about an import virtual machine or import snapshot tasks that are already created.

DescribeImportSnapshotTasks

Describes your import snapshot tasks.

DescribeInstanceAttribute

Describes the specified attribute of the specified instance. You can specify only one

DescribeInstanceCreditSpecifications

Describes the credit option for CPU usage of the specified burstable performance

DescribeInstanceEventNotificationAttributes

Describes the tag keys that are registered to appear in scheduled event notifications for

DescribeInstanceStatus

Describes the status of the specified instances or all of your instances. By default,

DescribeInstanceTypeOfferings

Returns a list of all instance types offered. The results can be filtered by location (Region or Availability

DescribeInstanceTypes

Describes the details of the instance types that are offered in a location. The results can be filtered by the

DescribeInstances

Describes the specified instances or all instances.

DescribeInternetGateways

Describes one or more of your internet gateways.

DescribeIpv6Pools

Describes your IPv6 address pools.

DescribeKeyPairs

Describes the specified key pairs or all of your key pairs.

DescribeLaunchTemplateVersions

Describes one or more versions of a specified launch template. You can describe all

DescribeLaunchTemplates

Describes one or more launch templates.

DescribeLocalGatewayRouteTableVirtualInterfaceGroupAssociations

Describes the associations between virtual interface groups and local gateway route tables.

DescribeLocalGatewayRouteTableVpcAssociations

Describes the specified associations between VPCs and local gateway route tables.

DescribeLocalGatewayRouteTables

Describes one or more local gateway route tables. By default, all local gateway route tables are described.

DescribeLocalGatewayVirtualInterfaceGroups

Describes the specified local gateway virtual interface groups.

DescribeLocalGatewayVirtualInterfaces

Describes the specified local gateway virtual interfaces.

DescribeLocalGateways

Describes one or more local gateways. By default, all local gateways are described.

DescribeManagedPrefixLists

Describes your managed prefix lists and any Amazon Web Services-managed prefix lists.

DescribeMovingAddresses

Describes your Elastic IP addresses that are being moved to the EC2-VPC platform, or that are being restored to the EC2-Classic platform. This request does not return information about any other Elastic IP addresses in your account.

DescribeNatGateways

Describes one or more of your NAT gateways.

DescribeNetworkAcls

Describes one or more of your network ACLs.

DescribeNetworkInsightsAnalyses

Describes one or more of your network insights analyses.

DescribeNetworkInsightsPaths

Describes one or more of your paths.

DescribeNetworkInterfaceAttribute

Describes a network interface attribute. You can specify only one attribute at a time.

DescribeNetworkInterfacePermissions

Describes the permissions for your network interfaces.

DescribeNetworkInterfaces

Describes one or more of your network interfaces.

DescribePlacementGroups

Describes the specified placement groups or all of your placement groups. For more

DescribePrefixLists

Describes available Amazon Web Services services in a prefix list format, which includes the prefix list

DescribePrincipalIdFormat

Describes the ID format settings for the root user and all IAM roles and IAM users

DescribePublicIpv4Pools

Describes the specified IPv4 address pools.

DescribeRegions

Describes the Regions that are enabled for your account, or all Regions.

DescribeReplaceRootVolumeTasks

Describes a root volume replacement task. For more information, see

DescribeReservedInstances

Describes one or more of the Reserved Instances that you purchased.

DescribeReservedInstancesListings

Describes your account's Reserved Instance listings in the Reserved Instance Marketplace.

DescribeReservedInstancesModifications

Describes the modifications made to your Reserved Instances. If no parameter is specified, information about all your Reserved Instances modification requests is returned. If a modification ID is specified, only information about the specific modification is returned.

DescribeReservedInstancesOfferings

Describes Reserved Instance offerings that are available for purchase. With Reserved Instances, you purchase the right to launch instances for a period of time. During that time period, you do not receive insufficient capacity errors, and you pay a lower usage rate than the rate charged for On-Demand instances for the actual time used.

DescribeRouteTables

Describes one or more of your route tables.

DescribeScheduledInstanceAvailability

Finds available schedules that meet the specified criteria.

DescribeScheduledInstances

Describes the specified Scheduled Instances or all your Scheduled Instances.

DescribeSecurityGroupReferences

[VPC only] Describes the VPCs on the other side of a VPC peering connection that are referencing the security groups you've specified in this request.

DescribeSecurityGroups

Describes the specified security groups or all of your security groups.

DescribeSnapshotAttribute

Describes the specified attribute of the specified snapshot. You can specify only one

DescribeSnapshots

Describes the specified EBS snapshots available to you or all of the EBS snapshots

DescribeSpotDatafeedSubscription

Describes the data feed for Spot Instances. For more information, see Spot

DescribeSpotFleetInstances

Describes the running instances for the specified Spot Fleet.

DescribeSpotFleetRequestHistory

Describes the events for the specified Spot Fleet request during the specified

DescribeSpotFleetRequests

Describes your Spot Fleet requests.

DescribeSpotInstanceRequests

Describes the specified Spot Instance requests.

DescribeSpotPriceHistory

Describes the Spot price history. For more information, see Spot Instance pricing history in the

DescribeStaleSecurityGroups

[VPC only] Describes the stale security group rules for security groups in a specified VPC. Rules are stale when they reference a deleted security group in a peer VPC, or a security group in a peer VPC for which the VPC peering connection has been deleted.

DescribeStoreImageTasks

Describes the progress of the AMI store tasks. You can describe the store tasks for

DescribeSubnets

Describes one or more of your subnets.

DescribeTags

Describes the specified tags for your EC2 resources.

DescribeTrafficMirrorFilters

Describes one or more Traffic Mirror filters.

DescribeTrafficMirrorSessions

Describes one or more Traffic Mirror sessions. By default, all Traffic Mirror sessions are described. Alternatively, you can filter the results.

DescribeTrafficMirrorTargets

Information about one or more Traffic Mirror targets.

DescribeTransitGatewayAttachments

Describes one or more attachments between resources and transit gateways. By default, all attachments are described.

DescribeTransitGatewayConnectPeers

Describes one or more Connect peers.

DescribeTransitGatewayConnects

Describes one or more Connect attachments.

DescribeTransitGatewayMulticastDomains

Describes one or more transit gateway multicast domains.

DescribeTransitGatewayPeeringAttachments

Describes your transit gateway peering attachments.

DescribeTransitGatewayRouteTables

Describes one or more transit gateway route tables. By default, all transit gateway route tables are described.

DescribeTransitGatewayVpcAttachments

Describes one or more VPC attachments. By default, all VPC attachments are described.

DescribeTransitGateways

Describes one or more transit gateways. By default, all transit gateways are described. Alternatively, you can

DescribeTrunkInterfaceAssociations

Describes one or more network interface trunk associations.

DescribeVolumeAttribute

Describes the specified attribute of the specified volume. You can specify only one

DescribeVolumeStatus

Describes the status of the specified volumes. Volume status provides the result of the

DescribeVolumes

Describes the specified EBS volumes or all of your EBS volumes.

DescribeVolumesModifications

Describes the most recent volume modification request for the specified EBS volumes.

DescribeVpcAttribute

Describes the specified attribute of the specified VPC. You can specify only one attribute at a time.

DescribeVpcClassicLink

Describes the ClassicLink status of one or more VPCs.

DescribeVpcClassicLinkDnsSupport

Describes the ClassicLink DNS support status of one or more VPCs. If enabled, the DNS

DescribeVpcEndpointConnectionNotifications

Describes the connection notifications for VPC endpoints and VPC endpoint

DescribeVpcEndpointConnections

Describes the VPC endpoint connections to your VPC endpoint services, including any

DescribeVpcEndpointServiceConfigurations

Describes the VPC endpoint service configurations in your account (your services).

DescribeVpcEndpointServicePermissions

Describes the principals (service consumers) that are permitted to discover your VPC

DescribeVpcEndpointServices

Describes available services to which you can create a VPC endpoint.

DescribeVpcEndpoints

Describes one or more of your VPC endpoints.

DescribeVpcPeeringConnections

Describes one or more of your VPC peering connections.

DescribeVpcs

Describes one or more of your VPCs.

DescribeVpnConnections

Describes one or more of your VPN connections.

DescribeVpnGateways

Describes one or more of your virtual private gateways.

DetachClassicLinkVpc

Unlinks (detaches) a linked EC2-Classic instance from a VPC. After the instance has been unlinked, the VPC security groups are no longer associated with it. An instance is automatically unlinked from a VPC when it's stopped.

DetachInternetGateway

Detaches an internet gateway from a VPC, disabling connectivity between the internet

DetachNetworkInterface

Detaches a network interface from an instance.

DetachVolume

Detaches an EBS volume from an instance. Make sure to unmount any file systems on the

DetachVpnGateway

Detaches a virtual private gateway from a VPC. You do this if you're planning to turn off the VPC and not use it anymore. You can confirm a virtual private gateway has been completely detached from a VPC by describing the virtual private gateway (any attachments to the virtual private gateway are also described).

DisableEbsEncryptionByDefault

Disables EBS encryption by default for your account in the current Region.

DisableFastSnapshotRestores

Disables fast snapshot restores for the specified snapshots in the specified Availability Zones.

DisableImageDeprecation

Cancels the deprecation of the specified AMI.

DisableSerialConsoleAccess

Disables access to the EC2 serial console of all instances for your account. By default,

DisableTransitGatewayRouteTablePropagation

Disables the specified resource attachment from propagating routes to the specified

DisableVgwRoutePropagation

Disables a virtual private gateway (VGW) from propagating routes to a specified route table of a VPC.

DisableVpcClassicLink

Disables ClassicLink for a VPC. You cannot disable ClassicLink for a VPC that has EC2-Classic instances linked to it.

DisableVpcClassicLinkDnsSupport

Disables ClassicLink DNS support for a VPC. If disabled, DNS hostnames resolve to

DisassociateAddress

Disassociates an Elastic IP address from the instance or network interface it's associated with.

DisassociateClientVpnTargetNetwork

Disassociates a target network from the specified Client VPN endpoint. When you disassociate the

DisassociateEnclaveCertificateIamRole

Disassociates an IAM role from an AWS Certificate Manager (ACM) certificate. Disassociating an IAM role

DisassociateIamInstanceProfile

Disassociates an IAM instance profile from a running or stopped instance.

DisassociateRouteTable

Disassociates a subnet or gateway from a route table.

DisassociateSubnetCidrBlock

Disassociates a CIDR block from a subnet. Currently, you can disassociate an IPv6 CIDR block only. You must detach or delete all gateways and resources that are associated with the CIDR block before you can disassociate it.

DisassociateTransitGatewayMulticastDomain

Disassociates the specified subnets from the transit gateway multicast domain.

DisassociateTransitGatewayRouteTable

Disassociates a resource attachment from a transit gateway route table.

DisassociateTrunkInterface

Removes an association between a branch network interface with a trunk network interface.

DisassociateVpcCidrBlock

Disassociates a CIDR block from a VPC. To disassociate the CIDR block, you must

EnableEbsEncryptionByDefault

Enables EBS encryption by default for your account in the current Region.

EnableFastSnapshotRestores

Enables fast snapshot restores for the specified snapshots in the specified Availability Zones.

EnableImageDeprecation

Enables deprecation of the specified AMI at the specified date and time.

EnableSerialConsoleAccess

Enables access to the EC2 serial console of all instances for your account. By default,

EnableTransitGatewayRouteTablePropagation

Enables the specified attachment to propagate routes to the specified

EnableVgwRoutePropagation

Enables a virtual private gateway (VGW) to propagate routes to the specified route table of a VPC.

EnableVolumeIO

Enables I/O operations for a volume that had I/O operations disabled because the data on

EnableVpcClassicLink

Enables a VPC for ClassicLink. You can then link EC2-Classic instances to your

EnableVpcClassicLinkDnsSupport

Enables a VPC to support DNS hostname resolution for ClassicLink. If enabled, the DNS

ExportClientVpnClientCertificateRevocationList

Downloads the client certificate revocation list for the specified Client VPN endpoint.

ExportClientVpnClientConfiguration

Downloads the contents of the Client VPN endpoint configuration file for the specified Client VPN endpoint. The Client VPN endpoint configuration

ExportImage

Exports an Amazon Machine Image (AMI) to a VM file. For more information, see Exporting a VM

ExportTransitGatewayRoutes

Exports routes from the specified transit gateway route table to the specified S3 bucket.

GetAssociatedEnclaveCertificateIamRoles

Returns the IAM roles that are associated with the specified AWS Certificate Manager (ACM) certificate.

GetAssociatedIpv6PoolCidrs

Gets information about the IPv6 CIDR block associations for a specified IPv6 address pool.

GetCapacityReservationUsage

Gets usage information about a Capacity Reservation. If the Capacity Reservation is shared, it shows usage information for the Capacity Reservation owner

GetCoipPoolUsage

Describes the allocations from the specified customer-owned address pool.

GetConsoleOutput

Gets the console output for the specified instance. For Linux instances, the instance

GetConsoleScreenshot

Retrieve a JPG-format screenshot of a running instance to help with

GetDefaultCreditSpecification

Describes the default credit option for CPU usage of a burstable performance instance family.

GetEbsDefaultKmsKeyId

Describes the default customer master key (CMK) for EBS encryption by default for your account in this Region.

GetEbsEncryptionByDefault

Describes whether EBS encryption by default is enabled for your account in the current

GetFlowLogsIntegrationTemplate

Generates a CloudFormation template that streamlines and automates the integration of VPC flow logs

GetGroupsForCapacityReservation

Lists the resource groups to which a Capacity Reservation has been added.

GetHostReservationPurchasePreview

Preview a reservation purchase with configurations that match those of your

GetLaunchTemplateData

Retrieves the configuration data of the specified instance. You can use this data

GetManagedPrefixListAssociations

Gets information about the resources that are associated with the specified managed prefix list.

GetManagedPrefixListEntries

Gets information about the entries for a specified managed prefix list.

GetPasswordData

Retrieves the encrypted administrator password for a running Windows instance.

GetReservedInstancesExchangeQuote

Returns a quote and exchange information for exchanging one or more specified

GetSerialConsoleAccessStatus

Retrieves the access status of your account to the EC2 serial console of all instances. By

GetTransitGatewayAttachmentPropagations

Lists the route tables to which the specified resource attachment propagates routes.

GetTransitGatewayMulticastDomainAssociations

Gets information about the associations for the transit gateway multicast domain.

GetTransitGatewayPrefixListReferences

Gets information about the prefix list references in a specified transit gateway route table.

GetTransitGatewayRouteTableAssociations

Gets information about the associations for the specified transit gateway route table.

GetTransitGatewayRouteTablePropagations

Gets information about the route table propagations for the specified transit gateway route table.

ImportClientVpnClientCertificateRevocationList

Uploads a client certificate revocation list to the specified Client VPN endpoint. Uploading a client certificate revocation list overwrites the existing client certificate revocation list.

ImportImage

Import single or multi-volume disk images or EBS snapshots into an Amazon Machine Image (AMI).

ImportInstance

Creates an import instance task using metadata from the specified disk image.

ImportKeyPair

Imports the public key from an RSA key pair that you created with a third-party tool.

ImportSnapshot

Imports a disk into an EBS snapshot.

ImportVolume

Creates an import volume task using metadata from the specified disk image.

ModifyAddressAttribute

Modifies an attribute of the specified Elastic IP address. For requirements, see Using reverse DNS for email applications.

ModifyAvailabilityZoneGroup

Changes the opt-in status of the Local Zone and Wavelength Zone group for your

ModifyCapacityReservation

Modifies a Capacity Reservation's capacity and the conditions under which it is to be released. You

ModifyClientVpnEndpoint

Modifies the specified Client VPN endpoint. Modifying the DNS server resets existing client connections.

ModifyDefaultCreditSpecification

Modifies the default credit option for CPU usage of burstable performance instances.

ModifyEbsDefaultKmsKeyId

Changes the default customer master key (CMK) for EBS encryption by default for your account in this Region.

ModifyFleet

Modifies the specified EC2 Fleet.

ModifyFpgaImageAttribute

Modifies the specified attribute of the specified Amazon FPGA Image (AFI).

ModifyHosts

Modify the auto-placement setting of a Dedicated Host. When auto-placement is enabled,

ModifyIdFormat

Modifies the ID format for the specified resource on a per-Region basis. You can

ModifyIdentityIdFormat

Modifies the ID format of a resource for a specified IAM user, IAM role, or the root

ModifyImageAttribute

Modifies the specified attribute of the specified AMI. You can specify only one attribute at a time.

ModifyInstanceAttribute

Modifies the specified attribute of the specified instance. You can specify only one

ModifyInstanceCapacityReservationAttributes

Modifies the Capacity Reservation settings for a stopped instance. Use this action to configure an

ModifyInstanceCreditSpecification

Modifies the credit option for CPU usage on a running or stopped burstable performance

ModifyInstanceEventStartTime

Modifies the start time for a scheduled Amazon EC2 instance event.

ModifyInstanceMetadataOptions

Modify the instance metadata parameters on a running or stopped instance. When you

ModifyInstancePlacement

Modifies the placement attributes for a specified instance. You can do the

ModifyLaunchTemplate

Modifies a launch template. You can specify which version of the launch template to

ModifyManagedPrefixList

Modifies the specified managed prefix list.

ModifyNetworkInterfaceAttribute

Modifies the specified network interface attribute. You can specify only one

ModifyReservedInstances

Modifies the Availability Zone, instance count, instance type, or network platform

ModifySnapshotAttribute

Adds or removes permission settings for the specified snapshot. You may add or remove

ModifySpotFleetRequest

Modifies the specified Spot Fleet request.

ModifySubnetAttribute

Modifies a subnet attribute. You can only modify one attribute at a time.

ModifyTrafficMirrorFilterNetworkServices

Allows or restricts mirroring network services.

ModifyTrafficMirrorFilterRule

Modifies the specified Traffic Mirror rule.

ModifyTrafficMirrorSession

Modifies a Traffic Mirror session.

ModifyTransitGateway

Modifies the specified transit gateway. When you modify a transit gateway, the modified options are applied to new transit gateway attachments only. Your existing transit gateway attachments are not modified.

ModifyTransitGatewayPrefixListReference

Modifies a reference (route) to a prefix list in a specified transit gateway route table.

ModifyTransitGatewayVpcAttachment

Modifies the specified VPC attachment.

ModifyVolume

You can modify several parameters of an existing EBS volume, including volume size, volume

ModifyVolumeAttribute

Modifies a volume attribute.

ModifyVpcAttribute

Modifies the specified attribute of the specified VPC.

ModifyVpcEndpoint

Modifies attributes of a specified VPC endpoint. The attributes that you can modify

ModifyVpcEndpointConnectionNotification

Modifies a connection notification for VPC endpoint or VPC endpoint service. You

ModifyVpcEndpointServiceConfiguration

Modifies the attributes of your VPC endpoint service configuration. You can change the

ModifyVpcEndpointServicePermissions

Modifies the permissions for your VPC endpoint service. You can add or remove permissions for service consumers (IAM users,

ModifyVpcPeeringConnectionOptions

Modifies the VPC peering connection options on one side of a VPC peering connection. You can do the following:

ModifyVpcTenancy

Modifies the instance tenancy attribute of the specified VPC. You can change the

ModifyVpnConnection

Modifies the customer gateway or the target gateway of an AWS Site-to-Site VPN connection. To modify the target gateway, the following migration

ModifyVpnConnectionOptions

Modifies the connection options for your Site-to-Site VPN connection.

ModifyVpnTunnelCertificate

Modifies the VPN tunnel endpoint certificate.

ModifyVpnTunnelOptions

Modifies the options for a VPN tunnel in an AWS Site-to-Site VPN connection. You can modify

MonitorInstances

Enables detailed monitoring for a running instance. Otherwise, basic monitoring is

MoveAddressToVpc

Moves an Elastic IP address from the EC2-Classic platform to the EC2-VPC platform. The

ProvisionByoipCidr

Provisions an IPv4 or IPv6 address range for use with your Amazon Web Services resources through bring your own IP

PurchaseHostReservation

Purchase a reservation with configurations that match those of your Dedicated Host.

PurchaseReservedInstancesOffering

Purchases a Reserved Instance for use with your account. With Reserved Instances, you pay a lower

PurchaseScheduledInstances

Purchases the Scheduled Instances with the specified schedule.

RebootInstances

Requests a reboot of the specified instances. This operation is asynchronous; it only

RegisterImage

Registers an AMI. When you're creating an AMI, this is the final step you must complete

RegisterInstanceEventNotificationAttributes

Registers a set of tag keys to include in scheduled event notifications for your resources.

RegisterTransitGatewayMulticastGroupMembers

Registers members (network interfaces) with the transit gateway multicast group. A member is a network interface associated

RegisterTransitGatewayMulticastGroupSources

Registers sources (network interfaces) with the specified transit gateway multicast group.

RejectTransitGatewayMulticastDomainAssociations

Rejects a request to associate cross-account subnets with a transit gateway multicast domain.

RejectTransitGatewayPeeringAttachment

Rejects a transit gateway peering attachment request.

RejectTransitGatewayVpcAttachment

Rejects a request to attach a VPC to a transit gateway.

RejectVpcEndpointConnections

Rejects one or more VPC endpoint connection requests to your VPC endpoint

RejectVpcPeeringConnection

Rejects a VPC peering connection request. The VPC peering connection must be in the

ReleaseAddress

Releases the specified Elastic IP address.

ReleaseHosts

When you no longer want to use an On-Demand Dedicated Host it can be released.

ReplaceIamInstanceProfileAssociation

Replaces an IAM instance profile for the specified running instance. You can use

ReplaceNetworkAclAssociation

Changes which network ACL a subnet is associated with. By default when you create a

ReplaceNetworkAclEntry

Replaces an entry (rule) in a network ACL. For more information, see Network ACLs in the

ReplaceRoute

Replaces an existing route within a route table in a VPC. You must provide only one of

ReplaceRouteTableAssociation

Changes the route table associated with a given subnet, internet gateway, or virtual private gateway in a VPC. After the operation

ReplaceTransitGatewayRoute

Replaces the specified route in the specified transit gateway route table.

ReportInstanceStatus

Submits feedback about the status of an instance. The instance must be in the

RequestSpotFleet

Creates a Spot Fleet request.

RequestSpotInstances

Creates a Spot Instance request.

ResetAddressAttribute

Resets the attribute of the specified IP address. For requirements, see Using reverse DNS for email applications.

ResetEbsDefaultKmsKeyId

Resets the default customer master key (CMK) for EBS encryption for your account in this Region

ResetFpgaImageAttribute

Resets the specified attribute of the specified Amazon FPGA Image (AFI) to its default value.

ResetImageAttribute

Resets an attribute of an AMI to its default value.

ResetInstanceAttribute

Resets an attribute of an instance to its default value. To reset the

ResetNetworkInterfaceAttribute

Resets a network interface attribute. You can specify only one attribute at a time.

ResetSnapshotAttribute

Resets permission settings for the specified snapshot.

RestoreAddressToClassic

Restores an Elastic IP address that was previously moved to the EC2-VPC platform back to the EC2-Classic platform. You cannot move an Elastic IP address that was originally allocated for use in EC2-VPC. The Elastic IP address must not be associated with an instance or network interface.

RestoreManagedPrefixListVersion

Restores the entries from a previous version of a managed prefix list to a new version of the prefix list.

RevokeClientVpnIngress

Removes an ingress authorization rule from a Client VPN endpoint.

RevokeSecurityGroupEgress

[VPC only] Removes the specified egress rules from a security group for EC2-VPC.

RevokeSecurityGroupIngress

Removes the specified ingress rules from a security group. To remove a rule, the values

RunInstances

Launches the specified number of instances using an AMI for which you have

RunScheduledInstances

Launches the specified Scheduled Instances.

SearchLocalGatewayRoutes

Searches for routes in the specified local gateway route table.

SearchTransitGatewayMulticastGroups

Searches one or more transit gateway multicast groups and returns the group membership information.

SearchTransitGatewayRoutes

Searches for routes in the specified transit gateway route table.

SendDiagnosticInterrupt

Sends a diagnostic interrupt to the specified Amazon EC2 instance to trigger a

StartInstances

Starts an Amazon EBS-backed instance that you've previously stopped.

StartNetworkInsightsAnalysis

Starts analyzing the specified path. If the path is reachable, the

StartVpcEndpointServicePrivateDnsVerification

Initiates the verification process to prove that the service provider owns the private

StopInstances

Stops an Amazon EBS-backed instance.

TerminateClientVpnConnections

Terminates active Client VPN endpoint connections. This action can be used to terminate a specific client connection, or up to five connections established by a specific user.

TerminateInstances

Shuts down the specified instances. This operation is idempotent; if you terminate an

UnassignIpv6Addresses

Unassigns one or more IPv6 addresses from a network interface.

UnassignPrivateIpAddresses

Unassigns one or more secondary private IP addresses from a network interface.

UnmonitorInstances

Disables detailed monitoring for a running instance. For more information, see Monitoring

UpdateSecurityGroupRuleDescriptionsEgress

[VPC only] Updates the description of an egress (outbound) security group rule. You

UpdateSecurityGroupRuleDescriptionsIngress

Updates the description of an ingress (inbound) security group rule. You can replace an

WithdrawByoipCidr

Stops advertising an address range that is provisioned as an address pool.