Struct aws_sdk_ec2::output::describe_instance_attribute_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DescribeInstanceAttributeOutput

Implementations

impl Builder

pub fn groups(self, input: impl Into<GroupIdentifier>) -> Self

pub fn set_groups(self, input: Option<Vec<GroupIdentifier>>) -> Self

pub fn block_device_mappings(
self,
input: impl Into<InstanceBlockDeviceMapping>
) -> Self

pub fn set_block_device_mappings(
self,
input: Option<Vec<InstanceBlockDeviceMapping>>
) -> Self

pub fn disable_api_termination(self, input: AttributeBooleanValue) -> Self

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

pub fn set_disable_api_termination(
self,
input: Option<AttributeBooleanValue>
) -> Self

pub fn ena_support(self, input: AttributeBooleanValue) -> Self

Indicates whether enhanced networking with ENA is enabled.

pub fn set_ena_support(self, input: Option<AttributeBooleanValue>) -> Self

pub fn enclave_options(self, input: EnclaveOptions) -> Self

To enable the instance for Amazon Web Services Nitro Enclaves, set this parameter to true; otherwise, set it to false.

pub fn set_enclave_options(self, input: Option<EnclaveOptions>) -> Self

pub fn ebs_optimized(self, input: AttributeBooleanValue) -> Self

Indicates whether the instance is optimized for Amazon EBS I/O.

pub fn set_ebs_optimized(self, input: Option<AttributeBooleanValue>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the instance.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn instance_initiated_shutdown_behavior(self, input: AttributeValue) -> Self

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

pub fn set_instance_initiated_shutdown_behavior(
self,
input: Option<AttributeValue>
) -> Self

pub fn instance_type(self, input: AttributeValue) -> Self

The instance type.

pub fn set_instance_type(self, input: Option<AttributeValue>) -> Self

pub fn kernel_id(self, input: AttributeValue) -> Self

The kernel ID.

pub fn set_kernel_id(self, input: Option<AttributeValue>) -> Self

pub fn product_codes(self, input: impl Into<ProductCode>) -> Self

pub fn set_product_codes(self, input: Option<Vec<ProductCode>>) -> Self

pub fn ramdisk_id(self, input: AttributeValue) -> Self

The RAM disk ID.

pub fn set_ramdisk_id(self, input: Option<AttributeValue>) -> Self

pub fn root_device_name(self, input: AttributeValue) -> Self

The device name of the root device volume (for example, /dev/sda1).

pub fn set_root_device_name(self, input: Option<AttributeValue>) -> Self

pub fn source_dest_check(self, input: AttributeBooleanValue) -> Self

Enable or disable source/destination checks, which ensure that the instance is either the source or the destination of any traffic that it receives. If the value is true, source/destination checks are enabled; otherwise, they are disabled. The default value is true. You must disable source/destination checks if the instance runs services such as network address translation, routing, or firewalls.

pub fn set_source_dest_check(self, input: Option<AttributeBooleanValue>) -> Self

pub fn sriov_net_support(self, input: AttributeValue) -> Self

Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

pub fn set_sriov_net_support(self, input: Option<AttributeValue>) -> Self

pub fn user_data(self, input: AttributeValue) -> Self

The user data.

pub fn set_user_data(self, input: Option<AttributeValue>) -> Self

pub fn build(self) -> DescribeInstanceAttributeOutput

Consumes the builder and constructs a DescribeInstanceAttributeOutput